\name{addEmptyRow2Dataset}
\alias{addEmptyRow2Dataset}
\title{Inserts an empty row}
\usage{
  addEmptyRow2Dataset(org.data.frame, new_row_position)
}
\arguments{
  \item{org.data.frame}{A data frame}

  \item{new_row_position}{Where to insert the empty row}
}
\value{
  Returns the same frame with an inserted row

  TODO: This should probably be changed to insertRow in the
  misc package
}
\description{
  Inserts an empty row
}
\author{
  max
}

