% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_pks.R
\name{additional_pks}
\alias{additional_pks}
\title{Additional useful packages}
\usage{
additional_pks()
}
\value{
Installed packages
}
\description{
Installs additional packages useful for geology, statistics, and visualization.
}
