% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{all_equal}
\alias{all_equal}
\title{checks that all items in a list are equal}
\usage{
all_equal(X)
}
\arguments{
\item{x}{list}
}
\value{
a boolean
}
\description{
checks that all items in a list are equal
}
\author{
G.A.Paleologo
}

