% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{all_identical}
\alias{all_identical}
\title{checks that all items in a list are identical}
\usage{
all_identical(X)
}
\arguments{
\item{x}{list}
}
\value{
a boolean
}
\description{
checks that all items in a list are identical
}
\author{
G.A.Paleologo
}

