% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{avg_list}
\alias{avg_list}
\title{Given as an put a list of vectors and an integer, averages the n
trailing adjacent list elements in a list of vectors;
the first n elements are only averaged partially.}
\usage{
avg_list(x, n)
}
\arguments{
\item{x}{list of vectors. elements must be named and vector elements
must be named}

\item{n}{number of trailing elements}
}
\value{
a list of vectors
}
\description{
Given as an put a list of vectors and an integer, averages the n
trailing adjacent list elements in a list of vectors;
the first n elements are only averaged partially.
}

