% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barton_Choubey.R
\name{Barton_Choubey}
\alias{Barton_Choubey}
\title{Barton-Choubey criterion}
\usage{
Barton_Choubey(JRC, JCS, phi.r, unit.weight, depth)
}
\arguments{
\item{JRC}{The Joint Roughness Coefficient (1-20)}

\item{JCS}{The Joint Compressive Strength in MPa}

\item{phi.r}{The residual friction angle of the rock}

\item{unit.weight}{The unit weight of the rock in kN/m3}

\item{depth}{The depth, in meters, to which calculate the parameters}
}
\value{
A list with 3 tibbles: data (normal stress and shear stress), stress level point, and parameters (c and phi). All stress values are in MPa
}
\description{
Calculates the Barton-Choubey criterion for the given data and for the given level of stress (unit weight and depth).
}
\examples{
JRC = 10
JCS = 30
phi.r = 26
unit.weight = 18.6
depth = 40
Barton_Choubey(JRC, JCS, phi.r, unit.weight, depth)

}
\references{
Barton, N. & Choubey, V. (1977). The shear strength of rock joints in theory and practice. Rock Mechanichs, 10: 1-54.
}
