% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barton_Choubey_plot.R
\name{Barton_Choubey_plot}
\alias{Barton_Choubey_plot}
\title{Barton-Choubey criterion plot}
\usage{
Barton_Choubey_plot(BC, units = "kPa")
}
\arguments{
\item{BC}{The result from running \code{Barton_Choubey()}}

\item{units}{The units to use in the plot, either "kPa" or "MPa". Default is "kPa". If "MPa" is desired you can use units = "" or units = "MPa"}
}
\value{
A ggplot object for the Barton-Choubey criterion
}
\description{
Draws the Barton-Choubey criterion for the given data and for the given level of stress (unit weight and depth).
}
\examples{
JRC = 10
JCS = 30
phi.r = 26
unit.weight = 18.6
depth = 40
BC = Barton_Choubey(JRC, JCS, phi.r, unit.weight, depth)
Barton_Choubey_plot(BC)

}
\references{
Barton, N. & Choubey, V. (1977). The shear strength of rock joints in theory and practice. Rock Mechanichs, 10: 1-54.
}
