% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearing_capacity.R
\name{bearing_capacity}
\alias{bearing_capacity}
\title{Bearing capacity}
\usage{
bearing_capacity(B, D, L = NULL, gamma.h, gamma.s, tau0, phi, wl, FS,
  footing = c("strip", "square", "rectangular", "circular"))
}
\arguments{
\item{B}{The footing's width (in meters)}

\item{D}{The footing's embedment depth (in meters)}

\item{L}{The footing's length for a rectangular footing (in meters)}

\item{gamma.h}{The wet unit weight of the soil (in kN/m3)}

\item{gamma.s}{The saturated unit weight of the soil (in kN/m3)}

\item{tau0}{The soil's cohesion (in kPa)}

\item{phi}{The soil's friction angle}

\item{wl}{The depth to the water level (in meters)}

\item{FS}{The Factor of Safety to use for the calculation of qa}

\item{footing}{Type of footing for which to calculate the bearing capacity (Default is "strip")}
}
\value{
A data frame with the solution (Depth in rows, Width in columns) and the respective plot
}
\description{
Calculates and plots the allowable bearing capacity for different footings.
}
\details{
The \code{B} and \code{D} parameters can be vectors for multiple cases comparisons or single values for a single case estimates. If  \code{FS = 1} then \code{qa = qu}.
For a total stress analysis (TSA) in a cohesive soil (plastic silts and clays) set the friction angle equal to zero (\code{phi = 0}) and the cohesion equal to the undrained shear strength (\code{tau0 = Su}).
For an effective stress analysis (ESA) in a cohesive soil use the effective friction angle and effective cohesion.
For a coarse-grained soil (gravels, sands, and non-plastic silts) usually TSA = ESA, and the friction angle and cohesion should be used, and if the material has no-cohesion then set cohesion equal to zero (tau0 = 0)
}
\examples{
B = seq(0.5, 2, 0.25)
D = seq(0, 2, 0.25)
L = NULL
gamma.h = 15.5
gamma.s = 18.5
tau0 = 10
phi = 30
FS = 3
wl = 1
bearing_capacity(B, D, L, gamma.h, gamma.s, tau0, phi, wl, FS)
bearing_capacity(B, D, L, gamma.h, gamma.s, tau0, phi, wl, FS, footing = "square")
bearing_capacity(B, D, L = 3, gamma.h, gamma.s, tau0, phi, wl, FS, footing = "rectangular")

}
\references{
Day, R. W. (2010). Foundation Engineering Handbook. McGraw Hill.
}
