% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs.R
\name{boxGrob}
\alias{boxGrob}
\title{Create a box with text}
\usage{
boxGrob(label, y = unit(0.5, "npc"), x = unit(0.5, "npc"), width, height,
  just = "center", bjust = "center", txt_gp = gpar(color = "black"),
  box_gp = gpar(fill = "#D8F0D1"))
}
\arguments{
\item{label}{The label to print}

\item{y}{The y position to put the box at. Can be either in \code{npc} (i.e. 0-1) or a \code{\link[grid]{unit}}.}

\item{x}{The x position to put the box at. Can be either in \code{npc} (i.e. 0-1) or a \code{\link[grid]{unit}}.}

\item{just}{The justification for the text: left, center or right.}

\item{bjust}{The justification for the box: left, center, right, top or bottom.
See the \code{just} option for the \code{\link[grid]{viewport}}}

\item{txt_gp}{The \code{\link[grid]{gpar}} style to apply to the text.}

\item{box_gp}{The \code{\link[grid]{gpar}} style to apply to the box.}
}
\value{
A grob
}
\description{
Creates a grob box with text inside it.
}
\examples{
library(grid)
grid.newpage()
boxGrob("My box")
}
\seealso{
Other box-functions: \code{\link{boxPropGrob}},
  \code{\link{connectGrob}}
}

