% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateLinesAndArrow}
\alias{calculateLinesAndArrow}
\title{Gets offsetted lines}
\usage{
calculateLinesAndArrow(x, y, offset, end_x = -1, end_y = -1,
  arrow_offset = -1)
}
\arguments{
  \item{x}{A numeric vector containing all the x-elements}

  \item{y}{A numeric vector containing all the y-elements}

  \item{offset}{The offset to add to the line}

  \item{end_x}{The x end of the line where the arrow
  occurrs (if < 0 arrow is skipped)}

  \item{end_y}{The y end of the line where the arrow
  occurrs (if < 0 arrow is skipped)}

  \item{arrow_offset}{The offset to add to the arrow
  section if any (if <= 0 arrow is skipped)}
}
\value{
\code{list(list(x = ..., y = ...))} Returns a list with the
right/left lines that in turn lists with \emph{x} and
\emph{y} elements
}
\description{
The function calculates new points according to the offset
that lie to the left/right of the provided line.
}

