% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carolina.R
\docType{data}
\name{carolina}
\alias{carolina}
\title{Data for directional examples}
\format{A vector or angular measurements}
\description{
Non-directional data of 99 "Carolina bays" - ellipsoidal depressions on southern USA.
}
\references{
Davis, J. C. (2002). Statistical and Data Analysis in Geology. 3rd ed. John Wiley & Sons.
}
\keyword{datasets}
