% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{chunk}
\alias{chunk}
\title{Split a vector into \code{n} size chunks.}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{vector.}

\item{n}{numeric value. Chunk size.}
}
\description{
Split a vector into \code{n} size chunks.
}
