% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_chisq.R
\name{ci_chisq}
\alias{ci_chisq}
\title{Confidence interval for standard deviation of a sample}
\usage{
ci_chisq(s, n, conf.level = 0.95, digits = 2)
}
\arguments{
\item{s}{Sample standard deviation}

\item{n}{Sample size}

\item{conf.level}{Confidence level to use for the confidence interval (Default is 0.95)}

\item{digits}{Number of digits to round to (Default is 2)}
}
\value{
A tibble with the standard deviation and variance, degrees of freedom, and lower and upper ends of the confidence interval
}
\description{
Calculates the confidence interval for the standard deviation of a sample (ChiSquare distribution).
}
\examples{
s <- 0.535
n <- 10
ci_chisq(s, n)

}
