% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_t.R
\name{ci_t}
\alias{ci_t}
\title{Confidence interval for mean (unknown sigma)}
\usage{
ci_t(x, s, n, conf.level = 0.95, digits = 2)
}
\arguments{
\item{x}{Sample mean}

\item{s}{Sample standard deviation}

\item{n}{Sample size}

\item{conf.level}{Confidence level to use for the confidence interval (Default is 0.95)}

\item{digits}{Number of digits to round to (Default is 2)}
}
\value{
A data frame with the mean, degrees of freedom, and lower and upper ends of the confidence interval
}
\description{
Calculates the confidence interval for the mean with unknown standard deviation, using the t-statistic (t distribution).
}
\details{
It can work for dependent (paired) samples using the mean and standard deviation of differences
}
\examples{
x <- 80
s <- 15
n <- 20
ci_t(x, s, n)

}
