% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_t2.R
\name{ci_t2}
\alias{ci_t2}
\title{Confidence interval for difference of the mean between 2 samples (unknown sigmas)}
\usage{
ci_t2(x1, s1, n1, x2, s2, n2, var.equal = FALSE, conf.level = 0.95, digits = 2)
}
\arguments{
\item{x1}{Sample's 1 mean}

\item{s1}{Sample's 1 standard deviation}

\item{n1}{Sample size of sample 1}

\item{x2}{Sample's 2 mean}

\item{s2}{Sample's 2 standard deviation}

\item{n2}{Sample size of sample 2}

\item{var.equal}{Logical indicating if the variances are equal or not}

\item{conf.level}{Confidence level to use for the confidence interval (Default is 0.95)}

\item{digits}{Number of digits to round to (Default is 2)}
}
\value{
A tibble with the difference of means, degrees of freedom, and lower and upper ends of the confidence interval
}
\description{
Calculates the confidence interval for the difference of the mean between 2 samples (t distribution).
}
\examples{
x1 <- 90
s1 <- 15
n1 <- 20
x2 <- 80
s2 <- 8
n2 <- 30
ci_t2(x1, s1, n1, x2, s2, n2)

}
