% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_Z.R
\name{ci_Z}
\alias{ci_Z}
\title{Confidence interval for mean of a population}
\usage{
ci_Z(mu, sig, n, conf.level = 0.95)
}
\arguments{
\item{mu}{A point estimate of the population mean (a vector of values can also be used)}

\item{sig}{A point estimate of the population standard deviation (a vector of values can also be used)}

\item{n}{Sample size from the population (a vector of values can also be used)}

\item{conf.level}{Confidence level to use for the confidence interval (Default is 0.95)}
}
\value{
A data frame with the mean, and lower and upper ends of the confidence interval
}
\description{
Calculates the confidence interval for a population mean with known mean and standard deviation, using the Z-statistic (normal distribution).
}
\examples{
mu <- 80
sig <- 15
n <- 20
ci_Z(mu, sig, n)

}
