% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_z2.R
\name{ci_z2}
\alias{ci_z2}
\title{Confidence interval for difference of the mean between 2 samples (known sigmas)}
\usage{
ci_z2(x1, sig1, n1, x2, sig2, n2, conf.level = 0.95, digits = 2)
}
\arguments{
\item{x1}{Sample's 1 mean}

\item{sig1}{Population's 1 standard deviation}

\item{n1}{Sample size of sample 1}

\item{x2}{Sample's 2 mean}

\item{sig2}{Population's 2 standard deviation}

\item{n2}{Sample size of sample 2}

\item{conf.level}{Confidence level to use for the confidence interval (Default is 0.95)}

\item{digits}{Number of digits to round to (Default is 2)}
}
\value{
A data frame with the difference of means, and lower and upper ends of the confidence interval
}
\description{
Calculates the confidence interval for the difference of the mean between 2 samples (Z distribution).
}
\examples{
x1 <- 42
sig1 <- 8
n1 <- 75
x2 <- 36
sig2 <- 6
n2 <- 50
ci_z2(x1, sig1, n1, x2, sig2, n2)

}
