% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPW.R
\name{CIPW}
\alias{CIPW}
\title{CIPW Norm}
\usage{
CIPW(data)
}
\arguments{
\item{data}{A dataframe (tibble) with the major oxides in the first column and their percent weight in the second column}
}
\value{
Summary table for the CIPW norm, with all the minerals and their percent weight, as well as separate tables for the ferromagentic minerals and its components
}
\description{
\code{CIPW()} calculates the mineral composition of an igneous rock sample, based on ther major oxides and weight percent
}
\details{
The names of the oxides must be: SiO2, TiO2, Al2O3, Fe2O3, FeO, MnO, MgO, CaO, Na2O, K2O, P2O5. If any of the major oxides is not present just use 0 for their percent weight
}
\examples{
d = data.frame(O = c('SiO2', 'TiO2', 'Al2O3', 'Fe2O3', 'FeO',
                     'MnO', 'MgO', 'CaO', 'Na2O', 'K2O', 'P2O5'),
               W = c(35.6,5.98,12.9,7.68,9.28,.05,
                     5.4,8.46,8.35,2.78,2.13))
CIPW(d)

}
