% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDA_2Group_H2.R
\name{CoDA_2Group_H2}
\alias{CoDA_2Group_H2}
\title{CoDA Two Group Hypothesis 2}
\usage{
CoDA_2Group_H2(comp1, comp2)
}
\arguments{
\item{comp1}{A matrix of observations for composition 2.}
}
\value{
A tibble with the statistic (Q), degrees of freedom (nu), p-value, and null hypothesis (H0)
}
\description{
Performs the calculations for the hypothesis of equal covariance with different centers for two diferent samples, using Compositional Data Analysis (CoDA) principles.
}
\examples{
data("Hongite")
data("Kongite")
CoDA_2Group_H2(Hongite,Kongite)

}
\references{
Pawlowsky-Hlahn, V., Egozcue, J.J & Tolosna-Delgado, R. (2015). Modeling and analysis of compositional data. John Wiley & Sons.

Aitchison, J. (1986). The statistical analysis of compositional data. Chapman and Hall.
}
