% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDA_Atyp_Idx.R
\name{CoDA_Atyp_Idx}
\alias{CoDA_Atyp_Idx}
\title{CoDA Atypicality Index}
\usage{
CoDA_Atyp_Idx(comp, sample = NULL)
}
\arguments{
\item{comp}{A matrix of observations for a given composition. Entries must be non-zero and positive.}

\item{sample}{A one sample vector to compare against the given composition.}
}
\value{
A matrix with the atypicality index or indices.
}
\description{
Calculates the atypicality index (ranges between 0 and 1, 1 being more atypical/different) for all the observations of a composition or compares one sample to the given composition, using Compositional Data Analysis (CoDA) principles.
}
\examples{
data("Hongite", package = 'compositions')
samp = c(44, 20.4, 13.9, 9.1, 12.6)
CoDA_Atyp_Idx(Hongite)
CoDA_Atyp_Idx(Hongite, samp)

}
\references{
Aitchison, J. (1986). The statistical analysis of compositional data. Chapman and Hall.
}
