% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_var.R
\name{coef_var}
\alias{coef_var}
\title{Coefficient of variation}
\usage{
coef_var(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The coefficient of variation for the given vector
}
\description{
Calculates the coefficient of variation, disregarding NAs.
}
\examples{
x = rnorm(50, 150, 30)
coef_var(x)
z = data.frame(N1 = rnorm(50, 150, 30), N2 = rnorm(50, 50, 10))
apply(z, 2, coef_var)

}
