% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertShowMissing.R
\name{convertShowMissing}
\alias{convertShowMissing}
\title{A function for converting a show_missing variable.}
\usage{
convertShowMissing(show_missing)
}
\arguments{
\item{show_missing}{Boolean or "no", "ifany", "always"}
}
\value{
string
}
\description{
The variable is supposed to be directly compatible with
\code{\link[base]{table}}(..., useNA = show_missing). It throws an error
if not compatible. It is mostly useful for custom describe functions.
}
\details{
\emph{Deprecated:} This function will be deprecated
as all functions now use the useNA style in order to comply
with standard R naming.
}
