% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs_coords.R
\name{coords}
\alias{coords}
\title{Get the box coordinates}
\usage{
coords(box)
}
\arguments{
\item{box}{The \code{\link{boxGrob}} or \code{\link{boxPropGrob}}}
}
\value{
A list with the coordinates
}
\description{
Retrieves the boxes \code{"coords"} attribute.
}
\examples{
box <- boxGrob("A test box")
coords(box)
}
\seealso{
Other flowchart components:
\code{\link[=align]{align()}},
\code{\link[=append]{append()}},
\code{\link[=boxGrob]{boxGrob()}},
\code{\link[=boxHeaderGrob]{boxHeaderGrob()}},
\code{\link[=boxPropGrob]{boxPropGrob()}},
\code{\link{boxShapes}},
\code{\link[=connectGrob]{connectGrob()}},
\code{\link[=distance]{distance()}},
\code{\link[=equalizeHeights]{equalizeHeights()}},
\code{\link[=equalizeWidths]{equalizeWidths()}},
\code{\link[=flowchart]{flowchart()}},
\code{\link[=insert]{insert()}},
\code{\link[=move]{move()}},
\code{\link[=moveBox]{moveBox()}},
\code{\link[=phaseLabel]{phaseLabel()}},
\code{\link[=spread]{spread()}}
}
\concept{flowchart components}
