% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_aic_eta.R
\name{cp_aic_eta}
\alias{cp_aic_eta}
\title{Changepoint analysis of different layer models}
\usage{
cp_aic_eta(x, m = 10, nl = 3)
}
\arguments{
\item{x}{A data frame containing the location variable (depth or distance) in the first column, and the value of interest in the second column}

\item{m}{The maximum number of breakpoints (# layers - 1) to look for}

\item{nl}{The minimum number of points per layer to be considered}
}
\value{
A ggplot and plotly objects showing the AIC and eta-squared statistics, and a data frame with all the data and possible layer models
}
\description{
Calculates and plots the AIC and eta-squared statistics for diferent layer models based on a changepoint analysis using the mean and variance.
}
\details{
The example data given is intended to show the structure needed for input data. The user should follow this structure, which in general corresponds with a data frame with a sequence in the first column and the observed/measured values in the second column
}
\examples{
cp_aic_eta(DPM_data, m = 10, nl = 3)

}
