% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_layers.R
\name{cp_layers}
\alias{cp_layers}
\title{Plots a layered model and summarizes the statistics for each layer from a Changepoint result}
\usage{
cp_layers(x, breaks)
}
\arguments{
\item{x}{The resulting objecto from \code{cp_aic_eta()}}

\item{breaks}{An integer giving the number of breakpoints to use (from 'Changepoint')}
}
\value{
A ggplot and plotly objects showing the layered model, another showing the confidence intervals, and a summary table
}
\description{
Given a number of breakpoints, plots a layered model of one variable against distance, plots the confidence intervals for each layer, and gives a summary table.
}
\examples{
cp = cp_aic_eta(DPM_data, m = 10, nl = 3)
cp_layers(cp, breaks = 2)

}
