% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPTu_data.R
\docType{data}
\name{CPTu_data}
\alias{CPTu_data}
\title{Example data for calculation of T2-statistic}
\format{A data frame with results from a CPTu survey. Depth in meters in the first column, and the measured variables in the next columns: point resistance (qc), sleeve friction (fs), and pore-water pressure (u)}
\description{
Example data from a CPTu survey.
}
\keyword{datasets}
