% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{cut_character}
\alias{cut_character}
\title{cuts a character vector using a set of character values}
\usage{
cut_character(x, v, align = c("left", "right"))
}
\arguments{
\item{x}{character vector}

\item{v}{character vector}

\item{align}{character, either 'right' or 'left'}
}
\value{
a character vector
}
\description{
cuts a character vector using a set of character values
}
\examples{
v <- letters[seq(2,length(letters),2)]
 cut(letters[1:10], v, 'right')
 cut(letters[1:10], v, 'left')
}

