% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degs.R
\name{degs}
\alias{degs}
\title{Degrees}
\usage{
degs(x)
}
\arguments{
\item{x}{Angle in radians}
}
\value{
Angle in degrees
}
\description{
Converts from radians to degrees.
}
\examples{
degs(0.7854)

}
