% Generated by roxygen2 (4.0.2): do not edit by hand
\name{descMerger}
\alias{descMerger}
\title{Prepares a matrix for htmlTable from a list}
\usage{
descMerger(tlist, ...)
}
\arguments{
  \item{tlist}{A list where each element comes from
  \code{\link{getDescriptionStatsBy}}. You can also provide
  pure output from the \code{\link{getDescriptionStatsBy}}
  function and have the function merge this together with
  the \code{...} argument. \emph{Note} that all elements
  myst have the same \code{by} argument or you will not be
  able to merge it into a list.}

  \item{...}{Additional lists/mtrx to be merged with the
  \code{tlist}}
}
\value{
\code{matrix} Returns a matrix object of class descList
}
\description{
By putting all the output from the
\code{\link{getDescriptionStatsBy}} into a list, naming
each element that we want in an rgroup we can automatically
merge everything and create an object ready for the
htmlTable.
}
\examples{
data(mtcars)

label(mtcars$mpg) <- "Gas"
units(mtcars$mpg) <- "Miles/(US) gallon"

label(mtcars$wt) <- "Weight"
units(mtcars$wt) <- "10<sup>3</sup> kg" # not sure the unit is correct

mtcars$am <- factor(mtcars$am, levels=0:1, labels=c("Automatic", "Manual"))
label(mtcars$am) <- "Transmission"

mtcars$gear <- factor(mtcars$gear)
label(mtcars$gear) <- "Gears"

# Make up some data for making it slightly more interesting
mtcars$col <- factor(sample(c("red", "black", "silver"),
                     size=NROW(mtcars), replace=TRUE))
label(mtcars$col) <- "Car color"

htmlTable(descMerger(getDescriptionStatsBy(mtcars$mpg, mtcars$am,
                                           header_count = TRUE,
                                           use_units = TRUE),
                     getDescriptionStatsBy(mtcars$wt, mtcars$am,
                                           header_count = TRUE,
                                           use_units = TRUE)),
          caption  = "Basic continuous stats from the mtcars dataset")

tll <- list()
tll[["Gear (3 to 5)"]] <- getDescriptionStatsBy(mtcars$gear, mtcars$am)
tll <- c(tll,
         list(getDescriptionStatsBy(mtcars$col, mtcars$am)))

htmlTable(descMerger(tll),
          caption  = "Factored variables")

tl_no_units <- list()
tl_no_units[["Gas (mile/gallons)"]] <-
  getDescriptionStatsBy(mtcars$mpg, mtcars$am,
                        header_count = TRUE)
tl_no_units[["Weight (10<sup>3</sup> kg)"]] <-
  getDescriptionStatsBy(mtcars$wt, mtcars$am,
                        header_count = TRUE)
htmlTable(descMerger(tl_no_units, tll),
          rgroupCSSstyle = "")

# A little more advanced
mtcars$mpg[sample(1:NROW(mtcars), size=5)] <- NA
getDescriptionStatsBy(mtcars$mpg, mtcars$am, statistics=TRUE)

# Do the horizontal version
getDescriptionStatsBy(mtcars$col, mtcars$am,
                      statistics=TRUE, hrzl_prop = TRUE)

mtcars$wt_with_missing <- mtcars$wt
mtcars$wt_with_missing[sample(1:NROW(mtcars), size=8)] <- NA
getDescriptionStatsBy(mtcars$wt_with_missing, mtcars$am, statistics=TRUE,
                      hrzl_prop = TRUE, total_col_show_perc = FALSE)


mtcars$col_with_missing <- mtcars$col
mtcars$col_with_missing[sample(1:NROW(mtcars), size=5)] <- NA
getDescriptionStatsBy(mtcars$col_with_missing, mtcars$am, statistics=TRUE,
                      hrzl_prop = TRUE, total_col_show_perc = FALSE)

}

