% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptionStats.R
\name{describeMean}
\alias{describeMean}
\title{Describe the mean}
\usage{
describeMean(
  x,
  html = TRUE,
  digits = 1,
  digits.nonzero = NA,
  number_first = TRUE,
  useNA = c("ifany", "no", "always"),
  useNA.digits = digits,
  percentage_sign = TRUE,
  plusmin_str,
  language = "en",
  ...
)
}
\arguments{
\item{x}{The variable that you want the statistics for}

\item{html}{If HTML compatible output should be used. If \code{FALSE}
it outputs LaTeX formatting}

\item{digits}{The number of decimals used}

\item{digits.nonzero}{The number of decimals used for values that are close to zero}

\item{number_first}{If the number should be given or if the percentage
should be presented first. The second is encapsulated in parentheses ().
This is only used together with the useNA variable.}

\item{useNA}{This indicates if missing should be added as a separate
row below all other. See \code{\link[base]{table}} for \code{useNA}-options.
\emph{Note:} defaults to ifany and not "no" as \code{\link[base]{table}} does.}

\item{useNA.digits}{The number of digits to use for the
missing percentage, defaults to the overall \code{digits}.}

\item{percentage_sign}{If you want to suppress the percentage sign you
can set this variable to FALSE. You can also choose something else that
the default \% if you so wish by setting this variable. Note, this is
only used when combined with the missing information.}

\item{plusmin_str}{Provide if you want anything other than the plus minus sign
suited for the given output format.}

\item{language}{The ISO-639-1 two-letter code for the language of
interest. Currently only english is distinguished from the ISO
format using a ',' as the separator in the \code{\link[htmlTable]{txtInt}}
function.}

\item{...}{Passed on to \code{\link{describeFactors}}}
}
\value{
\code{string} Returns a string formatted for either LaTeX or HTML
}
\description{
A function that returns a description of a continuous variable using
the mean together with the standard deviation. The standard deviation
is used as it is "industry standard" to use mean with standard deviation
and not because it's the only option.
}
\examples{
describeMean(1:10)
describeMean(c(1:10, NA), useNA = "always")
describeMean(c(1:10, NA), useNA = "no")
}
\seealso{
\code{\link{getDescriptionStatsBy}}

Other descriptive functions:
\code{\link[=describeFactors]{describeFactors()}},
\code{\link[=describeMedian]{describeMedian()}},
\code{\link[=describeProp]{describeProp()}},
\code{\link[=getDescriptionStatsBy]{getDescriptionStatsBy()}},
\code{\link[=getPvalWilcox]{getPvalWilcox()}}
}
\concept{descriptive functions}
