% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{df2v}
\alias{df2v}
\title{converts an n x 1 data frame into a vector}
\usage{
df2v(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
numerical vectors
}
\description{
converts an n x 1 data frame into a vector
}
\author{
G.A.Paleologo
}

