% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomy.R
\name{dichotomy}
\alias{dichotomy}
\title{Splits a vector into two classes}
\usage{
dichotomy(x, criteria = c("up-down", "median"))
}
\arguments{
\item{x}{A numeric vector of values to split}
}
\value{
A vector with the splitted classes
}
\description{
Splits (dichotomizes) a numeric vector into two classes using the "up-down" or "median" criteria. The results is intendend to be used in a runs test for tendency.
}
\examples{
x1 = c(17, 14, 13, 17, 9, 15, 36, 13, 26, 10, 25, 13, 25, 8, 28, 15, 19, 32, 22)
x2 = c(17, 14, 14, 17, 9, 15, 36, 13, 26, 10, 25, 13, 25, 8, 28, 15, 19, 32, 22)
dichotomy(x1)
dichotomy(x2)
dichotomy(x2, criteria = "median")

}
\references{
Davis, J. C. (2002). Statistical and Data Analysis in Geology. 3rd ed. John Wiley & Sons.
}
