% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{Dict}
\alias{Dict}
\alias{dict}
\title{R6 Based Key-Value Dictionary Implementation}
\usage{
dict(...)
}
\arguments{
\item{...}{Any length of key and value pairs. If you would like to use
a not valid R name as a key, you must wrap it by backquotes or convert it
using \code{\link{make.names}}.}
}
\value{
A \code{Dict} class object.
}
\description{
A key-value dictionary data structure based on R6 class which is designed to
be similar usages with other languages dictionary (e.g. Python) with
reference semantics and extendabilities by R6.
}
\examples{
ages <- Dict$new(
  Charlie = 40L,
  Alice = 30L,
  Bob = 25L
)
ages <- dict(
  Charlie = 40L,
  Alice = 30L,
  Bob = 25L
)

}
\section{Super class}{
\code{\link[gmisc:Mapping]{gmisc::Mapping}} -> \code{Dict}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Dict-clone}{\code{Dict$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="gmisc" data-topic="Mapping" data-id="as_list"><a href='../../gmisc/html/Mapping.html#method-Mapping-as_list'><code>gmisc::Mapping$as_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gmisc" data-topic="Mapping" data-id="clear"><a href='../../gmisc/html/Mapping.html#method-Mapping-clear'><code>gmisc::Mapping$clear()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gmisc" data-topic="Mapping" data-id="get"><a href='../../gmisc/html/Mapping.html#method-Mapping-get'><code>gmisc::Mapping$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gmisc" data-topic="Mapping" data-id="has"><a href='../../gmisc/html/Mapping.html#method-Mapping-has'><code>gmisc::Mapping$has()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gmisc" data-topic="Mapping" data-id="initialize"><a href='../../gmisc/html/Mapping.html#method-Mapping-initialize'><code>gmisc::Mapping$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gmisc" data-topic="Mapping" data-id="keys"><a href='../../gmisc/html/Mapping.html#method-Mapping-keys'><code>gmisc::Mapping$keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gmisc" data-topic="Mapping" data-id="pop"><a href='../../gmisc/html/Mapping.html#method-Mapping-pop'><code>gmisc::Mapping$pop()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gmisc" data-topic="Mapping" data-id="remove"><a href='../../gmisc/html/Mapping.html#method-Mapping-remove'><code>gmisc::Mapping$remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gmisc" data-topic="Mapping" data-id="set"><a href='../../gmisc/html/Mapping.html#method-Mapping-set'><code>gmisc::Mapping$set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gmisc" data-topic="Mapping" data-id="update"><a href='../../gmisc/html/Mapping.html#method-Mapping-update'><code>gmisc::Mapping$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gmisc" data-topic="Mapping" data-id="values"><a href='../../gmisc/html/Mapping.html#method-Mapping-values'><code>gmisc::Mapping$values()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dict-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Dict-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dict$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
