% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_2sample_test_3D.R
\name{dir_2sample_test_3D}
\alias{dir_2sample_test_3D}
\title{Directional statistics two-sample test (3D)}
\usage{
dir_2sample_test_3D(d1, i1, d2, i2, conf.level = 0.95)
}
\arguments{
\item{d1}{A vector of dip direction measurements in degrees for sample 1}

\item{i1}{A vector of dip measurements in degrees for sample 1}

\item{d2}{A vector of dip direction measurements in degrees for sample 2}

\item{i2}{A vector of dip measurements in degrees for sample 2}

\item{conf.level}{Confidence level for the F-statistic calculation and p-value interpretation (Default is 0.95)}
}
\value{
A list with the F-statistic, the critical value of F, the degrees of freedom, the p-value, and the interpretation of these values
}
\description{
Performes the two-sample test, for the 3D case, for a pair of angular measurements to determine if they could belong to the same population (event).
}
\examples{
d1 = c(12,18,22,15,10,20)
i1 = c(42,40,48,30,42,30)
d2 = c(111.7,109.2,185.1,83.1,235.2,226.7)
i2 = c(40.4,31.4,25.9,26.2,27,31.5)
dir_2sample_test_3D(d1, i1, d2, i2)

}
\references{
Borradaile, G. (2003). Statistics of Earth Science Data. Springer.
}
