% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_stats_2D.R
\name{dir_stats_2D}
\alias{dir_stats_2D}
\title{Directional statistics for the 2D case}
\usage{
dir_stats_2D(x, dir = 1, conf.level = 0.95)
}
\arguments{
\item{x}{A vector of angular measurements in degrees}

\item{dir}{A logical indicating if the data are directional (1) or non-directional (0)}

\item{conf.level}{Confidence level to use for the cone of confidence (Default is 0.95)}
}
\value{
A data frame with the mean direction, mean resultant length, circular variance, concentration parameter, and cone of confidence
}
\description{
Calculates the directional (circular) statistics of a sample of directional measurements for the 2D case (just the dip direction or strike).
}
\details{
Directional data refers to dipping planes or lines, data expressed as dip direction (trend). Non-directional data refers to strike or lineations expressed as azimuths that can take two angles (i.e. 45 or 225)
}
\examples{
x <- c(255, 239, 222, 231, 199, 271, 222, 274, 228,
       246, 177, 199, 257, 201, 237, 209, 216)
dir_stats_2D(x)
# non-directional data example
dir_stats_2D(carolina, dir = 0)

}
\references{
Davis, J. C. (2002). Statistical and Data Analysis in Geology. 3rd ed. John Wiley & Sons.

Swan, A. R. H. & Sandilands, M. (1995). Introduction to Geological Data Analysis. Blackwell Science.

Borradaile, G. (2003). Statistics of Earth Science Data. Springer.
}
