% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_stats_ps.R
\name{dir_stats_ps}
\alias{dir_stats_ps}
\title{Principal stresses}
\usage{
dir_stats_ps(dir, dip)
}
\arguments{
\item{dir}{A vector of dip directions measurements in degrees}

\item{dip}{A vector of dip angles measurements in degrees}
}
\value{
A data frame with the dip direction and dip angle for the principal stresses
}
\description{
Calculates the the principal stresses for a set of angular measurements.
}
\details{
This always considers a plane expressed as dip direction and dip angle, it could be a bedding plane or a fault plane
}
\examples{
dir = c(12,18,22,15,10,20)
dip = c(42,40,48,30,42,30)
dir_stats_ps(dir, dip)

}
