% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_unif_test_2D.R
\name{dir_unif_test_2D}
\alias{dir_unif_test_2D}
\title{Directional statistics test of uniformity (2D)}
\usage{
dir_unif_test_2D(x, dir = 1, conf.level = 0.95)
}
\arguments{
\item{x}{A vector of angular measurements in degrees}

\item{dir}{A logical indicating if the data are directional (1) or non-directional (0)}

\item{conf.level}{Confidence level to use for the critical value (Default is 0.95, and see details section below)}
}
\value{
A list with the mean resultant length, the critical mean resultant length, and the interpretation of these values
}
\description{
Performes the uniformity test, for the 2D case, for a set of angle measurements to determine if they come from a uniform distribution (no prefered mean direction).
}
\details{
Directional data refers to dipping planes or lines, data expressed as dip direction (trend). Non-directional data refers to strike or lineations expressed as azimuths that can take two angles (i.e. 45 or 225). The options for confidence level are 0.99, 0.975, 0.95, or 0.90
}
\examples{
x = c(255, 239, 222, 231, 199, 271, 222, 274, 228, 246,
      177, 199, 257, 201, 237, 209, 216, 180, 182, 250,
      219, 196, 197, 246, 218, 235, 232, 243, 232, 180,
      231, 254, 242, 149, 212, 210, 230, 205, 220, 268)
dir_unif_test_2D(x)

}
\references{
Davis, J. C. (2002). Statistical and Data Analysis in Geology. 3rd ed. John Wiley & Sons.

Swan, A. R. H. & Sandilands, M. (1995). Introduction to Geological Data Analysis. Blackwell Science.

Borradaile, G. (2003). Statistics of Earth Science Data. Springer.
}
