% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_unif_test_3D.R
\name{dir_unif_test_3D}
\alias{dir_unif_test_3D}
\title{Directional statistics test of uniformity (3D)}
\usage{
dir_unif_test_3D(dir, dip, conf.level = 0.95)
}
\arguments{
\item{dir}{A vector of dip directions measurements in degrees}

\item{dip}{A vector of dip angles measurements in degrees}

\item{conf.level}{Confidence level to use for the critical value (Default is 0.95, and see details section below)}
}
\value{
A list with the mean resultant length, the critical mean resultant length, and the interpretation of these values
}
\description{
Performes the uniformity test, for the 3D case, for a set of angle measurements to determine if they come from a uniform distribution (no prefered mean direction).
}
\details{
The options for confidence level are 0.99, 0.975, 0.95, or 0.90
}
\examples{
dir = c(12,18,22,15,10,20)
dip = c(42,40,48,30,42,30)
dir_unif_test_3D(dir, dip)

}
\references{
Davis, J. C. (2002). Statistical and Data Analysis in Geology. 3rd ed. John Wiley & Sons.

Swan, A. R. H. & Sandilands, M. (1995). Introduction to Geological Data Analysis. Blackwell Science.

Borradaile, G. (2003). Statistics of Earth Science Data. Springer.
}
