% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{dt2m}
\alias{dt2m}
\title{converts a data table into a matrix}
\usage{
dt2m(x, key = NULL)
}
\arguments{
\item{x}{data table}

\item{key}{character, optional name of the column to be used}
}
\value{
a matrix
}
\description{
converts a data table into a matrix by using a column
(by default, the first) to set the matrix row names.
The column is removed from the original data frame
}
\author{
G.A.Paleologo
}

