% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Entropy for a given split of a categorical variable}
\usage{
entropy(x, group, thres)
}
\arguments{
\item{x}{A numeric vector}

\item{group}{Categorical variable}

\item{thres}{Threshold value to use for the split}
}
\value{
A data frame with entropy before and after the split, as well as the gain and gain percent
}
\description{
Calculates the entropy of a categorical variable for a given split (threshold) value of a numeric variable.
}
\examples{
entropy(mtcars$mpg, mtcars$cyl, 15)
entropy(mtcars$mpg, mtcars$cyl, 21)
entropy(mtcars$mpg, mtcars$cyl, 25)

}
\references{
Kuhn, M. & Johnson, K. (2013). Applied Predictive Modeling. Springer.

Witten, I. H., Frank, E. & Hall, M. A. (2011). Data Mining: Practical Machine Learning Tools and Techniques. Elsevier.
}
