% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{enumerate}
\alias{enumerate}
\title{Create a list with the index and values of each element passed into ...}
\usage{
enumerate(..., .start = 1, .step = 1)
}
\arguments{
\item{...}{vectors or lists to be combined. They should be the same length.}

\item{.start}{numeric value. The first index value.}

\item{.step}{numeric value. Step size for each iteration.}
}
\description{
Create a list with the index and values of each element passed into ...
}
