% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figCaptionNo.R
\name{figCapNoNext}
\alias{figCapNoNext}
\title{Gets the next figure caption number}
\usage{
figCapNoNext(roman = getOption("fig_caption_no_roman", default = FALSE))
}
\arguments{
\item{roman}{Whether or not to use roman numbers instead
of Arabic. Can also be set through \code{options(fig_caption_no_roman = TRUE)}}
}
\description{
The function relies on \code{options("fig_caption_no")}
in order to keep track of the last number.
}
\examples{
org_opts <- options(fig_caption_no = 1)
figCapNoNext()
options(org_opts)
}
\seealso{
Other figure caption functions:
\code{\link[=figCapNo]{figCapNo()}},
\code{\link[=figCapNoLast]{figCapNoLast()}}
}
\concept{figure caption functions}
