% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{filter}
\alias{filter}
\title{Return only the values of a list that meet a certain predicate.}
\usage{
filter(x, f, negate = FALSE)
}
\arguments{
\item{x}{list. The list to filter.}

\item{f}{function. The functions should return TRUE or FALSE.}
}
\description{
Return only the values of a list that meet a certain predicate.
}
