% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{flatten}
\alias{flatten}
\title{Turn a list of lists into a list.}
\usage{
flatten(x, depth = -1)
}
\arguments{
\item{x}{list. The list of lists to unnest}
}
\description{
Turn a list of lists into a list.
}
