% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{flattenList}
\alias{flattenList}
\title{Converts a tree represented as a list as a data frame.}
\usage{
flattenList(x, colnames = NULL)
}
\arguments{
\item{x}{list}

\item{colnames}{character vector column names of the first column}
}
\value{
a data frame
}
\description{
Converts a tree represented as a list as a data frame.
}
\author{
G.A.Paleologo
}

