% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ftp_document}
\alias{ftp_document}
\title{Formatter wrapper for \code{\link[rmarkdown]{html_document}} facilitate easier LibreOffice porting}
\usage{
ftp_document(..., css = "custom.css", self_contained = FALSE,
  h1_style = "margin: 24pt 0pt 0pt 0pt;",
  other_h_style = "margin: 10pt 0pt 0pt 0pt;")
}
\arguments{
  \item{...}{Passed onto
  \code{\link[rmarkdown]{html_document}}.}

  \item{css}{Assumes that you have the ftp-specific
  \code{custom.css} file in the same folder as the knitr
  document from the Github
  \href{https://github.com/gforge/ftp/blob/master/custom.css}{gforge/ftp
  repository}. This can be over-ridden by providing your
  css files.}

  \item{self_contained}{Overrides the default \code{TRUE}
  for \code{\link[rmarkdown]{html_document}} to
  \code{FALSE} as LibreOffice can't hangs on long lines
  such as the base64 images included in the self-contained
  version.}

  \item{h1_style}{You can choose any css style formatting
  here that you want to be applied to all h1 elements.
  Note: this is only applied if LibreOffice_adapt is
  \code{TRUE}.}

  \item{other_h_style}{This is the formatting applied to
  any other h elements not included to the first. Note:
  this is only applied if LibreOffice_adapt is
  \code{TRUE}.}
}
\value{
R Markdown output format to pass to
\code{\link[rmarkdown]{render}}
}
\description{
This function adds the option of having adaptations needed
for seemless integration with word processor for importing
html-documents. The advantage of html documents is the
ability to create advanced formatting frequently needed in
publications and that is available in the
\code{\link{htmlTable}} function. You can view
\url{http://gforge.se/2014-07/fast-track-publishing-using-rmarkdown}{the
series} for more details regarding how to achieve
fast-track-publishing (ftp) together with knitr.
}
\examples{
# Possible yaml configuration at the top of the Rmd doc
\dontrun{
---
title: "Test"
author: "MG"
date: "Sunday, July 13, 2014"
output:
  Gmisc::rmd_ftp
---
}
}
\author{
Max Gordon
}

