% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bezier_private_functions.R
\name{getBezierAdj4Arrw}
\alias{getBezierAdj4Arrw}
\title{Gets the bezier points adjusted for an arrow}
\usage{
getBezierAdj4Arrw(end_points, spline_ctrl, arrow_length, internal.units, vp,
  rez_mltpl = 100)
}
\arguments{
  \item{end_points}{The start and end points}

  \item{spline_ctrl}{The spline control points}

  \item{arrow_length}{The desired length of the arrow}

  \item{internal.units}{The internal absolute unit that is
  used}

  \item{vp}{The viewport if any}

  \item{rez_mltpl}{Increases the resolution for the final
  bezier points by faking a larger line through
  multiplikation}
}
\value{
list
}
\description{
Gets the bezier points adjusted for an arrow
}

