\name{getBoundariesDataFrameFromCRR}
\alias{getBoundariesDataFrameFromCRR}
\title{Gets the boundaries for a fit}
\usage{
  getBoundariesDataFrameFromCRR(crr.fit, conf.int = 0.95)
}
\arguments{
  \item{crr.fit}{A model}

  \item{conf.int}{The interval of interest 0-1}
}
\value{
  A data frame with all the details
}
\description{
  TODO: Change the cofidence interval by hand into
  confint()
}
\author{
  max
}

