\name{getBoundariesDataFrameFromGlmFit}
\alias{getBoundariesDataFrameFromGlmFit}
\title{Gets the boundaries for a fit}
\usage{
  getBoundariesDataFrameFromGlmFit(glm.fit,
    conf.int = 0.95)
}
\arguments{
  \item{glm.fit}{A regression model}

  \item{conf.int}{The confidence interval 0-1}
}
\value{
  A frame with all the data
}
\description{
  TODO: Change the confidence interval calculation to
  confint()
}
\author{
  max
}

