\name{getGridVal}
\alias{getGridVal}
\title{Gets grid value}
\usage{
  getGridVal(x, default.units)
}
\arguments{
  \item{x}{Value}

  \item{default.units}{The unit type}
}
\value{
  float
}
\description{
  Returns the raw value in units if the provided is of type
  unit else it returns the default unit.
}
\author{
  Max
}

