% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getLineOffset}
\alias{getLineOffset}
\title{Gets offsetted lines}
\usage{
getLineOffset(x, y, offset)
}
\arguments{
  \item{x}{A numeric vector containing all the x-elements}

  \item{y}{A numeric vector containing all the y-elements}

  \item{offset}{The offset to add to the line}
}
\value{
\code{list(list(x = ..., y = ...))} Returns a list with the
right/left lines that in turn lists with \emph{x} and
\emph{y} elements
}
\description{
The function calculates new points according to the offset
that lie to the left/right of the provided line.
}

