% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getLines}
\alias{getLines}
\title{Gets the lines shifted according to width}
\usage{
getLines(bp, end_point, width, default.units, align_2_axis = TRUE)
}
\arguments{
  \item{bp}{Bezier points that define the line. Based upon
  a bezier grob.}

  \item{end_point}{A list with x & y for the end, only used
  for the last element's angle.}

  \item{width}{The width as a grid length
  \code{\link{unit}}}

  \item{default.units}{The default unit (see the grid
  package for available units)}

  \item{align_2_axis}{This indicates if the arrows origin
  should align it to an axis. Which is decided by which
  axis is the closes one.}
}
\value{
A list with left and right elements indicating the two
lines
}
\description{
The lines are the upper and the lower lines that will make
up the future polygon that will be used to generate the
arrow. These lines are separated from the points by a
defined width. The width is perpendicular to the lines
angle at that particular point.
}
\author{
max
}
\keyword{internal}

