% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bezier_private_functions.R
\name{getLinesWithArrow}
\alias{getLinesWithArrow}
\title{Gets the lines for an arrow}
\usage{
getLinesWithArrow(bp, arrow, end_points, width, default.units, align_2_axis)
}
\arguments{
  \item{bp}{Bezier points that define the line. Based upon
  a bezier grob.}

  \item{arrow}{The arrow list. This should contain a x and
  y element to indicate the vector in relation to the 0,0
  origo. The base element sets the arrow width while the
  length element gives the length of the arrow.}

  \item{end_points}{The end points of a line. This is a
  list with a start list(x, y, length) and a end list(x, y,
  length) as elements.}

  \item{width}{The width as a grid length
  \code{\link{unit}}}

  \item{default.units}{The default unit (see the grid
  package for available units)}

  \item{align_2_axis}{This indicates if the arrows origin
  should align it to an axis. Which is decided by which
  axis is the closes one.}
}
\value{
A list with left and right elements indicating the two
lines
}
\description{
The lines are the upper and the lower lines that are later
used to generate future polygon that will be used to
generate the arrow. These lines are separated from the
points by a defined width. The width is perpendicular to
the lines angle at that particular point.
}
\keyword{internal}

