% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_export.R
\name{ggplot_export}
\alias{ggplot_export}
\title{Save ggplot}
\usage{
ggplot_export(p, name = "Figure.eps", dpi = 300, w = 7, h = 4)
}
\arguments{
\item{p}{A ggplot object}

\item{name}{Name for the exported file}

\item{dpi}{Resolution for the exported file}

\item{w}{With in inches}

\item{h}{Height in inches}
}
\value{
Exported eps file
}
\description{
Saves/exports a ggplot object to eps.
}
\details{
The exported "eps" file can be manipulated further for final tweaking in the desired vector software or converted to "pdf" as is. This format allows for transparency to be preserved
}
\examples{
p = ggplot(mtcars, aes(wt, mpg)) + geom_point() + geom_smooth(method = "lm")
ggplot_export(p)
}
