% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini.R
\name{Gini}
\alias{Gini}
\title{Gini index for a given split of a categorical variable}
\usage{
Gini(x, group, thres)
}
\arguments{
\item{x}{A numeric vector}

\item{group}{Categorical variable}

\item{thres}{Threshold value to use for the split}
}
\value{
A data frame with Gini index before and after the split, as well as the gain and gain percent
}
\description{
Calculates the Gini index of a categorical variable for a given split (threshold) value of a numeric variable.
}
\examples{
Gini(mtcars$mpg, mtcars$cyl, 15)
Gini(mtcars$mpg, mtcars$cyl, 21)
Gini(mtcars$mpg, mtcars$cyl, 25)

}
\references{
Kuhn, M. & Johnson, K. (2013). Applied Predictive Modeling. Springer.

Witten, I. H., Frank, E. & Hall, M. A. (2011). Data Mining: Practical Machine Learning Tools and Techniques. Elsevier.
}
