% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{\%D*\%}
\alias{\%D*\%}
\title{Premultiplies a diagonal matrix by a matrix.}
\usage{
x \%D*\% Y
}
\arguments{
\item{x}{numeric, representing the diagonal of the matrix}

\item{Y}{matrix, having a number of rows equal to the length of x}
}
\value{
a matrix, equal to \code{diag(x) \%*\% Y}
}
\description{
Premultiplies a diagonal matrix by a matrix.
}
\author{
G.A.Paleologo
}

