% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{\%nin\%}
\alias{\%nin\%}
\title{the "not in" operator}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
you guessed it: a vector
}
\description{
the "not in" operator
}

