% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_and_has.R
\name{has}
\alias{has}
\title{An R alternative to the lodash \code{has} in JavaScript}
\usage{
has(sourceList, path)
}
\arguments{
\item{sourceList}{The \code{list()}/\code{c()} that is to be searched for the element}

\item{path}{A string that can be separated by [,] or ., the string \code{"elementname1.1.elementname"}
the validity of the path - it only separates and tries to address that element with `[[]]`.}
}
\value{
Returns a boolean.
}
\description{
This is a handy function for checking if item exist in a nested structure
}
\examples{
has(list(a = list(b = 1)), "a.b")

}
\seealso{
Other lodash similar functions:
\code{\link[=retrieve]{retrieve()}}
}
\concept{lodash similar functions}
