% Generated by roxygen2 (4.0.1): do not edit by hand
\name{heightDetails.xaxis}
\alias{heightDetails.xaxis}
\title{Gets the height for an x-axis object}
\usage{
\method{heightDetails}{xaxis}(x)
}
\arguments{
  \item{x}{The \code{\link[grid]{xaxisGrob}} object}
}
\value{
A unit object
}
\description{
A function that gets the height of an axisGrob. It is for
some reason not included by default in the
\code{grid}-package.
}
\examples{
grid.newpage()
xg <- xaxisGrob(c(1:3))
convertY(grobHeight(xg), "lines")
}
\author{
Max
}
\keyword{internal}

