% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hoek_Brown.R
\name{Hoek_Brown}
\alias{Hoek_Brown}
\title{Hoek-Brown criterion}
\usage{
Hoek_Brown(
  sig.ci,
  GSI,
  mi,
  MR,
  D,
  height,
  unit.weight,
  use = c("general", "tunnels", "slopes")
)
}
\arguments{
\item{sig.ci}{The unconfined compressive strength of the intact rock, in MPa}

\item{GSI}{The Geological Strength Index}

\item{mi}{The intact rock constant}

\item{MR}{the modulus ratio}

\item{D}{The quality of the explosive campaing design}

\item{height}{The height or depth for which to determine the stress level and parameters}

\item{unit.weight}{The unit weight of the rock, in kN/m3}

\item{use}{The use expected for the rock mass (general, tunnels, or slopes)}
}
\value{
A list with a list of the resulting parameters of the rock mass and 3 data frames: Shear envelopes for Mohr-Coulomb and Hoek-Brown criteria, Principal stresses envelopes for Mohr-Coulomb and Hoek-Brown criteria, and stress level for the given height and unit weight
}
\description{
Calculates the Hoek-Brown criterion for the given data and for the given level of stress (unit weight and depth).
}
\details{
Tables for some of the input parameters can be found in the directory where the package is installed, just run the following command to find this path: \code{system.file(package = "GMisc")}
}
\examples{
sig.ci = 16
GSI = 75
mi = 13
MR = 300
D = 0
height = 40
unit.weight = 18.6
Hoek_Brown(sig.ci, GSI, mi, MR, D, height, unit.weight)

}
\references{
Hoek, E. & Brown, E. T. (1980). Empirical strength criterion for rock masses. J. Geotech. Eng. Div. ASCE 106 (GT9): 1013-1035.

Hoek, E., Carranza-Torres, C. & Corkum, B. (2002). Hoek-Brown Failure Criterion - 2002 Edition. Proc. NARMS-TAC Conference, Toronto. 267-273.

Hoek, E. & Marinos, P. (2007). A brief history of the development of the Hoek-Brown criterion. Soils and Rocks, No. 2.
}
