% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hoek_Brown_plot.R
\name{Hoek_Brown_plot}
\alias{Hoek_Brown_plot}
\title{Hoek-Brown criterion plot}
\usage{
Hoek_Brown_plot(HB)
}
\arguments{
\item{HB}{The result from running \code{Hoek_Brown()}}
}
\value{
A list with two ggplot objects for the Hoek-Brown criterion (principal stresses and normal-shear stresses)
}
\description{
Draws the Hoek-Brown criterion for the given data and for the given level of stress (unit weight and depth).
}
\examples{
sig.ci = 16
GSI = 75
mi = 13
MR = 300
D = 0
height = 40
unit.weight = 18.6
HB = Hoek_Brown(sig.ci, GSI, mi, MR, D, height, unit.weight)
Hoek_Brown_plot(HB)

}
\references{
Hoek, E. & Brown, E. T. (1980). Empirical strength criterion for rock masses. J. Geotech. Eng. Div. ASCE 106 (GT9): 1013-1035.

Hoek, E., Carranza-Torres, C. & Corkum, B. (2002). Hoek-Brown Failure Criterion - 2002 Edition. Proc. NARMS-TAC Conference, Toronto. 267-273.

Hoek, E. & Marinos, P. (2007). A brief history of the development of the Hoek-Brown criterion. Soils and Rocks, No. 2.
}
