% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mergeDesc.R
\name{latex.descMrg}
\alias{htmlTable.descMrg}
\alias{latex.descMrg}
\title{A wrapper around the \code{\link[htmlTable]{htmlTable}}/\code{\link[Hmisc]{latex}} function}
\usage{
\method{latex}{descMrg}(x, ...)

\method{htmlTable}{descMrg}(x, ...)
}
\arguments{
  \item{x}{The \code{\link{mergeDesc}} object}

  \item{...}{Passed onto
  \code{\link[htmlTable]{htmlTable}}/\code{\link[Hmisc]{latex}}}
}
\description{
Calls the
\code{\link[htmlTable]{htmlTable}}/\code{\link[Hmisc]{latex}}
after extracting the \code{rgroup}/\code{n.rgroup}
arguments.
}
\keyword{internal}

