% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/htmlTable_tableOne.R
\name{htmlTable.TableOne}
\alias{htmlTable.CatTable}
\alias{htmlTable.ContTable}
\alias{htmlTable.TableOne}
\title{A function for converting a table generated from the \pkg{tablone} package}
\usage{
\method{htmlTable}{TableOne}(x, ..., css.rgroup = "",
  t1_args = list(printToggle = FALSE))

\method{htmlTable}{CatTable}(...)

\method{htmlTable}{ContTable}(...)
}
\arguments{
  \item{x}{The TableOne object}

  \item{...}{Passed onto
  \code{\link[htmlTable]{htmlTable}}}

  \item{css.rgroup}{Overrides the default bold text. See
  \code{\link[htmlTable]{htmlTable}} for details.}

  \item{t1_args}{Arguments passed to the
  \code{\link[tableone]{print.TableOne}}}
}
\value{
Returns an \code{\link[htmlTable]{htmlTable}} object
}
\description{
This function translates the \code{tableone}'s descriptive
table into an \code{\link[htmlTable]{htmlTable}}-object.
The function uses the raw output from the
\code{print.TableOne} and looks for \code{header},
\code{rgroup}, and \code{n.rgroup} elements.
}
\examples{
library(tableone)

## Load Mayo Clinic Primary Biliary Cirrhosis Data
library(survival)
data(pbc)
## Check variables
head(pbc)

## Make categorical variables factors
varsToFactor <- c("status","ascites","hepato","spiders","edema","stage")
pbc[varsToFactor] <- lapply(pbc[varsToFactor], factor)

## Create a variable list
dput(names(pbc))
vars <- c("time","status","age","sex","ascites","hepato",
          "spiders","edema","bili","chol","albumin",
          "copper","alk.phos","ast","trig","platelet",
          "protime","stage")


pbc$trt <- factor(pbc$trt,
                  labels = c("D-penicillmain", "Placebo"))
## Create Table 1 stratified by trt
tableOne <- CreateTableOne(vars = vars, strata = c("trt"), data = pbc)

htmlTable(tableOne,
          caption="A descriptive table with the Mayo Clinic trial
          in primary biliary cirrhosis (PBC) of the liver",
          col.rgroup = c("#FFF", "#F9F9F0"))
}
\keyword{internal}

