% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induced_stress.R
\name{induced_stress}
\alias{induced_stress}
\title{Induced stress on a surface}
\usage{
induced_stress(qs, B, L = NULL, z.end, footing = c("strip", "square",
  "rectangular", "circular"))
}
\arguments{
\item{qs}{The stress applied by the footing onto the surface}

\item{B}{The footing's width (in meters)}

\item{L}{The footing's length for a rectangular footing (in meters)}

\item{z.end}{The depth to which calculate the stresses (in meters)}

\item{footing}{Type of footing for which to calculate the stresses (Default is "strip")}
}
\value{
A stress distribution plot for the desired footing
}
\description{
Plots the stress distribution under the center of a footing of different shapes.
}
\details{
The stresses are calculated using the solutions from Boussinesq. The parameter \code{L} only applies for rectangular footings, otherwise is not considered, thus the default value of \code{NULL}. For rectangular footings the length (\code{L}) can be varied, while for the other footing shapes the width \code{B} can be varied. For circular footings the width is equal to the radius (\code{B=R})
}
\examples{
# Singular value of B for a strip footing
qs = 1
B = 1
L = NULL
z.end = 8
induced_stress(qs, B, L, z.end)
# Different values of B for a square footing
qs = 1
B = c(1, 2, 4)
L = NULL
z.end = 8
induced_stress(qs, B, L, z.end, footing = "square")
# Different values of L for a rectangular footing
qs = 1
B = 1
L = c(1, 2, 4)
z.end = 8
induced_stress(qs, B, L, z.end, footing = "rectangular")

}
\references{
Holtz, R. D., Kovacs, W. D. & Sheahan, T. C. (2011). An Introduction to Geotechnical Engineering. Prentice Hall.
}
