% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs_s3_append.R
\name{insert}
\alias{insert}
\alias{insert.default}
\alias{insert.Gmisc_list_of_boxes}
\title{Insert element into a list of boxes (S3)}
\usage{
insert(x, element, ..., after = NULL, before = NULL)

\method{insert}{default}(x, element, ..., after = NULL, before = NULL)

\method{insert}{Gmisc_list_of_boxes}(x, element, ..., after = NULL, before = NULL)
}
\arguments{
\item{x}{A `Gmisc_list_of_boxes`.}

\item{element}{The new box object to insert.}

\item{...}{Not used.}

\item{after}{The name or index of the box after which to insert.}

\item{before}{The name or index of the box before which to insert.}
}
\value{
The updated list of boxes with the new element inserted.
}
\description{
Inserts a box into a list of boxes and positions it relative to the
surrounding boxes.
}
\seealso{
Other flowchart components:
\code{\link[=align]{align()}},
\code{\link[=append]{append()}},
\code{\link[=boxGrob]{boxGrob()}},
\code{\link[=boxHeaderGrob]{boxHeaderGrob()}},
\code{\link[=boxPropGrob]{boxPropGrob()}},
\code{\link{boxShapes}},
\code{\link[=connectGrob]{connectGrob()}},
\code{\link[=coords]{coords()}},
\code{\link[=distance]{distance()}},
\code{\link[=equalizeWidths]{equalizeWidths()}},
\code{\link[=flowchart]{flowchart()}},
\code{\link[=move]{move()}},
\code{\link[=moveBox]{moveBox()}},
\code{\link[=spread]{spread()}}
}
\concept{flowchart components}
