% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_logit.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse logit}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{A vector of continuous values}
}
\value{
A vector of probabilities
}
\description{
Calculates the probability (0, 1) using the inverse logit function.
}
\examples{
x = seq(-3,3,.5)
inv_logit(x)

}
