% Generated by roxygen2 (4.0.0): do not edit by hand
\name{isHorizontal}
\alias{isHorizontal}
\title{Just a simple help with interpreting the radians}
\usage{
isHorizontal(angle_radian)
}
\arguments{
  \item{angle_radian}{The angle in radians}
}
\value{
boolean
}
\description{
Just a simple help with interpreting the radians
}
\author{
max
}
\keyword{internal}

