% Generated by roxygen2 (4.0.1): do not edit by hand
\name{isKnitting}
\alias{isKnitting}
\title{Detects if knitr is running}
\usage{
isKnitting()
}
\value{
\code{TRUE/FALSE} depending if knitting
}
\description{
The function tries to deduce if the environment contains
functions that indicate that knitr is running and
compiling. Previously all that was necessary was to detect
the package:knitr string within the environment but with
recent changes a more complicated detection algorithm has
been designed.
}
\examples{
isKnitting()
}
\author{
Max Gordon
}

