% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klimmeridge.R
\docType{data}
\name{Klimmeridge}
\alias{Klimmeridge}
\title{Klimmeridge shale organic matter content}
\format{
A data frame with organic matter content in sucesive intervals
}
\description{
Dataset of measurements of the organic matter content (%) in succesive intervals from the Klimmeridge shale.
}
\references{
Swan, A. R. H. & Sandilands, M. (1995). Introduction to Geological Data Analysis. Blackwell Science.
}
\keyword{datasets}
