% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_screeplot.R
\name{kmeans_screeplot}
\alias{kmeans_screeplot}
\title{kmeans_screeplot}
\usage{
kmeans_screeplot(x, n = 15)
}
\arguments{
\item{x}{A data frame containing the data to use for the kmeans algorithm}

\item{n}{The maximum number of clusters to try}
}
\value{
A scree plot
}
\description{
Creates a scree plot for a given number of clusters to determine the "optimal" number of clusters.
}
\examples{
kmeans_screeplot(CPTu_data[,2:ncol(CPTu_data)], n = 15)

}
