% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{l2m}
\alias{l2m}
\title{converts a raggedarray or list of vectors into a matrix}
\usage{
l2m(x, by.row = TRUE)
}
\arguments{
\item{x}{raggedarray or list of vectors}

\item{by.row}{boolean, should the vectors be rows}
}
\value{
a matrix
}
\description{
converts a raggedarray or list of vectors into a matrix
}
\author{
G.A.Paleologo
}

