% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{lapplyseq}
\alias{lapplyseq}
\title{sequential list processing.}
\usage{
lapplyseq(X, FUN, label = c("first", "last"), ...)
}
\arguments{
\item{FUN}{a function}

\item{...}{additional elements to give to FUN}

\item{X1}{list}
}
\value{
a list
}
\description{

}
\details{
Given a list X and a function FUN(x1,x2)
computes Z[[i]] <- FUN(X[[i-1]], X[[i-1]])
}
\examples{
X <- list(a=1,b=2, c=NA, d=4,e=5)
FUN <- function(x,y, ...) mean(c(x,y), ...)
procList(X, FUN, na.rm=F)
}
\author{
G.A.Paleologo
}

