% Generated by roxygen2 (4.0.2): do not edit by hand
\name{latex.descMrg}
\alias{latex.descMrg}
\title{A wrapper around the \code{\link[Hmisc]{latex}} function}
\usage{
\method{latex}{descMrg}(x, ...)
}
\arguments{
  \item{x}{The \code{\link{mergeDesc}} object}

  \item{...}{Passed onto \code{\link[Hmisc]{latex}}}
}
\description{
Calls the \code{\link[Hmisc]{latex}} after extracting the
\code{rgroup}/\code{n.rgroup} arguments.
}
\keyword{internal}

