% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{m2df}
\alias{m2df}
\title{converts a matrix into a data frame}
\usage{
m2df(x, key = NA)
}
\arguments{
\item{x}{matrix}

\item{key}{string, optional name of the first column}
}
\value{
a data frame
}
\description{
converts a matrix into a data frame by adding a first
column equal to the row names
}
\author{
G.A.Paleologo
}

