% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{m2l}
\alias{m2l}
\title{converts a matrix into a ragged array object}
\usage{
m2l(x, by.row = TRUE)
}
\arguments{
\item{x}{matrix}

\item{by.row}{boolean, should the vectors be rows}
}
\value{
a ragged array
}
\description{
converts a matrix into a ragged array object, i.e., a list of vectors;
each vector is a row or column of the original matrix
}
\author{
G.A.Paleologo
}

