% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{m2v}
\alias{m2v}
\title{converts an n x 1 matrix into a vector}
\usage{
m2v(x)
}
\arguments{
\item{x}{matrix}
}
\value{
numerical vector
}
\description{
converts an n x 1 matrix into a vector
}
\author{
G.A.Paleologo
}

