% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{m2zoo}
\alias{m2zoo}
\title{converts a matrix into a zoo object}
\usage{
m2zoo(x, by.row = TRUE)
}
\arguments{
\item{x}{matrix}

\item{by.row}{boolean, are the dates rows}
}
\value{
a zoo object
}
\description{
converts a matrix into a zoo object
}
\author{
G.A.Paleologo
}

