% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{map}
\alias{map}
\title{Iterate a function over a list.}
\usage{
map(x, f, ..., .method = c("default", "names", "values"))
}
\arguments{
\item{x}{list. The list-of-lists to iterate over.}

\item{f}{function. The function to apply.}

\item{...}{list. Additional optional arguments to pass to lapply.}
}
\description{
Iterate a function over a list.
}
