% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{Mapping}
\alias{Mapping}
\title{R6 based key-value mapping}
\description{
This class is basically a python dictionary.
}
\examples{

## ------------------------------------------------
## Method `Mapping$new`
## ------------------------------------------------

ages <- Mapping$new(
  Charlie = 40L,
  Alice = 30L,
  Bob = 25L
)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Mapping-new}{\code{Mapping$new()}}
\item \href{#method-Mapping-set}{\code{Mapping$set()}}
\item \href{#method-Mapping-get}{\code{Mapping$get()}}
\item \href{#method-Mapping-has}{\code{Mapping$has()}}
\item \href{#method-Mapping-remove}{\code{Mapping$remove()}}
\item \href{#method-Mapping-keys}{\code{Mapping$keys()}}
\item \href{#method-Mapping-values}{\code{Mapping$values()}}
\item \href{#method-Mapping-clear}{\code{Mapping$clear()}}
\item \href{#method-Mapping-as_list}{\code{Mapping$as_list()}}
\item \href{#method-Mapping-update}{\code{Mapping$update()}}
\item \href{#method-Mapping-pop}{\code{Mapping$pop()}}
\item \href{#method-Mapping-clone}{\code{Mapping$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mapping-new"></a>}}
\if{latex}{\out{\hypertarget{method-Mapping-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mapping$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Any length of key and value pairs a single list, named vector,
or data.frame. If you would like to use a not valid R name as a key, you
must wrap it by backquotes or convert it using \code{make.names}. If key
value pairs are provided, they are are constructed using rlang::list2,
thus the items are constructed sequentially.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ages <- Mapping$new(
  Charlie = 40L,
  Alice = 30L,
  Bob = 25L
)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mapping-set"></a>}}
\if{latex}{\out{\hypertarget{method-Mapping-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mapping$set(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{a string.}

\item{\code{value}}{any value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mapping-get"></a>}}
\if{latex}{\out{\hypertarget{method-Mapping-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mapping$get(key, default = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{a string.}

\item{\code{default}}{any value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mapping-has"></a>}}
\if{latex}{\out{\hypertarget{method-Mapping-has}{}}}
\subsection{Method \code{has()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mapping$has(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{a string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mapping-remove"></a>}}
\if{latex}{\out{\hypertarget{method-Mapping-remove}{}}}
\subsection{Method \code{remove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mapping$remove(key, silent = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{a string.}

\item{\code{silent}}{logical. Whether or not to show a message after removal.
Return all the keys of a dictionary.
Return all the values in the dictionary.
Remove all key value pairs from the dictionary.
Convert the dictionary to a list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mapping-keys"></a>}}
\if{latex}{\out{\hypertarget{method-Mapping-keys}{}}}
\subsection{Method \code{keys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mapping$keys()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mapping-values"></a>}}
\if{latex}{\out{\hypertarget{method-Mapping-values}{}}}
\subsection{Method \code{values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mapping$values()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mapping-clear"></a>}}
\if{latex}{\out{\hypertarget{method-Mapping-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mapping$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mapping-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-Mapping-as_list}{}}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mapping$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mapping-update"></a>}}
\if{latex}{\out{\hypertarget{method-Mapping-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mapping$update(inherits_mapping)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inherits_mapping}}{a subclass of Mapping}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mapping-pop"></a>}}
\if{latex}{\out{\hypertarget{method-Mapping-pop}{}}}
\subsection{Method \code{pop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mapping$pop(key, default = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{a string.}

\item{\code{default}}{any value.
Internal accessor for underlying data.
Very similar to as_list but the return value is a named list with the key and value.
The number of key-value pairs in the mapping.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mapping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Mapping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
