% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp.R
\name{mcp}
\alias{mcp}
\title{Multiple change-point analysis for layer boundary determination}
\usage{
mcp(data, R = 199, alpha = 2, sig.level = 0.01, min.perc = 15)
}
\arguments{
\item{data}{A data frame containing the depth/distance in the first column, and the variables of interest in the rest of the columns, for a CPTu test: point resistance (qc), sleeve friction (fs), and pore-water pressure (u)}

\item{R}{The number of random permutations}

\item{alpha}{A parameter between 0 (exclusive) and 2 (inclusive). lower values make allow for more variation in the search for changepoints}

\item{sig.level}{Significance level to determine significance of the changepoints}

\item{min.perc}{Minimum percentage of data points per layer (between changepoints)}
}
\value{
ggplot and plotly objects showing the layer distinction, statistical summary of the layers, and a summary table
}
\description{
Performs multiple change-point analysis in univariate or multivariate data, to find layer boundaries in a perforation log.
}
\details{
The example data given is intended to show the structure needed for input data. The user should follow this structure, which in general corresponds with a data frame with a sequence in the first column and the observed/measured values in the rest of the columns
}
\examples{
mcp(CPTu_data, R = 199, alpha = 2, sig.level = .01, min.perc = 15) # multivariate example
mcp(DPM_data, R = 199, alpha = 2, sig.level = .01, min.perc = 15) # univariate example

}
\references{
Nicholas A. James, David S. Matteson (2014). ecp: An R Package for Nonparametric Multiple Change Point Analysis of Multivariate Data, Journal of Statistical Software, 62(7), 1-25.
}
