% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeLists.R
\name{mergeLists}
\alias{mergeLists}
\title{Merging of multiple lists}
\usage{
mergeLists(..., lapplyOutput = NULL)
}
\arguments{
\item{...}{Any number of lists that you want to merge}

\item{lapplyOutput}{The \code{\link[base]{lapply}} function outputs a number
of lists and this is for specifically merging all of those.}
}
\value{
Returns a list with all the given lists.
}
\description{
The merge allows for a recursive component
where the lists are compared on the subelement.
If one does not contain that element it will get NA
in for those parameters.
}
\examples{
v1 <- list("a" = c(1, 2), b = "test 1", sublist = list(one = 20:21, two = 21:22))
v2 <- list("a" = c(3, 4), b = "test 2", sublist = list(one = 10:11, two = 11:12, three = 1:2))
mergeLists(v1, v2)
}
