% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs_units.R
\name{MidDistanceX}
\alias{MidDistanceX}
\alias{MidDistanceY}
\title{Calculate midpoint between two units}
\usage{
MidDistanceX(u1, u2)

MidDistanceY(u1, u2)
}
\arguments{
\item{u1}{The first unit}

\item{u2}{The second unit}
}
\value{
A unit object
}
\description{
Calculates the midpoint between two units. If the arguments are
numeric they are converted to npc units.
}
\examples{
MidDistanceX(unit(.1, "npc"), unit(.8, "npc"))
}
