% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{mintersect}
\alias{mintersect}
\title{applies intersect to multiple arguments}
\usage{
mintersect(...)
}
\arguments{
\item{...}{multiple arguments to which the intersect operation can be applied}
}
\value{
a vector/set equal to the intersection of the input sets
}
\description{
applies intersect to multiple arguments
}
\author{
G.A.Paleologo
}

