% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_val.R
\name{mode_val}
\alias{mode_val}
\title{Modal value}
\usage{
mode_val(x)
}
\arguments{
\item{x}{A vector of numbers}
}
\value{
The modal value of a numeric vector
}
\description{
Calculates the mode (modal value) of a vector of numbers
}
\examples{
x = rnorm(50, 80, 15)
mode_val(x)

}
