% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mohr_circle.R
\name{Mohr_Circle}
\alias{Mohr_Circle}
\title{Mohr circle solution and plot}
\usage{
Mohr_Circle(sigx, sigy, tauxy, theta = 0, digits = 5, units = "kPa")
}
\arguments{
\item{sigx}{A number corresponding with the mayor normal stress}

\item{sigy}{A number corresponding with the minor normal stress}

\item{tauxy}{A number corresponding with the shear stress}

\item{theta}{Angle of rotation for which to find the state of stresses}

\item{digits}{Number of significant digits to us}

\item{units}{A string with the units of the measurements}
}
\value{
A list containing a data frame with the results, and a ggplot object with the Mohr circle
}
\description{
Solves a Mohr circle problem for the given normal stresses, shear stress, and rotation angle, and plots the results.
}
\examples{
sigx = 143.6
sigy = 100.5
tauxy = -14.4
Mohr_Circle(sigx, sigy, tauxy)
Mohr_Circle(sigx, sigy, tauxy, theta = -35)

}
\references{
Coduto, D. P. (1999). Geotechnical Engineering - Principles and Practices. Prentice Hall.

Holtz, R. D., Kovacs, W. D. & Sheahan, T. C. (2011). An Introduction to Geotechnical Engineering. Prentice Hall.

Gonzalez de Vallejo, L. I. (2004). Ingenieria Geologica. Prentice Hall.
}
