% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs_s3_move.R
\name{move}
\alias{move}
\alias{move.default}
\alias{move.Gmisc_list_of_boxes}
\title{Move boxes (S3)}
\usage{
move(element, ...)

\method{move}{default}(element, ...)

\method{move}{Gmisc_list_of_boxes}(element, subelement = NULL, ...)
}
\arguments{
\item{element}{A `list` of boxes.}

\item{...}{Arguments passed to [`moveBox`].}

\item{subelement}{The name of the element to move (maps to `subelement` in `moveBox`)}
}
\value{
The moved list of boxes.
}
\description{
Move a box or list of boxes. A wrapper for [`moveBox`] that supports
piping of `Gmisc_list_of_boxes` while preserving attributes.
}
\seealso{
Other flowchart components:
\code{\link[=align]{align()}},
\code{\link[=append]{append()}},
\code{\link[=boxGrob]{boxGrob()}},
\code{\link[=boxHeaderGrob]{boxHeaderGrob()}},
\code{\link[=boxPropGrob]{boxPropGrob()}},
\code{\link{boxShapes}},
\code{\link[=connectGrob]{connectGrob()}},
\code{\link[=coords]{coords()}},
\code{\link[=distance]{distance()}},
\code{\link[=equalizeHeights]{equalizeHeights()}},
\code{\link[=equalizeWidths]{equalizeWidths()}},
\code{\link[=flowchart]{flowchart()}},
\code{\link[=insert]{insert()}},
\code{\link[=moveBox]{moveBox()}},
\code{\link[=phaseLabel]{phaseLabel()}},
\code{\link[=spread]{spread()}}
}
\concept{flowchart components}
