% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_avg_filter.R
\name{moving_avg_filter}
\alias{moving_avg_filter}
\title{Moving average filter}
\usage{
moving_avg_filter(x, y, k)
}
\arguments{
\item{x}{A numeric vector with the position values of the data to filter (a sequence)}

\item{y}{A numeric vector with the data to filter}

\item{k}{A numeric vector with the widths of the windows to use for the filter (must be an odd number)}
}
\value{
A list with two data frames: one with the filtered results plus the original data, and one with the residuals
}
\description{
Filters data using moving averages of diferent window sizes.
}
\examples{
data(nautilus)
head(nautilus)
k = c(3, 5, 11, 25)
moving_avg_filter(nautilus$x, nautilus$y, k)

}
\references{
Swan, A. R. H. & Sandilands, M. (1995). Introduction to Geological Data Analysis. Blackwell Science.
}
