% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_filter_plot_f.R
\name{moving_filter_plot_f}
\alias{moving_filter_plot_f}
\title{Moving (weighted) average filter plot (filtered)}
\usage{
moving_filter_plot_f(x, xlab = "X", ylab = "Data",
  filterlab = "Filter & data", plotk = "all")
}
\arguments{
\item{x}{The result of using the \code{moving_avg_filter()} or \code{moving_wt_filter()} functions}

\item{xlab}{The label for the x-axis}

\item{ylab}{The label for the y-axis}

\item{filterlab}{The label for the legend}

\item{plotk}{A string for choosing what to plot (Default is "all")}
}
\value{
A ggplot object with the filtered data and original data
}
\description{
Plots the filtered results of a moving (weighted) average filter routine.
}
\examples{
data(nautilus)
k = c(3, 5, 7, 9)
filt = moving_avg_filter(nautilus$x, nautilus$y, k)
moving_filter_plot_f(filt)
moving_filter_plot_f(filt, plotk = c("y", "k_9")) # plots data and filter of window 9
moving_filter_plot_f(filt, plotk = c("k_3", "k_9")) # plots filters of window 3 and 9

}
