% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_filter_plot_r.R
\name{moving_filter_plot_r}
\alias{moving_filter_plot_r}
\title{Moving (weighted) average filter plot (residual)}
\usage{
moving_filter_plot_r(x, xlab = "X", ylab = "Residual data",
  filterlab = "Filter", plotk = "all")
}
\arguments{
\item{x}{The result of using the \code{moving_avg_filter()} or \code{moving_wt_filter()} functions}

\item{xlab}{The label for the x-axis}

\item{ylab}{The label for the y-axis}

\item{filterlab}{The label for the legend}

\item{plotk}{A string for choosing what to plot (Default is "all")}
}
\value{
ggplot and plotly objects with the residual data
}
\description{
Plots the residual results of a moving (weighted) average filter routine.
}
\examples{
data(nautilus)
k = c(3, 5, 7, 9)
filt = moving_avg_filter(nautilus$x, nautilus$y, k)
moving_filter_plot_r(filt)
moving_filter_plot_r(filt, plotk = c("k_3", "k_9")) # plots residuals for filters of window 3 and 9

}
