% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_wt_filter.R
\name{moving_wt_filter}
\alias{moving_wt_filter}
\title{Moving weighted average filter}
\usage{
moving_wt_filter(x, y, k)
}
\arguments{
\item{x}{A numeric vector with the position values of the data to filter (a sequence)}

\item{y}{A numeric vector with the data to filter}

\item{k}{A numeric vector with the widths of the windows to use for the filter (must be an odd number)}
}
\value{
A list with two data frames: one with the filtered results plus the original data, and one with the residuals
}
\description{
Filters data using moving weighted averages of diferent window sizes.
}
\details{
The weigths for the window lengths are calculated in a way that farther observations from the central (estimated) value are given a lesser weight than observations near de central value
}
\examples{
data(nautilus)
head(nautilus)
k = c(5, 7, 9)
moving_wt_filter(nautilus$x, nautilus$y, k)

}
\references{
Swan, A. R. H. & Sandilands, M. (1995). Introduction to Geological Data Analysis. Blackwell Science.
}
