% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{munion}
\alias{munion}
\title{applies union to multiple arguments}
\usage{
munion(...)
}
\arguments{
\item{...}{multiple arguments to which the intersect operation can be applied}
}
\value{
a vector/set equal to the union of the input sets
}
\description{
applies union to multiple arguments
}
\author{
G.A.Paleologo
}

