% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{MVTSPlot}
\alias{MVTSPlot}
\title{multivariate plotting}
\usage{
MVTSPlot(X, obs = c("rows", "cols"), lowcol = "red", highcol = "green")
}
\arguments{
\item{X}{matrix,}
}
\value{
a plot
}
\description{
multivariate plotting
}
\author{
G.A.Paleologo
}

