% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{nalocf}
\alias{nalocf}
\title{Fast last-observation carried forward for matrices}
\usage{
nalocf(x, memory = 72L)
}
\arguments{
\item{x}{matrix}

\item{memory}{integer, how long missing data should be carried forward?}
}
\value{
a matrix
}
\description{
Fast last-observation carried forward for matrices
}

