% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nautilus.R
\docType{data}
\name{nautilus}
\alias{nautilus}
\title{Data for moving average filter}
\format{
A data frame with growth line spacing data from a Silurian nautiloid
}
\description{
Example data to be used with the \code{moving_avg_filter()} function.
}
\references{
Swan, A. R. H. & Sandilands, M. (1995). Introduction to Geological Data Analysis. Blackwell Science.
}
\keyword{datasets}
