% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewStats_2samples.R
\name{NewStats_2samples}
\alias{NewStats_2samples}
\title{New Statistics two samples (idependent & dependent)}
\usage{
NewStats_2samples(x, y, dep = F)
}
\arguments{
\item{x}{Sample 1}

\item{y}{Sample 2}

\item{dep}{Logical indicating if the samples are dependent (Default is F)}
}
\value{
A ggplot object
}
\description{
It plots the two samples and the difference between them with their confidence intervals (CI), based on the two samples being independent or dependent.
}
\details{
The labels of the plot are generic but they can be customised using the \code{labs()} and \code{scale_x_discrete(labels = c())} functions. For the dependent samples case the two means are joined by a line, to indicate the dependency
}
\examples{
set.seed(101)
x = rnorm(20, 15, 6)
y = rnorm(20, 10, 8)
y1 = rnorm(20,x/1.1,2)
NewStats_2samples(x, y, dep = F)
NewStats_2samples(x, y1, dep = T)

}
\references{
Cumming, G. (2014). The Statistics: Why and How. - Psychological Science, 25(1): 7-29.
}
