% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nq.R
\name{Nq}
\alias{Nq}
\title{Bearing capacity factor Nq}
\usage{
Nq(phi, case = "general", method = "Terzaghi")
}
\arguments{
\item{phi}{Friction angle in degrees}

\item{case}{"general" or "local" to indicate general or local shear failure ("general" is default)}

\item{method}{"Terzaghi" or "Vesic" ("Terzaghi" is default)}
}
\value{
Bearing capacity factor
}
\description{
Taken from "geotech" package. Calculate the bearing capacity factor Nq using either the Terzaghi or Vesic methods.
}
\examples{
Nq(phi = 20, case = "local", method = "Terzaghi")

}
\references{
Coduto, D.P., Kitch, W.A., and Yeung, M.R. (2016). Foundation Design: Principles and Practices, Pearson, Boston.

Terzaghi, K. (1943). Theoretical Soil Mechanics, John Wiley, New York

Vesic, A.S. (1973). Analysis of Ultimate Loads of Shallow Foundations, ASCE Journal of the Soil Mechanics and Foundations Division, Vol. 99, No. SM1, pp. 45-73.
}
