% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{opVectors}
\alias{opVectors}
\title{applies a function to two vectors element-wise, aligning them}
\usage{
opVectors(x, y, all = FALSE, all.x = all, all.y = all, na.values = 0,
  FUN = `+`, ...)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}

\item{all}{boolean. Should this be applied to all elements?}

\item{all.x}{boolean. Should all the elements of the first vector be used?}

\item{all.y}{boolean. Should all the elements of the second vector be used?}

\item{na.values}{numeric. Which value should be used in place of NAs?}
}
\value{
a vector
}
\description{
applies a function to two vectors element-wise, aligning them
}
\note{
\code{opVectors} requires that vectors have names for their elements.
}
\examples{
a <- b <- seq(letters)
names(a) <- names(b) <- letters
opVectors(a[6:15], a[1:10], all.x = TRUE)
opVectors(a[6:15], a[1:10], FUN = `/`)
}
\author{
G.A.Paleologo
}

