% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper_data_prep.R
\name{piper_data_prep}
\alias{piper_data_prep}
\title{Piper Data Preparation}
\usage{
piper_data_prep(data)
}
\arguments{
\item{data}{A dataframe with cations and anions, with the following names: Ca, Mg, Na, K, Cl, SO4, CO3, HCO3}
}
\value{
A dataframe with \code{x} and \code{y} coordinates and the rest of the original data
}
\description{
It transforms hydrochemical data (cations and anions) so they can be plotted on \code{piper_diagram()}.
}
\examples{

d = data.frame(Group = c('A','A','B','B'),
               Ca = c(120,150,110,52.6),
               Mg = c(78,160,110,28),
               Na = c(210,590,340,51.6),
               K = c(4.2,2,3.6,2.3),
               HCO3 = c(181,181,189,151),
               CO3 = 0,
               Cl = c(220,744,476,72.2),
               SO4 = c(560,1020,584,126))

piper_data_prep(d)

}
