% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDescriptionStatsBy_prAddEmptyVals.R
\name{prAddEmptyVals}
\alias{prAddEmptyVals}
\title{Convert the by-list into a matrix compatible format}
\usage{
prAddEmptyVals(t, missing_value)
}
\arguments{
\item{t}{Output from [prNumericDescs], [prPropDescs], or [prFactorDescs].}

\item{missing_value}{Value that is substituted for empty cells. Defaults to "-"}
}
\value{
A fixed list
}
\description{
Helper for [getDescriptionStatsBy] that fixes empty values in matrix
so that they are compatible with the matrix
}
