\name{prCaAddFactorRgroups}
\alias{prCaAddFactorRgroups}
\title{A function for adding factor rgroups}
\usage{
prCaAddFactorRgroups(value_mtrx, dataset, use_labels)
}
\arguments{
  \item{value_mtrx}{A matrix containing the values of prCa}

  \item{dataset}{The dataset that is used for the fit}

  \item{use_labels}{Wether or not to use labels}
}
\value{
\code{list} Returns the value matrix with the attributes
rgroup and n.rgroup
}
\description{
A function for adding factor rgroups
}
\author{
max
}

