\name{prCaAddReference}
\alias{prCaAddReference}
\title{Adds a reference to value matrix}
\usage{
  prCaAddReference(vn, matches, available_factors, ret,
    reference_zero_effect, ds, use_labels)
}
\arguments{
  \item{vn}{Variable name}

  \item{matches}{Rows that match}

  \item{available_factors}{Factors that are available}

  \item{ret}{The ret list}

  \item{reference_zero_effect}{The reference zero effect}

  \item{ds}{The data set}

  \item{use_labels}{If labels should be used for row names}
}
\value{
  \code{list} The ret list
}
\description{
  Adds a reference to value matrix
}
\author{
  max
}

