\name{prCaGetOrderVariables}
\alias{prCaGetOrderVariables}
\title{Re-order variables}
\usage{
prCaGetOrderVariables(names, order)
}
\arguments{
  \item{names}{The names of the variables}

  \item{order}{The order regular expression}
}
\value{
\code{vector} A vector containing the greps
}
\description{
Re-order variables
}
\author{
max
}

