\name{prCaGetRowname}
\alias{prCaGetRowname}
\title{Gets the labelled rowname if it exists}
\usage{
  prCaGetRowname(vn, use_labels, ds)
}
\arguments{
  \item{vn}{The variable name}

  \item{use_labels}{If labels should be used}

  \item{ds}{The dataset}
}
\value{
  \code{string} The rowname
}
\description{
  Gets the labelled rowname if it exists
}
\author{
  max
}

