% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prConvertShowMissing}
\alias{prConvertShowMissing}
\title{A functuon for converting a show_missing variable.}
\usage{
prConvertShowMissing(show_missing)
}
\arguments{
  \item{show_missing}{Boolean or "no", "ifany", "always"}
}
\value{
string
}
\description{
The variable is suppose to be directly compatible with
\code{\link[base]{table}}(..., useNA=show_missing). It
throughs an error if not compatible. \emph{Deprecated:}
This function will be deprecated as all functions now use
the useNA style in order to comply with standard R naming.
}
\keyword{internal}

