% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{prDescGetAndValidateDefaultRef}
\alias{prDescGetAndValidateDefaultRef}
\title{A helper function for the description stats}
\usage{
prDescGetAndValidateDefaultRef(x, default_ref)
}
\arguments{
\item{x}{The variable of interest with the levels}

\item{default_ref}{The default reference, either first,
the level name or a number within the levels. If left out
it defaults to the first value.}
}
\value{
\code{integer} The level number of interest
}
\description{
A helper function for the description stats
}
\keyword{internal}
