% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{prev}
\alias{prev}
\title{given a vector v and an elements x, returns
the previous element before x in v}
\usage{
prev(x, v)
}
\arguments{
\item{x}{element}

\item{v}{vector. All elements must be different.}
}
\value{
an element of v
}
\description{
given a vector v and an elements x, returns
the previous element before x in v
}
\author{
G.A.Paleologo
}

