% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDescriptionStatsBy_prFixDescRownames.R
\name{prFixDescRownames}
\alias{prFixDescRownames}
\title{Fix rownames for descriptive results}
\usage{
prFixDescRownames(results, t, name)
}
\arguments{
\item{results}{A matrix with the results}

\item{t}{The [base::by()] output}

\item{name}{Name if row names are missing or the results is a single row}
}
\value{
The results with fixed names
}
\description{
Helper for [getDescriptionStatsBy] that fixes row names
}
