\name{prFpFetchRowLabel}
\alias{prFpFetchRowLabel}
\title{Get the label}
\usage{
  prFpFetchRowLabel(label_type, labeltext, i, j)
}
\arguments{
  \item{label_type}{The type of label}

  \item{labeltext}{The text}

  \item{i}{The row}

  \item{j}{The column}
}
\value{
  An expression or a text
}
\description{
  A function used for fetching the text or expression from
  the supplied labeltext.
}
\author{
  max
}

