% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forestplot2_helpers.R
\name{prFpGetConfintFnList}
\alias{prFpGetConfintFnList}
\title{Get a function list}
\usage{
prFpGetConfintFnList(fn, no_rows, no_cols)
}
\arguments{
  \item{fn}{The function list/matrix. If a list it should
  be in the format [[row]][[col]], the function tries to
  handle this but in cases where the columns and rows are
  the same it will not know what is a column and what is a
  row.}

  \item{no_rows}{Number of rows}

  \item{no_cols}{Number of columns}
}
\value{
\code{list} The function returns a list that has the format
[[row]][[col]] where each element contains the function
that you need to call using the \code{\link[base]{as.call}}
and \code{\link[base]{eval}} functions:
\code{eval(as.call(list(fn[[row]][[col]], arg_1=1,
arg_2=2)))}
}
\description{
This function helps the \code{\link{forestplot2}} to deal
with multiple drawing functions for the confidence
intervals.
}
\keyword{internal}

