% Generated by roxygen2 (4.0.0): do not edit by hand
\name{prFpGetLabels}
\alias{prFpGetLabels}
\title{Gets the forestplot labels}
\usage{
prFpGetLabels(label_type, labeltext, align, nc, nr, is.summary,
  fontfamily.summary, fontfamily.labelrow, col, cex)
}
\arguments{
  \item{label_type}{The type of text labels}

  \item{labeltext}{The text labels}

  \item{align}{Alignment, should be equal to
  \code{length(nc}}

  \item{nc}{Number of columns}

  \item{nr}{Number of rows}

  \item{is.summary}{If the row is a summary}

  \item{fontfamily.summary}{The summary fontfamily}

  \item{fontfamily.labelrow}{The regular fontfamily
  (non-summary)}

  \item{col}{A list with the colors}

  \item{cex}{The font size adjustment}
}
\value{
\code{list} A list with \code{length(nc)} where each
element contains a list of \code{length(nr)} elements with
attributes width/height for each element and
max_width/max_height for the total
}
\description{
A function that gets all the labels
}
\author{
max
}
\keyword{internal}

