% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prFpGetLegendGrobs}
\alias{prFpGetLegendGrobs}
\title{Gets the forestplot legend grobs}
\usage{
prFpGetLegendGrobs(legend, legend.cex, legend.title = NULL)
}
\arguments{
  \item{legend}{Legen corresponding to the number of bars.}

  \item{legend.cex}{The cex size of the legend, defaults to
  80 \% procent of cex parameter.}

  \item{legend.title}{The title of the legend if any,
  default to NULL}
}
\value{
\code{list} A "Legend" class that derives from a list with
all the different legends. The list also contains
attributes such as height, width, max_height, max_width,
line_height_and_spacing. The title of the legend is saved
inside \code{attr("title")}
}
\description{
Gets the forestplot legend grobs
}
\keyword{internal}

