\name{prFpPrintLabels}
\alias{prFpPrintLabels}
\title{Plots the labels}
\usage{
  prFpPrintLabels(labels, nc, nr)
}
\arguments{
  \item{labels}{A list to the labels}

  \item{nc}{Number of columns}

  \item{nr}{Number of rows}
}
\value{
  void
}
\description{
  This is a helper function to the
  \code{\link{forestplot2}} function.
}
\author{
  Max
}

