% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_document.R
\name{prFtpHeaderStyle}
\alias{prFtpHeaderStyle}
\title{Helper to ftp_document}
\usage{
prFtpHeaderStyle(output_str, h1_style, other_h_style)
}
\arguments{
\item{output_str}{The output string from readLines()}

\item{h1_style}{You can choose any css style formatting here that you
want to be applied to all h1 elements. Note: this is only applied
if LibreOffice_adapt is \code{TRUE}.}

\item{other_h_style}{This is the formatting applied to any other
h elements not included to the first. Note: this is only applied
if LibreOffice_adapt is \code{TRUE}.}
}
\value{
string
}
\description{
Helper to ftp_document
}
\keyword{internal}
