% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prFtpOtherRemoval}
\alias{prFtpOtherRemoval}
\title{Removes other unwanted lines}
\usage{
prFtpOtherRemoval(output_str)
}
\arguments{
  \item{output_str}{The input from readLines()}
}
\value{
string Returns without the script elements
}
\description{
Removes other unwanted lines
}

