% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prFtpScriptRemoval}
\alias{prFtpScriptRemoval}
\title{Removes the <script>*</scrip>}
\usage{
prFtpScriptRemoval(output_str)
}
\arguments{
  \item{output_str}{The input from readLines()}
}
\value{
string Returns without the script elements
}
\description{
Removes the <script>*</scrip>
}

