% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prGetAndValidateDefaultRef}
\alias{prGetAndValidateDefaultRef}
\title{A helper function for the description stats}
\usage{
prGetAndValidateDefaultRef(x, default_ref)
}
\arguments{
  \item{x}{The variable of interest with the levels}

  \item{default_ref}{The default reference, either first,
  the level name or a number within the levels}
}
\value{
integer The level number of interest
}
\description{
A helper function for the description stats
}
\author{
max
}
\keyword{internal}

