% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs_spread_npc_norm_helpers.R
\name{prGetBoxAxisDefaults}
\alias{prGetBoxAxisDefaults}
\title{Get default from/to values for an axis}
\usage{
prGetBoxAxisDefaults(axis = c("x", "y"))
}
\arguments{
\item{axis}{Axis, either "x" or "y"}
}
\value{
A list with `from` and `to` values as `unit` objects
}
\description{
When axis is "x", the default is from 0 to 1 npc, i.e. left to right.
When axis is "y", the default is from 1 to 0 npc, i.e. top to bottom.
}
