% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDescriptionStatsBy_prGetDescHeader.R
\name{prGetDescHeader}
\alias{prGetDescHeader}
\title{Retrieve basic description stats by header}
\usage{
prGetDescHeader(by, html, header_count, already_table_format = FALSE)
}
\arguments{
\item{by}{The variable that you want to split into different
columns}

\item{html}{If HTML compatible output should be used. If \code{FALSE}
it outputs LaTeX formatting}

\item{header_count}{Set to \code{TRUE} if you want to add a header count,
e.g. Smoking; No. 25 observations, where there is a new line after the
factor name. If you want a different text for the second line you can
specifically use the \code{\link[base]{sprintf}} formatting, e.g. "No. \%s patients".}

\item{already_table_format}{Just a boolean as we use this in the total column}
}
\value{
A vector with basic headers
}
\description{
Helper for [getDescriptionStatsBy] that retrieves the basic
header names.
}
