% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prGetStatistics}
\alias{prGetStatistics}
\title{Get statistics according to the type}
\usage{
prGetStatistics(x, show_perc = FALSE, html = TRUE, digits = 1,
  numbers_first = TRUE, show_missing = TRUE, show_missing.digits = digits,
  show_all_values = FALSE, continuous_fn = describeMean,
  factor_fn = describeFactors, prop_fn = factor_fn,
  percentage_sign = percentage_sign)
}
\arguments{
  \item{x}{The variable that you want the statistics for}

  \item{html}{If HTML compatible output shoudl be used
  instead of default LaTeX}

  \item{digits}{The number of decimals used}

  \item{numbers_first}{If the number should be given or if
  the percentage should be presented first. The second is
  encapsulated in parentheses ().}

  \item{show_missing}{Show the missing values. This adds
  another row if there are any missing values.}

  \item{show_missing.digits}{The number of digits to use
  for the missing percentage, defaults to the overall
  \code{digits}.}

  \item{show_all_values}{This is by default false as for
  instance if there is no missing and there is only one
  variable then it is most sane to only show one option as
  the other one will just be a complement to the first. For
  instance sex - if you know gender then automatically you
  know the distribution of the other sex as it's 100 \% -
  other \%. To choose which one you want to show then set
  the \code{default_ref} parameter.}

  \item{continuous_fn}{The method to describe continuous
  variables. The default is \code{\link{describeMean}}.}

  \item{factor_fn}{The method used to describe factors, see
  \code{\link{describeFactors}}.}

  \item{prop_fn}{The method used to describe proportions,
  see \code{\link{describeProp}}.}

  \item{percentage_sign}{If you want to surpress the
  percentage sign you can set this variable to FALSE. You
  can also choose something else that the default \% if you
  so wish by setting this variable.}
}
\value{
A matrix or a vector depending on the settings
}
\description{
A simple function applied by the
\code{\link{getDescriptionStatsBy}} for the total column.
}
\keyword{internal}

