% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{prGetTextGrobCex}
\alias{prGetTextGrobCex}
\title{Just a simple acces to the gp$cex parameter}
\usage{
prGetTextGrobCex(x)
}
\arguments{
\item{x}{The text-grob of interest}
}
\value{
\code{numeric} The cex value, 1 if no cex was present
}
\description{
Just a simple acces to the gp$cex parameter
}
\keyword{internal}
