% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prHtAddAlign2Style}
\alias{prHtAddAlign2Style}
\title{Adds text align to style element}
\usage{
prHtAddAlign2Style(style, align)
}
\arguments{
  \item{style}{The CSS style for an element}

  \item{align}{The alignment for that element}
}
\value{
\code{string} Returns the style with the alignment sett
according to the align argument
}
\description{
Adds text align to style element
}
\seealso{
Other hidden helper functions for \code{\link{htmlTable}}:
\code{\link{prHtAddCells}};
\code{\link{prHtAddSemicolon2StrEnd}};
\code{\link{prHtGetCgroupHeader}};
\code{\link{prHtGetRowlabelPos}};
\code{\link{prHtGetTableCntr}};
\code{\link{prHtPrepareCgroup}}
}
\keyword{internal}

