% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prHtAddSemicolon2StrEnd}
\alias{prHtAddSemicolon2StrEnd}
\title{Add a ; at the end}
\usage{
prHtAddSemicolon2StrEnd(my_str)
}
\arguments{
  \item{my_str}{The string that is to be processed}
}
\value{
\code{string}
}
\description{
The CSS expects a semicolon at the end of each argument
this function just adds a semicolong if none is given and
remove multiple semicolon if such exist
}
\seealso{
Other hidden helper functions for \code{\link{htmlTable}}:
\code{\link{prHtAddCells}};
\code{\link{prHtGetCgroupHeader}};
\code{\link{prHtGetRowlabelPos}};
\code{\link{prHtGetStyle}}; \code{\link{prHtPrepareAlign}};
\code{\link{prHtPrepareCgroup}}; \code{\link{prHtTblNo}}
}
\keyword{internal}

