% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prHtGetAlign}
\alias{prHtGetAlign}
\title{Gets alignment}
\usage{
prHtGetAlign(align_req)
}
\arguments{
  \item{align_req}{The align option from
  \code{\link{htmlTable}}}
}
\description{
AVI added utility function to incorporate custom (LaTeX)
column alignment
}
\keyword{internal}

