% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prHtGetCgroupHeader}
\alias{prHtGetCgroupHeader}
\title{Retrieve a header row}
\usage{
prHtGetCgroupHeader(x, cgroup_vec, n.cgroup_vec, cgroup_vec.just, row_no,
  top_row_style, rowname, rowlabel, rowlabel.pos, cgroup_spacer_cells)
}
\arguments{
  \item{cgroup_vec}{The cgroup may be a matrix, this is
  just one row of that matrix}

  \item{n.cgroup_vec}{The same as above but for the
  counter}

  \item{cgroup_vec.just}{The same as above bot for the
  justificaiton}

  \item{row_no}{The row number within the header group.
  Useful for multirow headers when we need to output the
  rowlabel at the \code{rowlabel.pos} level.}

  \item{top_row_style}{The top row has a special style
  depending on the \code{ctable} option in the
  \code{htmlTable} call.}

  \item{cgroup_spacer_cells}{The spacer cells due to the
  multiple cgroup levels. With multiple rows in cgroup we
  need to keep track of how many spacer cells occur between
  the columns. This variable contains is of the size
  \code{ncol(x)-1} and 0 if there is no cgroup element
  between.}

  \item{x}{The matrix/data.frame with the data}

  \item{rowname}{Default is rownames of matrix or
  data.frame.}

  \item{rowlabel}{If x has row dimnames, rowlabel is a
  character string containing the column heading for the
  row dimnames.}

  \item{rowlabel.pos}{Where the rowlabel should be
  positioned. This value can be "top", "bottom", "header",
  or a integer between \code{1} and \code{nrow(cgroup) +
  1}. The options "bottom", "header" are the same, where
  the row label is presented at the same level as the
  header.}
}
\value{
\code{string}
}
\description{
This function retrieves a header row, i.e. a row within the
<th> elements on top of the table. Used by
\code{\link{htmlTable}}.
}
\seealso{
Other hidden helper functions for \code{\link{htmlTable}}:
\code{\link{prHtAddAlign2Style}};
\code{\link{prHtAddCells}};
\code{\link{prHtAddSemicolon2StrEnd}};
\code{\link{prHtGetRowlabelPos}};
\code{\link{prHtGetTableCntr}};
\code{\link{prHtPrepareCgroup}}
}
\keyword{internal}

