% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prHtGetRowlabelPos}
\alias{prHtGetRowlabelPos}
\title{Gets the rowlabel position}
\usage{
prHtGetRowlabelPos(cgroup, rowlabel.pos, headings)
}
\arguments{
  \item{cgroup}{a vector or a matrix of character strings
  defining major column headings. The default is to have
  none. This is also known as "the column spanner". If you
  want a column not to have a spanner then put that column
  as "". If you pass cgroup and n.crgroup as matrices you
  can have multiline cgroups. If the different levels have
  different number of elements you need to set the ones
  that lack elements to NA. For instance \code{cgroup =
  rbind(c("first", "second", NA), c("a", "b", "c"))}.}

  \item{rowlabel.pos}{Where the rowlabel should be
  positioned. This value can be "top", "bottom", "header",
  or a integer between \code{1} and \code{nrow(cgroup) +
  1}. The options "bottom", "header" are the same, where
  the row label is presented at the same level as the
  header.}

  \item{headings}{a vector of character strings specifying
  column headings, defaulting to \code{x}'s
  \code{colnames}}
}
\value{
\code{integer} Returns the position within the header rows
to print the \code{rowlabel} argument
}
\description{
Gets the rowlabel position
}
\seealso{
Other hidden helper functions for \code{\link{htmlTable}}:
\code{\link{prHtAddCells}};
\code{\link{prHtAddSemicolon2StrEnd}};
\code{\link{prHtGetCgroupHeader}};
\code{\link{prHtGetStyle}}; \code{\link{prHtPrepareAlign}};
\code{\link{prHtPrepareCgroup}}; \code{\link{prHtTblNo}}
}
\keyword{internal}

