% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prHtGetStyle}
\alias{prHtGetStyle}
\title{Gets the CSS style element}
\usage{
prHtGetStyle(styles, ...)
}
\arguments{
  \item{styles}{The styles can be provided as
  \code{vector}, \code{named vector}, or \code{string}.}

  \item{...}{All styles here are merged with the first
  parameter. If you provide a name, e.g.
  \code{styles="background: blue", align="center"} the
  function will convert the \code{align} into proper
  \code{align: center}.}
}
\value{
\code{string} Returns the codes merged into one string with
correct CSS ; and : structure.
}
\description{
A funciton for checking, merging, and more with a variety
of different style formats.
}
\seealso{
Other hidden helper functions for \code{\link{htmlTable}}:
\code{\link{prHtAddCells}};
\code{\link{prHtAddSemicolon2StrEnd}};
\code{\link{prHtGetCgroupHeader}};
\code{\link{prHtGetRowlabelPos}};
\code{\link{prHtGetTableCntr}};
\code{\link{prHtPrepareCgroup}}
}
\keyword{internal}

