% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prHtGetTableCntr}
\alias{prHtGetTableCntr}
\title{Gets the table counter string}
\usage{
prHtGetTableCntr()
}
\value{
\code{string} Returns a string formatted according to the
table_counter_str and table_counter_roman. The number is
decided by the table_counter variable
}
\description{
Returns the string used for htmlTable to number the
different tables. Uses options \code{table_counter},
\code{table_counter_str}, and \code{table_counter_roman} to
produce the final string. You can set each option by simply
calling \code{options()}.
}
\seealso{
Other hidden helper functions for \code{\link{htmlTable}}:
\code{\link{prHtAddCells}};
\code{\link{prHtAddSemicolon2StrEnd}};
\code{\link{prHtGetCgroupHeader}};
\code{\link{prHtGetRowlabelPos}};
\code{\link{prHtGetStyle}}; \code{\link{prHtPrepareAlign}};
\code{\link{prHtPrepareCgroup}}
}
\keyword{internal}

