% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prHtPrepareAlign}
\alias{prHtPrepareAlign}
\title{Prepares the align to match the columns}
\usage{
prHtPrepareAlign(align, x, rnames, default_rn = "l")
}
\arguments{
  \item{align}{a character strings specifying column
  alignments, defaulting to \code{paste(c("l",
  rep('c',ncol(n_table)-1)),collapse='')} to center. Valid
  alignments are l = left, c = center and r = right. You
  can also specify \code{align='c|c'} and other LaTeX
  tabular formatting.}

  \item{x}{The matrix/data.frame with the data}

  \item{rnames}{Default rownames are generated from
  \code{rownames(x)}. If you provide \code{FALSE} then it
  will skip the rownames. Note that even if you do
  \code{rownames(my_dataframe) <- NULL} it still has
  rownames, thus you need to use \code{FALSE} if you want
  to surpress those.}

  \item{default_rn}{The default rowname alignment. This is
  an option as the header uses the same function and there
  may be differences in how the alignments should be
  implemented.}
}
\description{
The alignment may be tricky and this function therefore
simplifies this process by extending/shortening the
alignment to match the correct number of columns.
}
\seealso{
Other hidden helper functions for \code{\link{htmlTable}}:
\code{\link{prHtAddCells}};
\code{\link{prHtAddSemicolon2StrEnd}};
\code{\link{prHtGetCgroupHeader}};
\code{\link{prHtGetRowlabelPos}};
\code{\link{prHtGetStyle}};
\code{\link{prHtPrepareCgroup}}; \code{\link{prHtTblNo}}
}
\keyword{internal}

