% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prHtSkipRownames}
\alias{prHtSkipRownames}
\title{Returns if rownames should be printed for the htmlTable}
\usage{
prHtSkipRownames(rnames)
}
\arguments{
  \item{rnames}{Default rownames are generated from
  \code{rownames(x)}. If you provide \code{FALSE} then it
  will skip the rownames. Note that even if you do
  \code{rownames(my_dataframe) <- NULL} it still has
  rownames, thus you need to use \code{FALSE} if you want
  to surpress those.}
}
\description{
Returns if rownames should be printed for the htmlTable
}
\keyword{internal}

