% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs_labelConnector.R
\name{print.Gmisc_connector_label}
\alias{print.Gmisc_connector_label}
\title{Print a connector label object}
\usage{
\method{print}{Gmisc_connector_label}(x, ...)
}
\arguments{
\item{x}{A `Gmisc_connector_label` object}

\item{...}{ignored}
}
\description{
Print a connector label object
}
