\name{print_fit}
\alias{print_fit}
\title{Prints a simple fit summary}
\usage{
  print_fit(fit, digits = 2)
}
\arguments{
  \item{fit}{Regression model object}

  \item{digits}{Number of digits you want to round to}
}
\value{
  The function only prints
}
\description{
  Sometimes you just want a simple fit print without all
  the details and with only the confidence intervals.
}
\author{
  max
}

