% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{prPasteVec}
\alias{prPasteVec}
\title{Collapses a vector for throwing errors}
\usage{
prPasteVec(x)
}
\arguments{
\item{x}{The vector}
}
\description{
The function collapses a vector into an output useful when throwing
errors, e.g. 1:3 becomes '1', '2', '3'
}
