% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_document.R
\name{prSetMaxWidth}
\alias{prSetMaxWidth}
\title{Updates the css max-width}
\usage{
prSetMaxWidth(max_width, css_file)
}
\arguments{
\item{max_width}{The new max-width css setting}

\item{css_file}{The CSS file name}
}
\value{
\code{string} The file name
}
\description{
Reads the file, changes the max-width
for the body element.
}
\keyword{internal}
