% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transitionClass_helpers.R
\name{prTcGetBoxPositions}
\alias{prTcGetBoxPositions}
\title{Gets the box positions}
\usage{
prTcGetBoxPositions(x_offset, width, proportions)
}
\arguments{
  \item{x_offset}{The x start position of the box}

  \item{width}{The box width}

  \item{proportions}{The vertical proportions that are
  \code{(1-sum(proportions)) == total_vertical_space}, i.e.
  the sum is always <= 1.}
}
\value{
\code{list} Returns a list with all the elements positions.
}
\description{
The box positions as a list with scalars for the positions:
\enumerate{ \item \emph{x} The center x-position \item
\emph{y} The center y-position \item \emph{right} The right
edge \item \emph{left} The left edge \item \emph{top} The
top edge \item \emph{bottom} The bottom edge \item
\emph{height} The box height \item \emph{width} The box
width }
}

