% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionClass_helpers.R
\name{prTcMatchClr}
\alias{prTcMatchClr}
\title{Finds the matching colors for the new addition based on the original set
of colors}
\usage{
prTcMatchClr(add, org, no_cols, no_rows, is3D)
}
\arguments{
\item{add}{The colors that we want to add}

\item{org}{The original colors}

\item{no_cols}{The no. of columns in the transition matrix}

\item{no_rows}{All the row counts in the transition plot}

\item{is3D}{Whether the transition matrix has 3 dimensions}
}
\value{
A list with colors that is appropriate for the transition object
}
\description{
Finds the matching colors for the new addition based on the original set
of colors
}
\keyword{internal}
