% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionClass_helpers.R
\name{prTcPlotArrows}
\alias{prTcPlotArrows}
\title{Plots the arrows within a single transition set}
\usage{
prTcPlotArrows(
  trnstn_set,
  widths,
  type = c("simple", "gradient"),
  clr,
  rez,
  origin_boxes,
  target_boxes,
  left_box_clrs,
  add_width,
  max_flow,
  min_width,
  max_width,
  abs_arrow_width = FALSE,
  clr_bar_subspace
)
}
\arguments{
\item{trnstn_set}{The set of transitions to plot}

\item{widths}{The arrow widths}

\item{type}{The type of arrow to be used}

\item{clr}{The arrow colors (a matrix)}

\item{rez}{The arrow resolution}

\item{origin_boxes}{The origin boxes' positions to the left}

\item{target_boxes}{The target boxes' positions to the right}

\item{left_box_clrs}{The colors of the left side boxes}

\item{add_width}{Adds a certain width - useful for background arrows}

\item{max_flow}{The maximum flow if it is common for the entire image}

\item{min_width}{The minimum line width}

\item{max_width}{The maximum line width}

\item{abs_arrow_width}{If the arrow width should be the same for all arrows}
}
\description{
Plots the arrows within a single transition set
}
\keyword{internal}
