% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transitionClass_helpers.R
\name{prTcPlotBoxColumn}
\alias{prTcPlotBoxColumn}
\title{Plots a column of boxes}
\usage{
prTcPlotBoxColumn(x_offset, width, proportions, fill, txt, txt_clr, cex)
}
\arguments{
  \item{x_offset}{The x start position of the box}

  \item{width}{The box width}

  \item{proportions}{The vertical proportions that are
  \code{(1-sum(proportions)) == total_vertical_space}, i.e.
  the sum is always <= 1.}

  \item{fill}{The fill colors of length equal to the
  proportions}

  \item{txt}{The texts of length equal to the proportions}

  \item{txt_clr}{The text colors of length equal to the
  proportions}

  \item{cex}{The fontsize multiplier}
}
\value{
\code{void}
}
\description{
Takes a set of box settings and plots them.
}

