% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transitionClass_helpers.R
\name{prTcValidateAndPrepClr}
\alias{prTcValidateAndPrepClr}
\title{Checks and prepares colors}
\usage{
prTcValidateAndPrepClr(value, transitions, tcObject)
}
\arguments{
  \item{value}{The color values}

  \item{transitions}{The transition matrix}

  \item{tcObject}{The Transtion-object}
}
\value{
\code{matrix} Returns a matrix of the same size as the
transition matrix
}
\description{
The colors need to match the size of the transition matrix
and therefore we need to have this input check. The output
is always a matrix of the exact same dimensions as the
transition matrix.
}
\keyword{internal}

