% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prTpGetBoxPositions}
\alias{prTpGetBoxPositions}
\title{Gets the box position}
\usage{
prTpGetBoxPositions(no, side, transitions, prop_start_sizes, prop_end_sizes,
  tot_spacing, box_width)
}
\arguments{
  \item{no}{The box number}

  \item{side}{The right or left side}

  \item{no_boxes}{The number of boxes}

  \item{transitions}{The flows to or from}

  \item{prop_start_sizes}{The size of the start boxes}

  \item{prop_end_sizes}{The size of the end boxes}

  \item{tot_spacing}{The total space between the boxes}

  \item{box_width}{The width of the box}
}
\value{
\code{list(top, left, bottom, right, width, height)}
}
\description{
Gets the box position
}
\keyword{internal}

