% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionPlot_helpers.R
\name{prTpGetBoxSizedTextGrob}
\alias{prTpGetBoxSizedTextGrob}
\title{Gets the text size for the box}
\usage{
prTpGetBoxSizedTextGrob(
  txt,
  txt_clr,
  txt_cex,
  txt_gpar = NULL,
  force_cex = FALSE,
  ...
)
}
\arguments{
\item{txt}{The text}

\item{txt_clr}{The color of the text}

\item{txt_cex}{The font size}

\item{txt_gpar}{The [grid::gpar()] object that is overridden with col & cex by other parameters}

\item{force_cex}{If font size should be forced}

\item{...}{Other options}
}
\description{
Gets the text size for the box
}
\keyword{internal}
