% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionPlot_helpers.R
\name{prTpPlotArrows}
\alias{prTpPlotArrows}
\title{Plots the arrows}
\usage{
prTpPlotArrows(
  type,
  box_row,
  transition_flow,
  max_flow,
  min_lwd,
  max_lwd,
  clr,
  box_clr,
  transition_arrow_props,
  prop_start_sizes,
  prop_end_sizes,
  tot_spacing,
  box_width,
  abs_arrow_width,
  color_bar_subspace,
  add_width = NA
)
}
\arguments{
\item{type}{The type of arrow used}

\item{box_row}{Which box do the arrows originate from}

\item{max_flow}{The largest transition flow}

\item{min_lwd}{The minimum line width}

\item{max_lwd}{The maximum line width}

\item{clr}{The color of the line}

\item{box_clr}{The color of the box}

\item{transition_arrow_props}{The proportions of the different transitions if
available.}

\item{prop_start_sizes}{The proportions to the left}

\item{prop_end_sizes}{The proportions to the right}

\item{tot_spacing}{Total spacing between boxes}

\item{box_width}{The box width}

\item{color_bar_subspace}{If there is little or no difference exists
at the low/high proportions of the spectrum then it
can be of interest to focus the color change to the center
leaving the tails constant}

\item{add_width}{Add a certain width}
}
\value{
\code{void}
}
\description{
Outputs all the arrows from a box row
}
\keyword{internal}
