% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prtrellis.R
\name{prtrellis}
\alias{prtrellis}
\title{Display Many Lattice Plots On One Page}
\usage{
prtrellis(x, layout = guess(x))
}
\arguments{
\item{x}{A list of trellis plots.}

\item{layout}{The number of columns and rows. By default a semi-intelligent guess is made based on the number of plots. FEATURE_TO_ADD: Should take the size of the plotting window into account. Also should use a similar approach for xqplot.}
}
\value{
NULL. The function is used to create a plot.
}
\description{
Display Many Lattice Plots On One Page
}
\examples{
library(lattice)
prtrellis( list(
lattice::barchart(HairEyeColor,auto.key=T),
lattice::xyplot( height ~ age, Loblolly, groups = Seed, auto.key =TRUE)))
}

