% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_chisq.R
\name{pval_chisq}
\alias{pval_chisq}
\title{P-value for proposed standard deviation of a population}
\usage{
pval_chisq(s, n, sigma, side = c("two", "one"))
}
\arguments{
\item{s}{Sample standard deviation}

\item{n}{Sample size}

\item{side}{To either get a two-tail or one-tail p-value (Default is "two")}

\item{sig}{Population standard deviation}
}
\value{
P-value
}
\description{
Calculates the p-value for a proposed standard deviation of a population (ChiSquare distribution).
}
\examples{
s <- 0.535
n <- 10
sig <- 1
pval_chisq(s, n, sigma)

}
