% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_corr.R
\name{pval_corr}
\alias{pval_corr}
\title{P-value for coefficient of correlation between two variables}
\usage{
pval_corr(r, n, side = c("two", "one"))
}
\arguments{
\item{r}{The coefficient of correlation to test}

\item{n}{Sample size from the sample}

\item{side}{To either get a two-tail or one-tail p-value (Default is "two")}
}
\value{
P-value
}
\description{
Calculates the p-value for a given coefficient of correlation between two variables, using the t-statistic (student-t distribution).
}
\examples{
r <- 0.597
n <- 10
pval_corr(r, n)

}
