% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_F.R
\name{pval_F}
\alias{pval_F}
\title{P-value for ratio of standard deviations of 2 samples}
\usage{
pval_F(s1, n1, s2, n2, side = c("two", "one"))
}
\arguments{
\item{s1}{Sample's 1 standard deviation}

\item{n1}{Sample size of sample 1}

\item{s2}{Sample's 2 standard deviation}

\item{n2}{Sample size of sample 2}

\item{side}{To either get a two-tail or one-tail p-value (Default is "two")}
}
\value{
P-value
}
\description{
Calculates the p-value for the ratio of the standard deviation of 2 samples (F distribution).
}
\examples{
s1 <- 3.1
n1 <- 15
s2 <- 0.8
n2 <- 12
pval_F(s1, n1, s2, n2)

}
