% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_t.R
\name{pval_t}
\alias{pval_t}
\title{P-value for proposed mean of a population}
\usage{
pval_t(x, s, n, mu = 0, side = c("two", "one"))
}
\arguments{
\item{x}{A point estimate of the population mean}

\item{s}{A point estimate of the sample standard deviation}

\item{n}{Sample size from the population}

\item{mu}{Proposed population mean to test for}

\item{side}{To either get a two-tail or one-tail p-value (Default is "two")}
}
\value{
P-value
}
\description{
Calculates the p-value for a proposed mean for a population with unknown standard deviation, using the t-statistic (student-t distribution).
}
\examples{
x <- 6.14
s <- 0.803
n <- 9
mu <- 7
pval_t(x, s, n, mu)

}
