% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_t2.R
\name{pval_t2}
\alias{pval_t2}
\title{P-value for proposed difference of the mean between 2 samples}
\usage{
pval_t2(x1, s1, n1, x2, s2, n2, dmu = 0, side = c("two", "one"))
}
\arguments{
\item{x1}{Sample's 1 mean}

\item{s1}{Sample's 1 standard deviation}

\item{n1}{Sample size of sample 1}

\item{x2}{Sample's 2 mean}

\item{s2}{Sample's 2 standard deviation}

\item{n2}{Sample size of sample 2}

\item{dmu}{Proposed difference between sample means (Default is 0)}

\item{side}{To either get a two-tail or one-tail p-value (Default is "two")}
}
\value{
P-value
}
\description{
Calculates the p-value for the proposed difference of the mean between 2 samples (t distribution).
}
\examples{
x1 <- 34.67
s1 <- 4.97
n1 <- 15
x2 <- 40.87
s2 <- 5.36
n2 <- 15
pval_t2(mu1, s1, n1, mu2, s2, n2)

}
