% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_t2_pair.R
\name{pval_t2_pair}
\alias{pval_t2_pair}
\title{P-value for proposed difference of the mean between 2 paired samples}
\usage{
pval_t2_pair(x, s, n, dmu = 0, side = c("two", "one"))
}
\arguments{
\item{x}{A point estimate of the mean of the difference between observed values}

\item{s}{A point estimate of the standar deviaton of the difference between observed values}

\item{n}{Sample size from the population}

\item{dmu}{Proposed difference between sample means (Default is 0)}

\item{side}{To either get a two-tail or one-tail p-value (Default is "two")}
}
\value{
P-value
}
\description{
Calculates the p-value for the proposed difference of the mean between 2 paired samples (student-t distribution).
}
\examples{
x <- 0.1
s <- 0.125
n <- 10
pval_t2_pair(x, s, n)

}
