% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_93.R
\name{Q_93}
\alias{Q_93}
\title{Q system (1993)}
\usage{
Q_93(Q, De, dot.col = "blue")
}
\arguments{
\item{Q}{A numeric vector with the Q values}

\item{De}{A numeric vector with the eqivalent dimensions}

\item{dot.col}{Color for the points}
}
\value{
A ggplot object
}
\description{
Draws data points on top of the Q calssification system of 1993.
}
\examples{
De = 9
Q = c(.026, 1.4, .078, .56, .35)
Q_93(Q, De)
}
\references{
Grimstad, E. & Barton, N. (1993). Updating the Q-System for NMT. Proc. int. symp. on sprayed concrete - modern use of wet mixed sprayed concrete for underground supports, 46-66. Oslo, Norway.
}
