% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_crit.R
\name{R_crit}
\alias{R_crit}
\title{Mean resultant length critical value}
\usage{
R_crit(n, case = "2D", conf.level = 0.95)
}
\arguments{
\item{n}{Sample size of the correlated values}

\item{case}{String idicating for which case to calculate the critical value, 2D or 3D}

\item{conf.level}{Confidence level to use for the critical value (Default is 0.95, and see details section below)}
}
\value{
Mean resultant length's critical value for a given sample size
}
\description{
Calculates the critical value for the mean resultant length for directional statistics.
}
\details{
The result of this function is mostly to be used with \code{dir_unif_test_xD()}, but can be used as standalone. The options for confidence level are 0.99, 0.975, 0.95, or 0.90
}
\examples{
R_crit(10)
R_crit(10, conf.level = 0.9)
R_crit(10, case = "3D", conf.level = 0.9)

}
\references{
Davis, J. C. (2002). Statistical and Data Analysis in Geology. 3rd ed. John Wiley & Sons.
}
