% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rads.R
\name{rads}
\alias{rads}
\title{Radians}
\usage{
rads(x)
}
\arguments{
\item{x}{Angle in degrees}
}
\value{
Angle in radians
}
\description{
Converts from degrees to radians.
}
\examples{
rads(45)

}
