% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{reapply}
\alias{reapply}
\title{Repeated apply the same function to \code{x}, using the output as the input for
each successive call.}
\usage{
reapply(x, f, times)
}
\arguments{
\item{x}{any R object.}

\item{f}{function. The function should only accept one argument.}

\item{times}{a number.}
}
\description{
Repeated apply the same function to \code{x}, using the output as the input for
each successive call.
}
