% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{str_length}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{stringr}{\code{\link[stringr]{str_length}}}
}}

