% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{repl}
\alias{repl}
\title{utility function to create lists with repeated elements
creates a list with length(x) elements, all identically equal to v
the names of the list elements are the names of x}
\usage{
repl(x, v)
}
\arguments{
\item{x}{character vector}

\item{v}{object}
}
\value{
a list
}
\description{
utility function to create lists with repeated elements
creates a list with length(x) elements, all identically equal to v
the names of the list elements are the names of x
}

