% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{Reverse}
\alias{Reverse}
\title{Reverse the order of a vector}
\usage{
Reverse(X)
}
\arguments{
\item{X}{vector}
}
\value{
vector
}
\description{
Reverse the order of a vector
}
\examples{
v <- letters[1:10]
Reverse(v)
}

