% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_T_plot.R
\name{RI_T_plot}
\alias{RI_T_plot}
\title{Plots the results of an RI or T_stat object}
\usage{
RI_T_plot(x, ylab = "Depth [m]")
}
\arguments{
\item{x}{A list containing the results of the \code{RI()} or \code{T_stat()} functions}

\item{ylab}{The label to use for the y-axis}
}
\value{
A ggplot object based on the results from an RI object
}
\description{
Plots the results of an intraclass correlation coefficient (RI) object or T_stat oject, to easily visualize the perforation log and the suggested layer boundaries.
}
\examples{
testRI = RI(DPM_data, k = 7)
testT = T_stat(DPM_data, k = 7)
RI_T_plot(testRI)
RI_T_plot(testT)

}
\references{
Mora, R. (2013). Uso de metodos estadisticos para la determinacion de capas homogeneas de suelos volcanicos en un sitio de las laderas del Volcan Irazu, Cartago, Costa Rica. - Rev. Geol. Amer. Central, 49: 101-108.

Mora, R. (2013). Uso de metodos estadisticos para la identifacion de capas de suelos volcanicos con el ensayo del cono de pentracion en los terrenos de la Universidad de Costa Rica, Montes de Oca, San Jose, Costa Rica. - Rev. Geol. Amer. Central, 49: 109-120.
}
