% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rmd_ftp}
\alias{rmd_ftp}
\title{Formatter wrapper for \code{\link[rmarkdown]{html_document}} facilitate easier LibreOffice porting}
\usage{
rmd_ftp(..., LibreOffice_adapt = TRUE,
  h1_style = "margin: 24pt 0pt 0pt 0pt;",
  other_h_style = "margin: 10pt 0pt 0pt 0pt;")
}
\arguments{
  \item{...}{Passed onto
  \code{\link[rmarkdow]{html_document}}.}

  \item{LibreOffice_adapt}{\code{TRUE} if the function
  should adapt the html for optimal compatibility with
  LibreOffice.}

  \item{h1_style}{You can choose any css style formatting
  here that you want to be applied to all h1 elements.
  Note: this is only applied if LibreOffice_adapt is
  \code{TRUE}.}

  \item{other_h_style}{This is the formatting applied to
  any other h elements not included to the first. Note:
  this is only applied if LibreOffice_adapt is
  \code{TRUE}.}
}
\value{
String
}
\description{
This function adds the option of having adaptations needed
for seemless integration with word processor for importing
html-documents. The advantage of html documents is the
ability to create advanced formatting frequently needed in
publications and that is available in the
\code{\link{htmlTable}} function. You can view the series
on \url{http://gforge.se} for more details regarding how to
achieve fast-track-publishing (ftp) together with knitr.
}
\examples{
# Possible yaml configuration at the top of the Rmd doc
\dontrun{
---
title: "Test"
author: "MG"
date: "Sunday, July 13, 2014"
output:
  Gmisc::rmd_ftp:
    css: custom.css
---
}
}
\author{
Max Gordon
}

