% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rotateWidthAccAngle}
\alias{rotateWidthAccAngle}
\title{Rotates a width to match a vector}
\usage{
rotateWidthAccAngle(angle, x_origo = NA, y_origo = NA, width = 0,
  default.units, perpendicular = TRUE, prev_angle = NA)
}
\arguments{
  \item{angle}{Angle to rotate according to}

  \item{x_origo}{The x origin if other than 0}

  \item{y_origo}{The y origin if other than 0}

  \item{width}{The width of the desired left & right
  vectors}

  \item{default.units}{As defined by the \pkg{grid}
  package}

  \item{perpendicular}{If the width is perpendicular to the
  provided vector the set this to true}

  \item{prev_angle}{If you want an average from this and
  the previous angle then set this parameter to the
  previous angle}
}
\value{
list A list with left & right vectors of c(x, y) format and
a angle element
}
\description{
This function is provided with an angle and a width. It
then creates a vector that uses the angle and the origo
points to generate a width vector used in
\code{\link{getLines}}, \code{\link{getLinesWithArrow}}
}
\keyword{internal}

