% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2d.R
\name{s2d}
\alias{s2d}
\title{Conversion from strike to dip directions}
\usage{
s2d(x)
}
\arguments{
\item{x}{A vector of directions in degrees}
}
\value{
A vector or directions
}
\description{
Calculates dip directions from strike directions .
}
\examples{
x <- runif(min = 120, max = 170, n = 20)
dip.dir <- s2d(x)
strike <- d2s(strike)

}
