% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_bic.R
\name{sc_bic}
\alias{sc_bic}
\title{Structural change analysis of different layer models}
\usage{
sc_bic(x, h = 0.1)
}
\arguments{
\item{x}{A data frame containing the location variable (depth or distance) in the first column, and the value of interest in the second column}

\item{h}{Percent of data points per layer}
}
\value{
A ggplot and plotly objects showing the BIC statistic
}
\description{
Performs a structural change analysis on diferent layer models to find points where a change in the sequence might be occuring.
}
\details{
The example data given is intended to show the structure needed for input data. The user should follow this structure, which in general corresponds with a data frame with a sequence in the first column and the observed/measured values in the second column
}
\examples{
sc_bic(DPM_data, h = 0.1)

}
