% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_layers.R
\name{sc_layers}
\alias{sc_layers}
\title{Plots a layered model and summarizes the statistics for each layer from a Strucchange result}
\usage{
sc_layers(x, h = 0.1, breaks)
}
\arguments{
\item{x}{A data frame containing the location variable (depth or distance) in the first column, and the value of interest in the second column}

\item{h}{Percent of data points per layer}

\item{breaks}{An integer giving the number of breakpoints to use (from 'Strucchange')}
}
\value{
A ggplot and plotly objects showing the layered model, another showing the confidence intervals, and a summary table
}
\description{
Given a number of breakpoints, plots a layered model of one variable against distance, plots the confidence intervals for each layer, and gives a summary table.
}
\examples{
sc_layers(DPM_data, h = 0.1, breaks = 2)

}
