\name{scrucca_factor2ind}
\alias{scrucca_factor2ind}
\title{Scruccas code that is a helper for factoring for the cmprsk package}
\usage{
  scrucca_factor2ind(x, baseline)
}
\arguments{
  \item{x}{The variable, a vector (column)}

  \item{baseline}{The baseline to reference}
}
\value{
  Returns a matrix with all the dummy variables
}
\description{
  The cmprsk package isn't fully evolved and this function
  helps it by creating dummy variables.
}
\author{
  max
}

