% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_difference}
\alias{set_difference}
\title{Get the values of x that only exist in x and not y.}
\usage{
set_difference(x, ..., .names_only = FALSE)
}
\description{
Get the values of x that only exist in x and not y.
}
