% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_intersection}
\alias{set_intersection}
\title{Get the values that intersection between x and all containers in ...}
\usage{
set_intersection(x, ..., .names_only = FALSE)
}
\description{
Get the values that intersection between x and all containers in ...
}
