% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_coef.R
\name{similarity_coef}
\alias{similarity_coef}
\title{Similarity Coefficient}
\arguments{
\item{x}{A data frame of observations in rows and features (variables) in columns. Features must be numeric and positive.}

\item{samp.names}{A character vector with the names of the observations (samples)}

\item{digits}{Number of digits to use}
}
\value{
A matrix of similarity coefficients
}
\description{
Calculates the similarity coeficient between observations based on their features (columns). Coefficient ranges between 0 and 1, with values close to 1 indicating more similarity.
}
\examples{
dat = iris \%>\% group_by(Species) \%>\% sample_n(2)
similarity_coef(dat[-5])

}
\references{
Borchardt, G.A. (1974). The SIMAN Coefficient for Similarity Analysis. Classification Society Bulletin, 3(2), 2–8.
}
