% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spearman_crit.R
\name{Spearman_crit}
\alias{Spearman_crit}
\title{Spearman critical value}
\usage{
Spearman_crit(n, conf.level = 0.95)
}
\arguments{
\item{n}{Sample size of the correlated values}

\item{conf.level}{Confidence level to use for the critical value (Default is 0.95, and see details section below)}
}
\value{
Spearman's critical value for a given sample size
}
\description{
Calculates the critical value for Spearman's correlation coeficient.
}
\details{
The result of this function can be used to estimate the statistical signficance of Spearman's correlation coeficient. The options for confidence level are 0.99, 0.95, or 0.90
}
\examples{
Spearman_crit(10)
Spearman_crit(10, conf.level = 0.9)

}
\references{
Swan, A. R. H. & Sandilands, M. (1995). Introduction to Geological Data Analysis. Blackwell Science.
}
