% Generated by roxygen2 (4.0.0): do not edit by hand
\name{splitLines4Table}
\alias{splitLines4Table}
\title{If you want a row to span two or more lines}
\usage{
splitLines4Table(..., html = FALSE)
}
\arguments{
  \item{...}{The lines that you want to be joined}

  \item{html}{If it's suppose to be in html or LaTeX.
  Default is LaTeX.}
}
\value{
string
}
\description{
This function helps you to do a multiline table header in
both html and in LaTeX. In html this isn't that tricky, you
just use the <br /> command but in LaTeX I often find
myself writing vbox/hbox stuff and therefore I've created
this simple helper function
}
\examples{
splitLines4Table("hello", "world")
splitLines4Table("hello", "world", html=TRUE)
splitLines4Table("hello", "world", list("A list", "is OK"))
}
\author{
max
}

