% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereo_mean.R
\name{stereo_mean}
\alias{stereo_mean}
\title{Stereonet (mean)}
\usage{
stereo_mean(dir, dip, conf.level = 0.95, points.shape = 3,
  points.col = "blue", plane.col = "red", add = FALSE)
}
\arguments{
\item{dir}{A vector of dir directions in degrees}

\item{dip}{A vector of dip angles in degrees}

\item{conf.level}{Confidence level for the cone of confidence (Default is 0.95)}

\item{points.shape}{Shape value (pch) for the data points}

\item{points.col}{Color for the data points}

\item{plane.col}{Color for the mean plane and cone of confidence}

\item{add}{A logical indicating to add or not another set of measurements and its calculations}
}
\value{
A stereonet plot of data points and its mean estimations
}
\description{
Draws the mean great circle, mean direction and cone of confidence for a set of measurements.
}
\examples{
dir <- runif(min = 30, max = 80, n = 20)
dip <- runif(min = 10, max = 60, n = 20)
stereo_mean(dir, dip)

}
