% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereo_plane.R
\name{stereo_plane}
\alias{stereo_plane}
\title{Stereonet (planes)}
\usage{
stereo_plane(az, dip, dir = c("strike", "dip"), col = "blue",
  add = FALSE)
}
\arguments{
\item{az}{A vector of angular measurements in degrees (strike or dip directions)}

\item{dip}{A vector of dip angles in degrees}

\item{dir}{A string indicating which direction is being given: "strike" (default) or "dip"}

\item{col}{Color for the great circles}

\item{add}{A logical indicating to add or not the great circles to an existing plot}
}
\value{
A stereonet plot
}
\description{
Draws the great circles of planes on an equal area stereonet.
}
\details{
If "strike" is given it assumes it is using the 'right-hand-rule'. This can be added to a stereonet from \code{stereo_pole()}
}
\examples{
strike <- runif(min = 30, max = 80, n = 20)
dip.dir <- s2d(strike)
dip <- runif(min = 10, max = 60, n = 20)
strike2 <- runif(min = 240, max = 290, n = 20)
dip2 <- runif(min = 10, max = 60, n = 20)
stereo_plane(strike, dip)
stereo_plane(dip.dir, dip, dir = "dip")
stereo_plane(strike2, dip2, col = "red", add = TRUE)

}
