% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereo_pole.R
\name{stereo_pole}
\alias{stereo_pole}
\title{Stereonet (poles)}
\usage{
stereo_pole(dir, dip, points.col = "blue", points.shape = 3, add = FALSE)
}
\arguments{
\item{dir}{A vector of dip directions in degrees}

\item{dip}{A vector of dip angles in degrees}

\item{points.col}{Color for the poles}

\item{points.shape}{Shape value (pch) for the poles}

\item{add}{A logical indicating to add or not the poles to an existing plot}
}
\value{
A stereonet plot
}
\description{
Draws the poles to planes (great circles) on an equal area stereonet.
}
\details{
This can be added to a stereonet from \code{stereo_plane()}
}
\examples{
dip.dir <- runif(min = 120, max = 170, n = 20)
dip <- runif(min = 10, max = 60, n = 20)
stereo_pole(dip.dir, dip)
stereo_plane(dip.dir, dip, dir = "dip")
stereo_pole(dip.dir, dip, add = TRUE)

}
