% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereo_ps.R
\name{stereo_ps}
\alias{stereo_ps}
\title{Stereonet (principal stresses)}
\usage{
stereo_ps(dir, dip)
}
\arguments{
\item{dir}{A vector of dir directions in degrees}

\item{dip}{A vector of dip angles in degrees}
}
\value{
A stereonet plot of principal stresses
}
\description{
Adds the principal stresses to a stereonet.
}
\examples{
dir <- runif(min = 30, max = 80, n = 20)
dip <- runif(min = 10, max = 60, n = 20)
stereo_mean(dir, dip)
stereo_ps(dir, dip)

}
