% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_contains}
\alias{str_contains}
\title{Check whether a string contains the substring.}
\usage{
str_contains(string, substring, negate = FALSE, ignore_case = FALSE)
}
\description{
Check whether a string contains the substring.
}
