% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_count}
\alias{str_count}
\title{Count the number of times \code{substring} occurs within \code{string}.}
\usage{
str_count(string, substring)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}
}
\description{
Count the number of times \code{substring} occurs within \code{string}.
}
