% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_count_regex}
\alias{str_count_regex}
\title{Count the number of times a regular expression occurs in \code{string}.}
\usage{
str_count_regex(string, pattern, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[stringr:modifiers]{stringr::regex}}
  \describe{
    \item{\code{pattern}}{Pattern to modify behaviour.}
    \item{\code{ignore_case}}{Should case differences be ignored in the match?}
    \item{\code{multiline}}{If \code{TRUE}, \code{$} and \code{^} match
the beginning and end of each line. If \code{FALSE}, the
default, only match the start and end of the input.}
    \item{\code{comments}}{If \code{TRUE}, white space and comments beginning with
\verb{#} are ignored. Escape literal spaces with \verb{\\\\ }.}
    \item{\code{dotall}}{If \code{TRUE}, \code{.} will also match line terminators.}
  }}
}
\description{
Count the number of times a regular expression occurs in \code{string}.
}
