% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_endswith}
\alias{str_endswith}
\title{Check whether a string ends with \code{suffix}.}
\usage{
str_endswith(string, suffix, negate = FALSE, ignore_case = FALSE)
}
\description{
Check whether a string ends with \code{suffix}.
}
