% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_isalnum}
\alias{str_isalnum}
\title{Check whether \code{string} only contains letters and numbers.}
\usage{
str_isalnum(string)
}
\description{
Check whether \code{string} only contains letters and numbers.
}
