% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_isgraph}
\alias{str_isgraph}
\title{Check whether \code{string} only contains punctuation, letters, and numbers.}
\usage{
str_isgraph(string)
}
\description{
Check whether \code{string} only contains punctuation, letters, and numbers.
}
