% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_isspace}
\alias{str_isspace}
\title{Check whether \code{string} only contains whitespace characters.}
\usage{
str_isspace(string)
}
\description{
Check whether \code{string} only contains whitespace characters.
}
