% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_join}
\alias{str_join}
\title{Join multiple strings into a single string}
\usage{
str_join(sep, ...)
}
\arguments{
\item{sep}{string inserted between each character}
}
\description{
If a single object is passed in after the separator,
it's assumed the user wants to flatten.
If several objects are passed, it's assumed that the
user wants to concatenate.
}
