% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_lstrip}
\alias{str_lstrip}
\title{Remove \code{chars} from the left end of a string. If \code{chars} is NULL,
this function removes whitespace.}
\usage{
str_lstrip(string, chars = NULL)
}
\description{
Remove \code{chars} from the left end of a string. If \code{chars} is NULL,
this function removes whitespace.
}
