% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_normalize_whitespace}
\alias{str_normalize_whitespace}
\title{Simplify redundant whitespace characters.}
\usage{
str_normalize_whitespace(string)
}
\arguments{
\item{string}{A character vector.}
}
\description{
Simplify redundant whitespace characters.
}
