% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_replace}
\alias{str_replace}
\title{Replace all occurences of \code{substring} with \code{replacement} in \code{string}.}
\usage{
str_replace(string, substring, replacement)
}
\description{
Replace all occurences of \code{substring} with \code{replacement} in \code{string}.
}
