% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_replace_map}
\alias{str_replace_map}
\title{Same as str_replace, but ... is named arguments where each name is
a substring and its value is the replacement.}
\usage{
str_replace_map(.string, ...)
}
\description{
Same as str_replace, but ... is named arguments where each name is
a substring and its value is the replacement.
}
