% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_rstrip}
\alias{str_rstrip}
\title{Remove \code{chars} from the right end of a string. If \code{chars} is NULL,
this function removes whitespace.}
\usage{
str_rstrip(string, chars = NULL)
}
\description{
Remove \code{chars} from the right end of a string. If \code{chars} is NULL,
this function removes whitespace.
}
