% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_sentence}
\alias{str_sentence}
\title{Convert a string to sentence case.}
\usage{
str_sentence(string, locale = "en")
}
\arguments{
\item{string}{String to modify}

\item{locale}{Locale to use for translations. Defaults to "en" (English)
to ensure consistent default ordering across platforms.}
}
\description{
Convert a string to sentence case.
}
