% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_split}
\alias{str_split}
\title{Split \code{string} on \code{char}. Passing a value for \code{n}
determines the number of splits to apply.}
\usage{
str_split(string, char = " ", n = -1L)
}
\description{
Split \code{string} on \code{char}. Passing a value for \code{n}
determines the number of splits to apply.
}
