% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_strip}
\alias{str_strip}
\title{Remove \code{chars} from both ends of a string. If \code{chars} is NULL,
this function removes whitespace.}
\usage{
str_strip(string, chars = NULL)
}
\description{
Remove \code{chars} from both ends of a string. If \code{chars} is NULL,
this function removes whitespace.
}
