% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{subset.raggedarray}
\alias{subset.raggedarray}
\title{subsets a ragged array}
\usage{
\method{subset}{raggedarray}(x, v)
}
\arguments{
\item{x}{a ragged array}

\item{v}{a ragged array containing elements named
with names in x, and values of each vector in v
in the names of a corresponding element of x}
}
\value{
a ragged array
}
\description{
subsets a ragged array
}
\author{
G.A.Paleologo
}

