% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{summary.raggedarray}
\alias{summary.raggedarray}
\title{summarizes a ragged array}
\usage{
\method{summary}{raggedarray}(x)
}
\arguments{
\item{x}{ragged array}
}
\value{
a matrix with the summary data
}
\description{
summarizes a ragged array
}
\author{
G.A.Paleologo
}

