% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{symmsetdiff}
\alias{symmsetdiff}
\title{symmetric set difference}
\usage{
symmsetdiff(x, y)

symmsetdiff(x, y)
}
\arguments{
\item{x}{set}

\item{y}{set}

\item{...}{multiple arguments to which the intersect operation can be applied}
}
\value{
a vector/set equal to the symmetric set difference of the input sets

a vector/set equal to the symmetricsetdifference of the input sets
}
\description{
symmetric set difference

computes symmetric set difference
}
\author{
G.A.Paleologo

G.A.Paleologo
}

