% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_stat.R
\name{T_stat}
\alias{T_stat}
\title{T-statistic for layer boundary determination}
\usage{
T_stat(x, k = 7)
}
\arguments{
\item{x}{A data frame containing the depth of perforation in the first column, and the SPT values in the seconds column}

\item{k}{The window length for the number of data points to include in the calculation of T-statistic. Always and odd number}
}
\value{
A list of data frames: original data plus the T-statistic, locations where the coefficient is significant to 0.95 and 0.99, and one with the degrees of freedom and the corresponding critical T-values for the 0.95 and 0.99 signficance
}
\description{
Calculates the T-statistic for a perforation log from a Dynamic Probing Machine (DPM), and finds the locations where this statistic is significant to 0.95 and 0.99. The coefficient is used to find layer boundaries in a perforation log.
}
\details{
The example data given is intended to show the structure needed for input data. The user should follow this structure, which in general corresponds with a data frame with a sequence in the first column and the observed/measured values in the second column
}
\examples{
testT = T_stat(DPM_data, k = 7)

}
\references{
Mora, R. (2013). Uso de metodos estadisticos para la identifacion de capas de suelos volcanicos con el ensayo del cono de pentracion en los terrenos de la Universidad de Costa Rica, Montes de Oca, San Jose, Costa Rica. - Rev. Geol. Amer. Central, 49: 109-120.
}
