% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAS.R
\name{TAS}
\alias{TAS}
\title{TAS Diagram}
\usage{
TAS(output = c("ggplot", "plotly"), language = c("en", "es"))
}
\arguments{
\item{output}{The output format: "ggplot" or "plotly" (default is "ggplot")}

\item{language}{The language to be displayed: "en" for english or "es" for spanish (deafult is "en")}
}
\value{
TAS diagram in the desired format (object)
}
\description{
\code{TAS()} draws either a static or interactive diagram, in english or spanish. It is a base diagram where data can be ploted.
}
\details{
The examples show basic usage and how to add data, which can be more customizable. In general, just map the silica content to the x-axis and alkali content to the y-axis.
}
\examples{
library(ggplot2)
library(plotly)

mydata = data.frame(silica=c(58,50,70), alkali=c(8,4,5))

TAS()
TAS('plotly')

# adding data to ggplot object
TAS(language = 'es') + geom_point(aes(x=silica, y=alkali), data=mydata)

# adding data to plotly object
TAS('plotly') \%>\%
add_markers(x=~silica, y=~alkali,
            data = mydata,
            name = "My data",
            marker = list(size=8,color='orange',symbol=3,opacity=.9)) \%>\%
layout(showlegend = TRUE)

}
