% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_qap_um.R
\name{ternary_qap_um}
\alias{ternary_qap_um}
\title{QAP Ternary Diagram for Ultramafic Rocks}
\usage{
ternary_qap_um(
  output = c("ggplot", "plotly"),
  language = c("en", "es"),
  opacity = 0.5
)
}
\arguments{
\item{output}{The output format: "ggplot" or "plotly" (default is "ggplot")}

\item{language}{The language to be displayed: "en" for english or "es" for spanish (deafult is "en")}

\item{opacity}{Transparency level (default is 0.5)}
}
\value{
QAP ternary diagram for ultramafic rocks in the desired format (object)
}
\description{
\code{ternary_qap_um()} draws either a static or interactive ternary diagram, in english or spanish. It is a base diagram where data can be ploted.
}
\details{
For ploting data on the ggplot object it would be easier if the names of the dataframe are "opx", "ol", and "cpx", that way it gets mapped automatically, if not make sure to use "aes(x=opx,y=ol,z=cpx)".
For ploting on the plotly object the mapping of the new data should be as shown in the example: \code{a = ~ol, b = ~opx, c = ~cpx}, where \code{a} refers to the top ("ol"), \code{b} refers to the bottom left ("opx"), and \code{c} refers to the bottom right ("cpx").
The examples show basic usage and how to add data, which can be more customizable.
}
\examples{
library(ggplot2)
library(plotly)

d = data.frame(ol=c(23,26.9,25.3),
               opx=c(27,23.7,5.1),
               cpx=c(50,49.4,59.6))

# adding data to ggplot object
ternary_qap_um() + geom_point(data = d)

# adding data to plotly object
ternary_qap_um('plotly') \%>\%
add_trace(a = ~ol, b = ~opx, c = ~cpx,
          data = d,
          name = 'My data',
          type = "scatterternary",
          mode = "markers",
          marker = list(size=8,color='coral',symbol=3,opacity=.9),
          hovertemplate = paste0('Ol: \%{a}<br>',
                                 'Opx: \%{b}<br>',
                                 'Cpx: \%{c}'))

}
