% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_qmflt.R
\name{ternary_qmflt}
\alias{ternary_qmflt}
\title{QmFLt Ternary Diagram for Provenance Analysis}
\usage{
ternary_qmflt(
  output = c("ggplot", "plotly"),
  language = c("en", "es"),
  opacity = 0.5
)
}
\arguments{
\item{output}{The output format: "ggplot" or "plotly" (default is "ggplot")}

\item{language}{The language to be displayed: "en" for english or "es" for spanish (deafult is "en")}

\item{opacity}{Transparency level (default is 0.5)}
}
\value{
QmFLt ternary diagram for provenance analysis in the desired format (object)
}
\description{
\code{ternary_qmflt()} draws either a static or interactive ternary diagram, in english or spanish. It is a base diagram where data can be plotted.
}
\details{
For plotting data on the ggplot object it would be easier if the names of the dataframe are "f", "qm", and "lt", that way it gets mapped automatically, if not make sure to use "aes(x=f,y=qm,z=lt)".
For plotting on the plotly object the mapping of the new data should be as shown in the example: \code{a = ~qm, b = ~f, c = ~lt}, where \code{a} refers to the top ("qm"), \code{b} refers to the bottom left ("f"), and \code{c} refers to the bottom right ("lt").
The examples show basic usage and how to add data, which can be more customizable.
}
\examples{
library(ggplot2)
library(plotly)

d = data.frame(qm=c(23,26.9,25.3),
               f=c(27,23.7,5.1),
               lt=c(50,49.4,59.6))

# adding data to ggplot object
ternary_qmflt() + geom_point(data = d)

# adding data to plotly object
ternary_qmflt('plotly') \%>\%
  add_trace(a = ~qm, b = ~f, c = ~lt,
            data = d,
            name = 'My data',
            type = "scatterternary",
            mode = "markers",
            marker = list(size=8,color='coral',
                          symbol=3,opacity=.9),
            hovertemplate = paste0('Qm: \%{a}<br>',
                                   'F: \%{b}<br>',
                                   'Lt: \%{c}'))

}
