% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_span_txt.R
\name{time_to_span_txt}
\alias{time_to_span_txt}
\title{A dense time-span text}
\usage{
time_to_span_txt(
  times,
  month_format = getOption("Gmisc_time_to_span_txt_month", default = "\%b"),
  full_year_format = getOption("Gmisc_time_to_span_txt_month", default = "\%e \%b \%y"),
  start_stop_glue_txt = getOption("Gmisc_time_to_span_txt_template", default =
    "{start} to {stop}")
)
}
\arguments{
\item{times}{The dates or posix timestamps to used for timespan}

\item{month_format}{The \code{\link[base]{strftime}} format string for months}

\item{full_year_format}{The \code{\link[base]{strftime}} format string for the full year}

\item{start_stop_glue_txt}{The string used in the \code{\link[glue]{glue}} for putting the 
start and stop dates together into one string}
}
\value{
\code{string} A string describing the time span
}
\description{
When adding a time span text we often don't want to write *3 jun - 10 jun* but
shorten it to *3 - 10 jun* while retaining month and year info only if the span
crosses between months or years.
}
\examples{
time_to_span_txt(as.POSIXct(c("2020-01-02", "2020-03-01", NA)))
# 2 jan to 1 mar

}
