% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transitionClass.R
\docType{class}
\name{transitionClass-class}
\alias{transitionClass}
\alias{transitionClass-class}
\title{An object for generating transition plots}
\description{
An object for generating transition plots
}
\section{Fields}{

\describe{
\item{\code{transitions}}{This is a 3 dimensional array storing all the transitions.}
}}
\section{Methods}{

\describe{
\item{\code{addTransitions(mtrx, txt)}}{Add a transition matrix. The input has to be a numerical matrix between 2 and 3 dimensions.}

\item{\code{boxSizes(col)}}{Gets the size of the boxes. Generally an integer but you can also set \code{col = 'last'}.}

\item{\code{getDim()}}{Gets the current dimensions of the transitions}

\item{\code{initialize(transitions, fill_clr, txt_clr, ...)}}{Set up a transitionClass object. The \code{transitions} should be a 2D or 3D matrix
as defined in the \code{$addTransitions} section and not as later internally stored.}

\item{\code{noCols()}}{Gets the number of columns}

\item{\code{noRows()}}{Gets the number of boxes in each row}
}}

