% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCS.R
\name{UCS}
\alias{UCS}
\title{Uniaxial Compressive Strength (UCS)}
\usage{
UCS(W, D, P, digits = 2)
}
\arguments{
\item{W}{A vector of sample widths, in mm}

\item{D}{A vector of sample diameters, in mm}

\item{P}{A vector of loads at failure for each sample, in kN}

\item{digits}{Number of significant digits to display (Default is 2)}
}
\value{
A comparison table and graph
}
\description{
Calculates different possible values of UCS based on point load test measurements. It performs several regressions (Gaussian and robust), and also calculates corrected value based on the point load index.
}
\examples{
W = c(70.0,95.0,80.0,90.0,98.0,110.0,100.0,100.0,110.0)
D = c(40,55,60,50,25,60,40,70,50)
P = c(4.3,3.4,4.5,3.1,5.0,6.4,2.6,11,4.1)
UCS(W = W, D = D, P = P)

}
\references{
ASTM (1995). Standard Test Method for Determination of the Point Load Strength Index of Rock.
}
