% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{v2df}
\alias{v2df}
\title{converts a vector into an n x 1 data frame}
\usage{
v2df(x, keys = c("V1", "V2"))
}
\arguments{
\item{x}{vector}

\item{key}{character, optional name of first column}
}
\value{
data.frame
}
\description{
converts a vector into an n x 1 data frame
}
\author{
G.A.Paleologo
}

