% Generated by roxygen2 (4.0.0): do not edit by hand
\name{validateAndConvertVectorInputs}
\alias{validateAndConvertVectorInputs}
\title{Checks the input of a vector}
\usage{
validateAndConvertVectorInputs(x, y, x_origo = NA, y_origo = NA)
}
\arguments{
  \item{x}{The x point of the vector}

  \item{y}{The y point of the vector}

  \item{x_origo}{The x origin if other than 0}

  \item{y_origo}{The y origin if other than 0}
}
\value{
list with the input variables checked and converted if
necessary.
}
\description{
It checks that a vector makes sense in its grid parameters.
It also creates an origo at point 0 if none is provided.
}
\author{
max
}
\keyword{internal}

