% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{withNames}
\alias{withNames}
\title{creates names for a vector equal to its elements}
\usage{
withNames(x)
}
\arguments{
\item{x}{vector}
}
\value{
a vector with names equal to its elements
}
\description{
creates names for a vector equal to its elements
}
\author{
G.A.Paleologo
}

