% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs_s3_append.R
\name{append}
\alias{append}
\alias{append.default}
\alias{append.Gmisc_list_of_boxes}
\title{Append to list of boxes (S3)}
\usage{
append(x, values, after = length(x))

\method{append}{default}(x, values, after = length(x))

\method{append}{Gmisc_list_of_boxes}(x, values, after = length(x))
}
\arguments{
\item{x}{A `Gmisc_list_of_boxes` or `list`.}

\item{values}{The object(s) to append. If a single `boxGrob` is provided, it is wrapped in a list.}

\item{after}{Subscript, after which the values are to be appended.}
}
\value{
Updated list of boxes.
}
\description{
Append an element to a list of boxes, preserving attributes/class.
Masks `base::append` to support S3 dispatch.
}
\seealso{
Other flowchart components:
\code{\link[=align]{align()}},
\code{\link[=boxGrob]{boxGrob()}},
\code{\link[=boxHeaderGrob]{boxHeaderGrob()}},
\code{\link[=boxPropGrob]{boxPropGrob()}},
\code{\link{boxShapes}},
\code{\link[=connectGrob]{connectGrob()}},
\code{\link[=coords]{coords()}},
\code{\link[=distance]{distance()}},
\code{\link[=equalizeWidths]{equalizeWidths()}},
\code{\link[=flowchart]{flowchart()}},
\code{\link[=insert]{insert()}},
\code{\link[=move]{move()}},
\code{\link[=moveBox]{moveBox()}},
\code{\link[=spread]{spread()}}
}
\concept{flowchart components}
