% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_unit}
\alias{assert_unit}
\alias{assert_label}
\alias{assert_just}
\title{Checks if unit or number 0-1}
\usage{
assert_unit(var, allow_missing = TRUE)

assert_label(var, allow_missing = TRUE)

assert_just(var, allow_missing = TRUE)
}
\arguments{
\item{var}{Variable 2 check}

\item{allow_missing}{Allow missing variables}
}
\value{
void

void

void
}
\description{
Checks if unit or number 0-1

Assert that its a string or number

Assert valid color
}
\keyword{internal}
