% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs.R
\name{print.box}
\alias{plot.box}
\alias{print.box}
\title{The print/plot calls the \code{\link[grid]{grid.draw}} function on the object}
\usage{
\method{print}{box}(x, ...)

\method{plot}{box}(x, ...)
}
\description{
The print/plot calls the \code{\link[grid]{grid.draw}} function on the object
}

