% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs.R
\name{boxGrob}
\alias{boxGrob}
\alias{plot.box}
\alias{print.box}
\title{Create a grid box with text}
\usage{
boxGrob(label, y = unit(0.5, "npc"), x = unit(0.5, "npc"), width, height,
  just = "center", bjust = "center", txt_gp = gpar(color = "black"),
  box_gp = gpar(fill = "#D8F0D1"))

\method{print}{box}(x, ...)

\method{plot}{box}(x, ...)
}
\arguments{
\item{fill}{}
}
\description{
Create a grid box with text

The print/plot calls the \code{\link[grid]{grid.draw}} function on the object
}

