% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_t.R
\name{ci_t}
\alias{ci_t}
\title{Confidence interval for mean of a sample}
\usage{
ci_t(mu, s, n, conf.level = 0.95)
}
\arguments{
\item{mu}{A point estimate of the population mean (a vector of values can also be used)}

\item{s}{A point estimate of the sample standard deviation (a vector of values can also be used)}

\item{n}{Sample size (a vector of values can also be used)}

\item{conf.level}{Confidence level to use for the confidence interval (Default is 0.95)}
}
\value{
A data frame with the mean, degrees of freedom, and lower and upper ends of the confidence interval
}
\description{
Calculates the confidence interval for a population mean with known mean and unknown standard deviation, using the t-statistic (student-t distribution).
}
\examples{
mu <- 80
s <- 15
n <- 20
ci_t(mu, s, n)

}
