% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_t2.R
\name{ci_t2}
\alias{ci_t2}
\title{Confidence interval for difference of the mean between 2 samples}
\usage{
ci_t2(mu1, s1, n1, mu2, s2, n2, conf.level = 0.95)
}
\arguments{
\item{mu1}{A point estimate of sample's 1 mean}

\item{s1}{A point estimate of sample's 1 standard deviation}

\item{n1}{Sample size of sample 1}

\item{mu2}{A point estimate of sample's 2 mean}

\item{s2}{A point estimate of sample's 2 standard deviation}

\item{n2}{Sample size of sample 2}

\item{conf.level}{Confidence level to use for the confidence interval (Default is 0.95)}
}
\value{
A data frame with the difference of means, degrees of freedom, and lower and upper ends of the confidence interval
}
\description{
Calculates the confidence interval for the difference of the mean between 2 samples (student-t distribution).
}
\examples{
mu1 <- 90
s1 <- 15
n1 <- 20
mu2 <- 80
s2 <- 8
n2 <- 30
ci_t2(mu1, s1, n1, mu2, s2, n2)

}
