% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_z.R
\name{ci_z}
\alias{ci_z}
\title{Confidence interval for mean (known sigma)}
\usage{
ci_z(x, sig, n, conf.level = 0.95, digits = 2)
}
\arguments{
\item{x}{Sample mean}

\item{sig}{Population standard deviation}

\item{n}{Sample size}

\item{conf.level}{Confidence level to use for the confidence interval (Default is 0.95)}

\item{digits}{Number of digits to round to (Default is 2)}
}
\value{
A data frame with the mean, and lower and upper ends of the confidence interval
}
\description{
Calculates the confidence interval for the mean with known standard deviation, using the Z-statistic (normal distribution).
}
\examples{
x <- 80
sig <- 15
n <- 20
ci_z(x, sig, n)

}
