% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs.R
\name{coords}
\alias{coords}
\title{Get the box coordinates}
\usage{
coords(box)
}
\arguments{
\item{box}{The boxGrob}
}
\value{
A list with the cooordinates
}
\description{
Retrieves the boxes \code{"coords"} attribute.
}
\examples{
box <- boxGrob("A test box")
coords(box)
}

