% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs_coords.R
\name{coords}
\alias{coords}
\title{Get the box coordinates}
\usage{
coords(box)
}
\arguments{
\item{box}{The \code{\link{boxGrob}} or \code{\link{boxPropGrob}}}
}
\value{
A list with the coordinates
}
\description{
Retrieves the boxes \code{"coords"} attribute.
}
\examples{
box <- boxGrob("A test box")
coords(box)
}
\seealso{
Other flowchart components: 
\code{\link{align}},
\code{\link{boxGrob}()},
\code{\link{boxPropGrob}()},
\code{\link{connectGrob}()},
\code{\link{distance}()},
\code{\link{moveBox}()},
\code{\link{spread}}
}
\concept{flowchart components}
