% Generated by roxygen2 (4.0.1): do not edit by hand
\name{describeMean}
\alias{describeMean}
\title{Describe the mean}
\usage{
describeMean(x, html = FALSE, digits = 1, number_first = TRUE,
  show_missing, show_missing.digits = digits, percentage_sign = TRUE,
  plusmin_str, language = "en", ...)
}
\arguments{
  \item{x}{The variable that you want the statistics for}

  \item{digits}{The number of decimals used}

  \item{html}{If HTML compatible output shoudl be used
  instead of default LaTeX}

  \item{number_first}{If the number should be given or if
  the percentage should be presented first. The second is
  encapsulated in parentheses (). This is only used
  together with the show_missing variable.}

  \item{show_missing}{This indicates if missing should be
  added as a separate row below all other.}

  \item{show_missing.digits}{The number of digits to use
  for the missing percentage, defaults to the overall
  \code{digits}.}

  \item{percentage_sign}{If you want to suppress the
  percentage sign you can set this variable to FALSE. You
  can also choose something else that the default \% if you
  so wish by setting this variable. Note, this is only used
  when combined with the missing information.}

  \item{plusmin_str}{Provide if you want anything other
  than the plus minus sign suited for the given output
  format.}

  \item{language}{The ISO-639-1 two-letter code for the
  language of interest. Currently only english is
  distinguished from the ISO format using a ',' as the
  separator in the \code{\link{outputInt}} function.}

  \item{...}{Passed on to \code{\link{describeFactors}}}
}
\value{
\code{string} A string formatted for either LaTeX or HTML
}
\description{
A function that returns a description of a continuous
variable using the mean together with the standard
deviation. The standard deviation is used as it is
"industry standard" to use mean with standard deviation and
not because it's the only option.
}
\examples{
describeMean(1:10)
describeMean(c(1:10, NA), show_missing=TRUE)
}
\seealso{
\code{\link{getDescriptionStatsBy}}

Other description functions: \code{\link{describeFactors}};
\code{\link{describeMedian}}; \code{\link{describeProp}}
}

