% Generated by roxygen2 (4.0.1): do not edit by hand
\name{describeProp}
\alias{describeProp}
\title{A function that returns a description proportion that contains
the number and the percentage}
\usage{
describeProp(x, html = FALSE, digits = 1, number_first = TRUE,
  show_missing = FALSE, show_missing.digits = digits, default_ref,
  percentage_sign = TRUE, language = "en", ...)
}
\arguments{
  \item{x}{The variable that you want the statistics for}

  \item{html}{If HTML compatible output shoudl be used
  instead of default LaTeX}

  \item{digits}{The number of decimals used}

  \item{number_first}{If the number should be given or if
  the percentage should be presented first. The second is
  encapsulated in parentheses (). This is only used
  together with the show_missing variable.}

  \item{show_missing}{This indicates if missing should be
  added as a separate row below all other.}

  \item{show_missing.digits}{The number of digits to use
  for the missing percentage, defaults to the overall
  \code{digits}.}

  \item{percentage_sign}{If you want to suppress the
  percentage sign you can set this variable to FALSE. You
  can also choose something else that the default \% if you
  so wish by setting this variable. Note, this is only used
  when combined with the missing information.}

  \item{language}{The ISO-639-1 two-letter code for the
  language of interest. Currently only english is
  distinguished from the ISO format using a ',' as the
  separator in the \code{\link{outputInt}} function.}

  \item{...}{Passed on to \code{\link{describeFactors}}}

  \item{default_ref}{The default reference, either first,
  the level name or a number within the levels. If left out
  it defaults to the first value.}
}
\value{
\code{string} A string formatted for either LaTeX or HTML
}
\description{
A function that returns a description proportion that
contains the number and the percentage
}
\examples{
describeProp(factor(sample(50, x=c("A","B"), replace=TRUE)))
describeProp(factor(sample(50, x=c("A","B", NA), replace=TRUE)), show_missing=TRUE)
}
\seealso{
Other description functions: \code{\link{describeFactors}};
\code{\link{describeMean}}; \code{\link{describeMedian}}
}

