% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier_private_functions.R
\name{getBezierAdj4Arrw}
\alias{getBezierAdj4Arrw}
\title{Gets the bezier points adjusted for an arrow}
\usage{
getBezierAdj4Arrw(x, y, arrow_length, length_out = 100)
}
\arguments{
\item{x}{The x start and end points}

\item{y}{The spline control points}

\item{arrow_length}{The desired length of the arrow}

\item{length_out}{Increases the resolution for the final bezier
points, i.e. generating more fine-grained intervals}
}
\value{
list
}
\description{
Gets the bezier points adjusted for an arrow
}
