% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getGridVal}
\alias{getGridVal}
\title{Gets grid value}
\usage{
getGridVal(x, default.units, axisTo = "x")
}
\arguments{
  \item{x}{Value}

  \item{default.units}{The unit type}

  \item{axisTo}{The axis that is used, useful for "npc"
  where there is a big difference in height and width}
}
\value{
float
}
\description{
Returns the raw value in units if the provided is of type
unit else it returns the default unit.
}
\author{
Max
}
\keyword{internal}

