% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier_private_functions.R
\name{getVectorAngle}
\alias{getVectorAngle}
\title{Gets an angle}
\usage{
getVectorAngle(x, y, x_origo = NA, y_origo = NA)
}
\arguments{
\item{x}{The x point of the vector}

\item{y}{The y point of the vector}

\item{x_origo}{The x origin if other than 0}

\item{y_origo}{The y origin if other than 0}
}
\value{
angle in radians (see \code{\link{atan2}})
}
\description{
Uses a vector to get an angle by \code{\link{atan2}}.
}
\keyword{internal}
