\name{getXTicks}
\alias{getXTicks}
\title{Gets the x-ticks in a formatted version}
\usage{
  getXTicks(low, high, clip)
}
\arguments{
  \item{low}{lower bound}

  \item{high}{upper bound}

  \item{clip}{if the ci are clipped}
}
\value{
  Returns a vector with the ticks apropriate
}
\description{
  Gets the x-ticks in a formatted version
}
\author{
  max
}

