% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{Gmisc-deprecated}
\alias{Gmisc-deprecated}
\alias{forestplot2}
\title{Deprecated function(s) in \pkg{Gmisc}}
\usage{
forestplot2(...)
}
\arguments{
\item{...}{Parameters to be passed to the new version of the function}
}
\description{
The \code{forestplot2} has moved to \pkg{forestplot} package and changed name to \code{\link[forestplot]{forestplot}}
}
\keyword{internal}
