% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{Gmisc-deprecated}
\alias{Gmisc-deprecated}
\alias{Gmisc-deprecated-package}
\alias{forestplot2}
\title{Deprecated function(s) in \pkg{Gmisc}}
\usage{
forestplot2(...)
}
\arguments{
  \item{...}{Parameters to be passed to the new version of
  the function}
}
\description{
The \code{forestplot2} has move to \pkg{forestplot} package
and changed name to \code{\link[forestplot]{forestplot}}
}
\keyword{internal}

