% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{Gmisc-package}
\alias{Gmisc-package}
\title{A package for plotting, tables, and basic outputs.}
\description{
This is a collection of functions that I've found useful in
my research. The package is inspired by Frank Harrel's
Hmisc package. The main focus is on tables and plots.
}
\section{The htmlTable}{
  For tables you'll find the convenient
  \code{\link{htmlTable}} that I have used for advanced
  table layout. A major focus has been to have it
  compatible with LibreOffice (you can copy/past from there
  into word) as I generally want to be able to send my
  documents to a journal in .doc/.docx format. \bold{Note:}
  it is now often possible to copy->paste directly from the
  viewer into a MS Word document without any layout loss.
}

\section{Description functions}{
  To accompany the \code{\link{htmlTable}} function is the
  \code{\link{getDescriptionStatsBy}} it is a straight
  forward function that aims at helping you to generate
  descriptive table stratified by different variables. In
  other words, the function returns everything you need for
  generating a \emph{Table 1} ready for publication.
}

\section{Plots}{
  The main plots are \code{\link{forestplot2}} and
  \code{\link{transitionPlot}}. The forest plot is a more
  general version of the original
  \code{\link[rmeta]{forestplot}} function aimed at using
  the forest plot for more than just meta-analyses.

  The \code{\link{transitionPlot}} is for descriptive
  purposes in order to illustrate the size of change
  between one state and the next, i.e. a transition. This
  is basically a graph of \code{table(var1, var2)}. It uses
  the \code{\link{bezierArrowSmpl}} alternatives for
  showing off elegant arrows that should nicely bend into
  each box.

  Additionally there is the
  \code{\link{getSvdMostInfluential}} function that tries
  to illustrate and use the SVD in order to select the ones
  having most influence on the V-matrix.
}

