\name{Gmisc-package}
\alias{Gmisc-package}
\alias{Gmisc}
\docType{package}
\title{
\packageTitle{Gmisc}
}
\description{
\packageDescription{Gmisc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Gmisc}
\packageIndices{Gmisc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Gmisc}

Maintainer: \packageMaintainer{Gmisc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
