% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmisc.R
\docType{package}
\name{gmisc}
\alias{gmisc}
\alias{gmisc-package}
\title{gmisc: A Collection of Miscellaneous Functions}
\description{
gmisc: A Collection of Miscellaneous Functions
}
\section{Graphics functions}{


prtrellis - show many lattice plots on one page
}

