\name{heightDetails.xaxis}
\alias{heightDetails.xaxis}
\title{Gets the height for an x-axis object}
\usage{
\method{heightDetails}{xaxis}(x)
}
\arguments{
  \item{x}{The \code{\link[grid]{xaxisGrob}} object}
}
\value{
A unit object
}
\description{
A hidden function that gets the height. It is for some
reason not included by default in the \code{grid}-package.
}
\author{
Max
}

