% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDesc.R
\name{latex.descMrg}
\alias{latex.descMrg}
\alias{htmlTable.descMrg}
\alias{print.descMrg}
\alias{knit_print.descMrg}
\title{A wrapper around the \code{\link[htmlTable]{htmlTable}}/\code{\link[Hmisc]{latex}} function}
\usage{
\method{latex}{descMrg}(x, ...)

\method{htmlTable}{descMrg}(x, ...)

\method{print}{descMrg}(x, ...)

\method{knit_print}{descMrg}(x, ...)
}
\arguments{
\item{x}{The \code{\link{mergeDesc}} object}

\item{...}{Passed onto \code{\link[htmlTable]{htmlTable}}/\code{\link[Hmisc]{latex}}}
}
\description{
Calls the \code{\link[htmlTable]{htmlTable}}/\code{\link[Hmisc]{latex}} after extracting the
\code{rgroup}/\code{n.rgroup} arguments.
}
\keyword{internal}
