% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/htmlTable_tableOne.R
\name{htmlTable.TableOne}
\alias{htmlTable.TableOne}
\title{A function for converting a table generated from the \pkg{tablone} package}
\usage{
\method{htmlTable}{TableOne}(x, ..., css.rgroup = "")
}
\arguments{
  \item{x}{The TableOne object}

  \item{...}{Passed onto
  \code{\link[htmlTable]{htmlTable}}}
}
\value{
Returns an \code{\link[htmlTable]{htmlTable}} object
}
\description{
This function translates the \code{tableone}'s descriptive
table into an \code{\link[htmlTable]{htmlTable}}-object.
The function uses the raw output from the
\code{print.TableOne} and looks for \code{header},
\code{rgroup}, and \code{n.rgroup} elements.
}

