% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs_s3_append.R
\name{insert}
\alias{insert}
\alias{insert.default}
\alias{insert.Gmisc_list_of_boxes}
\title{Insert element into a list of boxes (S3)}
\usage{
insert(x, element, ..., after = NULL, before = NULL)

\method{insert}{default}(x, element, ..., after = NULL, before = NULL)

\method{insert}{Gmisc_list_of_boxes}(x, element, ..., after = NULL, before = NULL)
}
\arguments{
\item{x}{A `Gmisc_list_of_boxes`.}

\item{element}{The new box object to insert.}

\item{...}{Not used.}

\item{after}{The name or index of the box after which to insert.}

\item{before}{The name or index of the box before which to insert.}
}
\value{
The updated list of boxes with the new element inserted.
}
\description{
Inserts a box into a list of boxes and positions it relative to the
surrounding boxes.
}
\seealso{
Other flowchart components: 
\code{\link{align}()},
\code{\link{append}()},
\code{\link{boxGrob}()},
\code{\link{boxHeaderGrob}()},
\code{\link{boxPropGrob}()},
\code{\link{boxShapes}},
\code{\link{connectGrob}()},
\code{\link{coords}()},
\code{\link{distance}()},
\code{\link{flowchart}()},
\code{\link{move}()},
\code{\link{moveBox}()},
\code{\link{spread}()}
}
\concept{flowchart components}
