% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier_private_functions.R
\name{isHorizontal}
\alias{isHorizontal}
\title{Just a simple help with interpreting the radians}
\usage{
isHorizontal(angle_radian)
}
\arguments{
\item{angle_radian}{The angle in radians}
}
\value{
boolean
}
\description{
Just a simple help with interpreting the radians
}
\keyword{internal}
