% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_window.R
\name{layers_window}
\alias{layers_window}
\title{Plots a layered model and summarizes the statistics for each layer}
\usage{
layers_window(x, breaks)
}
\arguments{
\item{x}{A data frame containing the location variable (depth or distance) in the first column, and the value(s) of interest in the rest of the columns}

\item{breaks}{A vector containg the breakpoints (from 'RI', 'Cohen d' or 'Mahalanobis D2')}
}
\value{
A ggplot and plotly objects showing the layered model, another showing the confidence intervals, and a summary table
}
\description{
Given a set of breakpoints (depths/distances), plots a layered model of the data against distance, plots the confidence intervals for each layer, and gives a summary table.
}
\examples{
layers_window(DPM_data, breaks = c(3.8,8.9))
layers_window(CPTu_data, breaks = c(1.2,3.8,5.1))

}
