% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeLists.R
\name{mergeLists_internal}
\alias{mergeLists_internal}
\title{A helper function for mergeLists that does the actual work.}
\usage{
mergeLists_internal(o_element, n_element)
}
\arguments{
\item{o_element}{Old element}

\item{n_element}{New element to add to the old element}
}
\value{
list with the two merged
}
\description{
A helper function for mergeLists that does the actual work.
}
\keyword{internal}
