% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs_s3_move.R
\name{move}
\alias{move}
\alias{move.default}
\alias{move.Gmisc_list_of_boxes}
\title{Move boxes (S3)}
\usage{
move(element, ...)

\method{move}{default}(element, ...)

\method{move}{Gmisc_list_of_boxes}(element, subelement = NULL, ...)
}
\arguments{
\item{element}{A `list` of boxes.}

\item{...}{Arguments passed to [`moveBox`].}

\item{subelement}{The name of the element to move (maps to `subelement` in `moveBox`)}
}
\value{
The moved list of boxes.
}
\description{
Move a box or list of boxes. A wrapper for [`moveBox`] that supports
piping of `Gmisc_list_of_boxes` while preserving attributes.
}
\seealso{
Other flowchart components: 
\code{\link{align}()},
\code{\link{append}()},
\code{\link{boxGrob}()},
\code{\link{boxHeaderGrob}()},
\code{\link{boxPropGrob}()},
\code{\link{boxShapes}},
\code{\link{connectGrob}()},
\code{\link{coords}()},
\code{\link{distance}()},
\code{\link{flowchart}()},
\code{\link{insert}()},
\code{\link{moveBox}()},
\code{\link{spread}()}
}
\concept{flowchart components}
