% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper_diagram.R
\name{piper_diagram}
\alias{piper_diagram}
\title{Piper diagram}
\usage{
piper_diagram(data, group = NULL)
}
\arguments{
\item{data}{A data frame containing the concentrations of cations (Ca,Mg,Na,K) and anions (SO4,Cl,HCO3,CO3) in mg/L}

\item{group}{Grouping variable in the data frame, name in quotes. If not provided the samples will be numbered sequentially}
}
\value{
ggplot object
}
\description{
Plots hydrochemical data into a Piper diagram.
}
\examples{
data = data.frame(Group = 'Test',Ca = c(120,150,110,52.6), Mg = c(78,160,110,28), Na = c(210,590,340,51.6), K = c(4.2,2,3.6,2.3), HCO3 = c(181,181,189,151), CO3 = 0, Cl = c(220,744,476,72.2), SO4 = c(560,1020,584,126))
piper_diagram(data, 'Group')
piper_diagram(data)

}
