% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_summary.R
\name{post_summary}
\alias{post_summary}
\title{Summary of vectors, matrices or data frames}
\usage{
post_summary(x, probs = seq(0, 1, 0.25), digits = 3)
}
\arguments{
\item{x}{A vector, matrix or data frame}

\item{probs}{Desired quantiles to display}

\item{digits}{Number of significant digits to display}
}
\value{
A data frame with summary statistics
}
\description{
Summarises vectors, matrices or data frames by showing their mean, gometric mean, median, standard deviation, median absolute deviation,inter quartile range, and desired quantiles.
}
\examples{
x1 <- rnorm(100, 30, 6)
x2 <- matrix(x1, nrow = 10)
x3 <- data.frame(N = rnorm(100, 30, 6), T = rt(100, 10, 30))
post_summary(x1)
post_summary(x2)
post_summary(x3)

}
