% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDescriptionStatsBy_prAddDescUnitColumn.R
\name{prAddDescUnitColumn}
\alias{prAddDescUnitColumn}
\title{Add a units column to the results}
\usage{
prAddDescUnitColumn(results, x, use_units, units_column_name)
}
\arguments{
\item{results}{The results that we want to add the column to}

\item{x}{If a data.frame it will be used as the data source for the variables in the \code{...} parameter.
If it is a single variable it will be the core value that want the statistics for.
In the print this is equivalent to the output of this function.}

\item{use_units}{If the Hmisc package's units() function has been employed
it may be interesting to have a column at the far right that indicates the
unit measurement. If this column is specified then the total column will
appear before the units (if specified as last). You can also set the value to
\code{"name"} and the units will be added to the name as a parenthesis,
e.g. Age (years).}

\item{units_column_name}{The name of the units column. Used if use_units = TRUE}
}
\value{
results with added column
}
\description{
Add a units column to the results
}
