% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_document.R
\name{prCaptionFix}
\alias{prCaptionFix}
\title{Fixes the caption for elements without caption}
\usage{
prCaptionFix(outFile)
}
\arguments{
\item{outFile}{The name of the file}
}
\value{
outFile The name of the file
}
\description{
Fixes the caption for elements without caption
}
\keyword{internal}
