% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrob_prConvert2Coords.R
\name{prConvert2Coords}
\alias{prConvert2Coords}
\title{Converts an object to coordinates}
\usage{
prConvert2Coords(obj)
}
\arguments{
\item{obj}{A \code{\link{boxGrob}}, \code{\link{boxPropGrob}}, \code{\link{coords}} output, \code{\link[grid:unit]{unit}} or a number ranging to
be converted to a \code{npc} \code{\link[grid:unit]{unit}}}
}
\value{
A \code{list} with all the points that \code{\link{coords}} returns
}
\description{
Sometimes we have an object that can be either a box,
a coprdinate, a unit or a numerical value and all we
want is a \code{list} of coordinates that we can use for
calculating distance, alignment and other things.
}
