% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs_prCreateBoxCoordinates.R
\name{prCreateBoxCoordinates}
\alias{prCreateBoxCoordinates}
\title{Creates coordinates for box}
\usage{
prCreateBoxCoordinates(viewport_data, extra_coordinate_functions = NULL)
}
\arguments{
\item{viewport_data}{The arguments that will be used for generating the \code{viewport}}

\item{extra_coordinate_functions}{A list with named functions if we want additional
parameters}
}
\value{
\code{list} of class \code{coords}
}
\description{
Creates coordinates for box
}
