% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forestplot2_helpers.R
\name{prFpConvertMultidimArray}
\alias{prFpConvertMultidimArray}
\title{Converts a 2D or 3D array to mean, lower, upper}
\usage{
prFpConvertMultidimArray(x)
}
\arguments{
  \item{x}{The array to convert}
}
\value{
\code{list(mean = mean, lower = lower, upper = upper)}
}
\description{
Converts a 2D or 3D array to mean, lower, upper
}
\keyword{internal}

