% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prFpDrawLegend}
\alias{prFpDrawLegend}
\title{Draw the forestplot legend}
\usage{
prFpDrawLegend(lGrobs, legend.pos, col, colgap, legend.gp, legend.r,
  legend.padding, legendMarkerFn, ...)
}
\arguments{
  \item{lGrobs}{A list with all the grobs, see
  \code{\link{prFpGetLegendGrobs}}}

  \item{legend.pos}{Specifies if the legend is horizontal
  or not. Can either be a list or a string.}

  \item{col}{The colors of the legends.}

  \item{colgap}{The gap between the box and the text}

  \item{legend.gp}{The \code{\link[grid]{gpar}} options for
  background fill, border etc. If NULL this is not used,}

  \item{legend.r}{The radius for the box if any (see
  \code{\link[grid]{grid.roundrect}})}

  \item{legend.padding}{The padding for the legend box,
  only used if box is drawn. This is the distance from the
  border to the text/boxes of the legend.}

  \item{legendMarkerFn}{The function for drawing the
  marker}

  \item{...}{Passed to the legend \code{legendMarkerFn}}
}
\value{
\code{void}
}
\description{
Takes the grobs and outputs the legend inside the current
viewport.
}
\author{
max
}
\keyword{internal}

