% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prFpFindWidestGrob}
\alias{prFpFindWidestGrob}
\title{Finds the widest grob in the current list of grobs}
\usage{
prFpFindWidestGrob(grob.list, return_unit = "mm")
}
\arguments{
  \item{grob.list}{A list of grobs}

  \item{return_unit}{A valid \code{\link[grid]{unit}}
  specifier}
}
\value{
\code{grid::unit} Returns the widest grob and its width
}
\description{
Finds the widest grob in the current list of grobs
}

