% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forestplot2_helpers.R
\name{prFpFindWidestGrob}
\alias{prFpFindWidestGrob}
\title{Finds the widest grob in the current list of grobs}
\usage{
prFpFindWidestGrob(grob.list, return_unit = "mm")
}
\arguments{
  \item{grob.list}{A list of grobs}

  \item{return_unit}{A valid \code{\link[grid]{unit}}
  specifier}
}
\value{
\code{grid::unit} Returns the width
\code{\link[grid]{unit}} for the widest grob
}
\description{
Finds the widest grob in the current list of grobs
}
\keyword{internal}

