% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forestplot2_helpers.R
\name{prFpGetGraphTicksAndClips}
\alias{prFpGetGraphTicksAndClips}
\title{A helper function to forestplot2}
\usage{
prFpGetGraphTicksAndClips(xticks, xticks.digits, xlog, xlab, lwd.xaxis, col,
  cex, cex.axis, clip, zero, x_range, nc, mean)
}
\arguments{
  \item{xticks}{Optional user-specified x-axis tick marks.
  Specify NULL to use the defaults, numeric(0) to omit the
  x-axis.}

  \item{xticks.digits}{The number of digits to allow in the
  x-axis if this is created by default.}

  \item{xlog}{If TRUE, x-axis tick marks are exponentiated}

  \item{xlab}{x-axis label}

  \item{lwd.xaxis}{lwd for the xaxis}

  \item{col}{See \code{\link{fpColors}}}

  \item{cex}{The font adjustment}

  \item{cex.axis}{The font adjustment for the x-xaxis,
  defaults to 60 \% of the cex parameter.}

  \item{clip}{Lower and upper limits for clipping
  confidence intervals to arrows}

  \item{zero}{x-axis coordinate for zero line. If you
  provide a vector of length 2 it will print a rectangle
  instead of just a line.}

  \item{x_range}{The range that the values from the
  different confidence interval span}

  \item{nc}{Number of columns}

  \item{mean}{The original means, either matrix or vector}
}
\value{
\code{list} Returns a list with axis_vp, axisGrob, labGrob,
zero and clip
}
\description{
Gets the x-label and zero-bar details
}
\keyword{internal}

