% Generated by roxygen2 (4.0.0): do not edit by hand
\name{prFpGetLegendGrobs}
\alias{prFpGetLegendGrobs}
\title{Gets the forestplot legend grobs}
\usage{
prFpGetLegendGrobs(legend, legend.cex, legend.title = NULL)
}
\arguments{
  \item{legend}{The legend names}

  \item{legend.cex}{The cex for the text size}

  \item{legend.title}{The title of the legend if any}
}
\value{
\code{list} A "Legend" class that derives from a list with
all the different legends. The list also contains
attributes such as height, width, max_height, max_width,
line_height_and_spacing. The title of the legend is saved
inside \code{attr("title")}
}
\description{
Gets the forestplot legend grobs
}
\author{
max
}
\keyword{internal}

