% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forestplot2_helpers.R
\name{prFpGetLegendGrobs}
\alias{prFpGetLegendGrobs}
\title{Gets the forestplot legend grobs}
\usage{
prFpGetLegendGrobs(legend, cex, title)
}
\arguments{
  \item{legend}{Legend corresponding to the number of
  bars.}

  \item{cex}{The font adjustment}

  \item{title}{The title of the legend if any}
}
\value{
\code{list} A "Legend" class that derives from a list with
all the different legends. The list also contains
attributes such as height, width, max_height, max_width,
line_height_and_spacing. The title of the legend is saved
inside \code{attr("title")}
}
\description{
Gets the forestplot legend grobs
}
\keyword{internal}

