\name{prFpPrepareLegendMarker}
\alias{prFpPrepareLegendMarker}
\title{Prepares the legend marker function}
\usage{
prFpPrepareLegendMarker(legendMarkerFn, col_no, confintNormalFn)
}
\arguments{
  \item{legendMarkerFn}{The unknown parameter}

  \item{col_no}{The number of columns}

  \item{confintNormalFn}{The original confintNormalFn
  input}
}
\value{
\code{list}
}
\description{
Prepares the legend marker function
}
\author{
Max
}

