% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forestplot2_helpers.R
\name{prFpPrepareLegendMarker}
\alias{prFpPrepareLegendMarker}
\title{Prepares the legend marker function}
\usage{
prFpPrepareLegendMarker(legendMarkerFn, col_no, confintNormalFn)
}
\arguments{
  \item{legendMarkerFn}{The unknown parameter}

  \item{col_no}{The number of columns}

  \item{confintNormalFn}{The original confintNormalFn
  input}
}
\value{
\code{list}
}
\description{
Prepares the legend marker function
}
\keyword{internal}

