% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forestplot2_helpers.R
\name{prFpPrintLabels}
\alias{prFpPrintLabels}
\title{Plots the labels}
\usage{
prFpPrintLabels(labels, nc, nr)
}
\arguments{
  \item{labels}{A list to the labels}

  \item{nc}{Number of columns}

  \item{nr}{Number of rows}
}
\value{
\code{void}
}
\description{
This is a helper function to the \code{\link{forestplot2}}
function.
}
\keyword{internal}

