% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forestplot2_helpers.R
\name{prFpPrintXaxis}
\alias{prFpPrintXaxis}
\title{Plots the x-axis for forestplot2}
\usage{
prFpPrintXaxis(axisList, col, lwd.zero)
}
\arguments{
  \item{axisList}{The list from
  \code{\link{prFpGetGraphTicksAndClips}}}

  \item{col}{See \code{\link{fpColors}}}

  \item{lwd.zero}{lwd for the vertical line that gives the
  no-effect line}
}
\value{
void
}
\description{
A helper function to the \code{\link{forestplot2}}
function.
}
\keyword{internal}

