% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forestplot2_helpers.R
\name{prFpValidateLabelList}
\alias{prFpValidateLabelList}
\title{Validate the forestplot label list}
\usage{
prFpValidateLabelList(labelList)
}
\arguments{
  \item{labelList}{The list of labels}
}
\value{
\code{boolean} TRUE or FALSE
}
\description{
Checks that all list elements have equal length, i.e. there
is a m x n relation
}
\keyword{internal}

