% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prFpXrange}
\alias{prFpXrange}
\title{Gets the x-axis range}
\usage{
prFpXrange(upper, lower, clip, zero, xticks, xlog)
}
\arguments{
  \item{upper}{The upper bound of the confidence interval
  for the forestplot, needs to be the same format as the
  mean, i.e. matrix/vector of equal columns & length}

  \item{lower}{The lower bound of the confidence interval
  for the forestplot, needs to be the same format as the
  mean, i.e. matrix/vector of equal columns & length}

  \item{clip}{Lower and upper limits for clipping
  confidence intervals to arrows}

  \item{zero}{x-axis coordinate for zero line}

  \item{xticks}{Optional user-specified x-axis tick marks.
  Specify NULL to use the defaults, numeric(0) to omit the
  x-axis.}

  \item{xlog}{If TRUE, x-axis tick marks are exponentiated}
}
\value{
\code{vector} Contains a min and max value
}
\description{
If the borders are smaller than the upper/lower limits then
clip the graph. The line will have arrows indicating that
it continues beyond the graph The zero bar has to be on the
chart though!
}
\keyword{internal}

