% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_document.R
\name{prFtpOtherRemoval}
\alias{prFtpOtherRemoval}
\title{Removes other unwanted lines}
\usage{
prFtpOtherRemoval(output_str)
}
\arguments{
\item{output_str}{The input from readLines()}
}
\value{
string Returns without the unwanted lines
}
\description{
Removes other unwanted lines
}
\keyword{internal}
