% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_document.R
\name{prFtpScriptRemoval}
\alias{prFtpScriptRemoval}
\title{Removes the <script>*</scrip>}
\usage{
prFtpScriptRemoval(output_str)
}
\arguments{
\item{output_str}{The input from readLines()}
}
\value{
string Returns without the script elements
}
\description{
Removes the <script>*</scrip>
}
\keyword{internal}
