\name{prGetModelVariables}
\alias{prGetModelVariables}
\title{Get the models variables}
\usage{
  prGetModelVariables(model, remove_splines = TRUE)
}
\arguments{
  \item{model}{A model fit}

  \item{remove_splines}{If splines, etc. should be cleaned
  from the variables as these no longer are "pure"
  variables}
}
\value{
  vector with names
}
\description{
  Get the models variables
}
\author{
  max
}

