% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prHtGetAlign}
\alias{prHtGetAlign}
\title{Gets alignment}
\usage{
prHtGetAlign(align, index)
}
\arguments{
  \item{align}{a character strings specifying column
  alignments, defaulting to \code{paste(c("l",
  rep('c',ncol(n_table)-1)),collapse='')} to center. Valid
  alignments are l = left, c = center and r = right. You
  can also specify \code{align='c|c'} and other LaTeX
  tabular formatting.}

  \item{index}{The index of the align parameter of
  interest}
}
\description{
Gets alignment
}
\keyword{internal}

