% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGrobs_print.Gmisc_list_of_boxes.R
\name{print.Gmisc_list_of_boxes}
\alias{print.Gmisc_list_of_boxes}
\title{Output boxes}
\usage{
\method{print}{Gmisc_list_of_boxes}(x, ...)
}
\arguments{
\item{x}{A list of a set of [`boxGrob`]/[`boxPropGrob`] to plot}

\item{...}{Ignored argument}
}
\description{
Outputs a list of boxes as produced by either the
spread or align functions for boxGrobs.
}
