% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{prPushMarginViewport}
\alias{prPushMarginViewport}
\title{Pushes viewport with margins}
\usage{
prPushMarginViewport(bottom, left, top, right, name = "margin")
}
\arguments{
\item{bottom}{The margin object, either in npc or a \code{\link[grid]{unit}} object,
or a vector of 4 if you want the same margins allover.}

\item{left}{The margin object, either in npc or a \code{\link[grid]{unit}} object.}

\item{top}{The margin object, either in npc or a \code{\link[grid]{unit}} object}

\item{right}{The margin object, either in npc or a \code{\link[grid]{unit}} object}

\item{name}{The name of the last viewport}
}
\value{
\code{2} - the number of pushed viewports
}
\description{
A \code{\link[grid]{grid.layout}} object is used to
generate the margins. A second viewport selecting the
mid-row/col is used to create the effect of margins
}
\keyword{internal}
