% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionClass_helpers.R
\name{prTcPlotBoxColumn}
\alias{prTcPlotBoxColumn}
\title{Plots a column of boxes}
\usage{
prTcPlotBoxColumn(box_positions, proportions, fill, txt, txt_clr, cex)
}
\arguments{
\item{box_positions}{The box positions}

\item{proportions}{The box vertical proportions (needed for the prop attribute)}

\item{fill}{The fill colors of length equal to the proportions}

\item{txt}{The texts of length equal to the proportions}

\item{txt_clr}{The text colors of length equal to the proportions}

\item{cex}{The fontsize multiplier}
}
\value{
\code{void}
}
\description{
Takes a set of box settings and plots them.
}
\keyword{internal}
