% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionClass_helpers.R
\name{prTcValidateAndPrepClr}
\alias{prTcValidateAndPrepClr}
\title{Checks and prepares colors}
\usage{
prTcValidateAndPrepClr(value, no_cols, no_rows, is3D)
}
\arguments{
\item{value}{The color values}

\item{no_cols}{The no. of columns in the transition matrix}

\item{no_rows}{All the row counts in the transition plot}

\item{is3D}{Whether the transition matrix has 3 dimensions}
}
\value{
\code{matrix} Returns a matrix of the same size as the
 transition matrix
}
\description{
The colors need to match the size of the transition matrix
and therefore we need to have this input check. The output is
always a matrix of the exact same dimensions as the transition
matrix.
}
\keyword{internal}
