% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionPlot_helpers.R
\name{prTpGetBoxPositions}
\alias{prTpGetBoxPositions}
\title{Gets the box position}
\usage{
prTpGetBoxPositions(
  no,
  side,
  transitions,
  prop_start_sizes,
  prop_end_sizes,
  tot_spacing,
  box_width
)
}
\arguments{
\item{no}{The box number}

\item{side}{The right or left side}

\item{transitions}{The flows to or from}

\item{prop_start_sizes}{The size of the start boxes}

\item{prop_end_sizes}{The size of the end boxes}

\item{tot_spacing}{The proportion of the vertical space that is to be left
empty. It is then split evenly between the boxes.}

\item{box_width}{The width of the box. By default the box is one fourth of
the plot width.}

\item{no_boxes}{The number of boxes}
}
\value{
\code{list(top, left, bottom, right, width, height)}
}
\description{
Gets the box position
}
\keyword{internal}
