% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prTpGetBoxPositions}
\alias{prTpGetBoxPositions}
\title{Gets the box position}
\usage{
prTpGetBoxPositions(no, colNum, totalColNum, transitions, prop_sizes,
  tot_spacing, box_width)
}
\arguments{
  \item{no}{The box number}

  \item{colNum}{the box column number}

  \item{totalColNum}{the total number of box columns}

  \item{transitions}{The flows to or from}

  \item{prop_sizes}{The size of the start boxes}

  \item{tot_spacing}{The total space between the boxes}

  \item{box_width}{The width of the box}

  \item{prop_end_sizes}{The size of the end boxes}
}
\value{
\code{list(top, left, bottom, right, width, height)}
}
\description{
Gets the box position
}
\keyword{internal}

