% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionPlot_helpers.R
\name{prTpGetColors}
\alias{prTpGetColors}
\title{Gets a set of colors or just one color}
\usage{
prTpGetColors(colors, proportion, space = NULL, no = 101)
}
\arguments{
\item{colors}{A set of min. two colors that is used
for \code{\link[grDevices:colorRamp]{colorRampPalette}}.}

\item{proportion}{A proportion or a set of proportions
between 0 and 1. If you leave this out then the full color
span will be returned.}

\item{space}{If there is little or no difference
at the low/high proportions of the spectrum then it
can be of interest to focus the color change to the center
leaving the tails constant}

\item{The}{color resolution to use}
}
\value{
\code{character} The function can return both single colors
 or multiple colors as character vector (see \code{\link[grDevices]{rgb}})
}
\description{
Used in order to illustrate the mix between two
proportions in the \code{\link{transitionPlot}}.
}
\keyword{internal}
