% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionPlot_helpers.R
\name{prTpPlotBox}
\alias{prTpPlotBox}
\title{Plots the box for the transition plot}
\usage{
prTpPlotBox(bx, bx_txt, fill, txt_clr, cex, line_col, lwd, prop = NA)
}
\arguments{
\item{bx}{A list with the x, y, height and width parameters}

\item{bx_txt}{The box text}

\item{fill}{The fill color}

\item{txt_clr}{The text color}

\item{cex}{The font size}

\item{line_col}{The line color around the box}

\item{lwd}{The line width}

\item{prop}{Provide a proportion if the box should be split (0-1)}
}
\value{
\code{void}
}
\description{
Plots the box for the transition plot
}
\keyword{internal}
