% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_t.R
\name{pval_t}
\alias{pval_t}
\title{P-value for proposed mean of a population (unknown sigma)}
\usage{
pval_t(x, s, n, mu, side = c("two", "one"))
}
\arguments{
\item{x}{Sample mean}

\item{s}{Sample standard deviation}

\item{n}{Sample size}

\item{mu}{Population mean}

\item{side}{To either get a two-tail or one-tail p-value (Default is "two")}
}
\value{
P-value
}
\description{
Calculates the p-value for a proposed mean for a population with unknown standard deviation, using the t-statistic (t distribution).
}
\examples{
x <- 6.14
s <- 0.803
n <- 9
mu <- 7
pval_t(x, s, n, mu)

}
