% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_z.R
\name{pval_z}
\alias{pval_z}
\title{P-value for proposed mean of a population}
\usage{
pval_z(x, sig, n, mu = 0, side = c("two", "one"))
}
\arguments{
\item{x}{A point estimate of the population mean}

\item{sig}{A point estimate of the population standard deviation}

\item{n}{Sample size from the population}

\item{mu}{Proposed population mean to test for}

\item{side}{To either get a two-tail or one-tail p-value (Default is "two")}
}
\value{
P-value
}
\description{
Calculates the p-value for a proposed mean for a population with known standard deviation, using the Z-statistic (normal distribution).
}
\examples{
x <- 2.6
sig <- 0.3
n <- 36
mu <- 2.5
pval_z(x, sig, n, mu)

}
