% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose_diag_3D.R
\name{rose_diag_3D}
\alias{rose_diag_3D}
\title{Rose diagram (3D)}
\usage{
rose_diag_3D(dir, dip, width = 30, conf.level = 0.95)
}
\arguments{
\item{dir}{A vector of dip directions in degrees}

\item{dip}{A vector of dip angles in degrees}

\item{width}{Petal width}

\item{conf.level}{Confidence level to use for the cone of confidence (Default is 0.95)}
}
\value{
A ggplot obejct
}
\description{
Plots a rose diagram for angular (direction and dip) data.
}
\details{
The mean direction and cone of confidence are displayed as red and cyan lines respectively.
}
\examples{
dir = c(12,18,22,15,10,20)
dip = c(42,40,48,30,42,30)
rose_diag_3D(dir, dip)

}
