% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bezier_private_functions.R
\name{rotateWidthAccVector}
\alias{rotateWidthAccVector}
\title{Rotates a width to match a vector}
\usage{
rotateWidthAccVector(x, y, x_origo = NA, y_origo = NA, default.units, ...)
}
\arguments{
  \item{x}{The x point of the vector}

  \item{y}{The y point of the vector}

  \item{x_origo}{The x origin if other than 0}

  \item{y_origo}{The y origin if other than 0}

  \item{default.units}{The \code{\link[grid]{unit}} type}

  \item{...}{Passed on to \code{\link{rotateWidthAccAngle}}
  after calculating the angle from the vector.}
}
\value{
list A list with left & right vectors of c(x, y) format and
a angle element
}
\description{
This function is provided with a vector and a width. The
vector is for knowing the direction/direction and this is
used for generating an angle. It then calls the subfunction
\code{\link{rotateWidthAccAngle}}.
}
\keyword{internal}

