% Generated by roxygen2 (4.0.2): do not edit by hand
\name{splitLines4Table}
\alias{splitLines4Table}
\title{A helper function for html/LaTeX line formatting}
\usage{
splitLines4Table(..., html = TRUE)
}
\arguments{
  \item{...}{The lines that you want to be joined}

  \item{html}{If HTML compatible output should be used. If
  \code{FALSE} it outputs LaTeX formatting}
}
\value{
string
}
\description{
This function helps you to do a multiline table header in
both html and in LaTeX. In html this isn't that tricky, you
just use the <br /> command but in LaTeX I often find
myself writing vbox/hbox stuff and therefore I've created
this simple helper function
}
\examples{
splitLines4Table("hello", "world")
splitLines4Table("hello", "world", html=TRUE)
splitLines4Table("hello", "world", list("A list", "is OK"))
}

