% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereo_mean.R
\name{stereo_mean}
\alias{stereo_mean}
\title{Stereonet (mean)}
\usage{
stereo_mean(dir, dip, conf.level = 0.95, mean.col = "red", add = FALSE)
}
\arguments{
\item{dir}{A vector of dir directions in degrees}

\item{dip}{A vector of dip angles in degrees}

\item{conf.level}{Confidence level for the cone of confidence (Default is 0.95)}

\item{mean.col}{Color for the mean and cone of confidence}

\item{add}{A logical indicating to add or not another set of measurements and its calculations}
}
\value{
A stereonet plot of data points and its mean estimations
}
\description{
Draws the mean direction and cone of confidence for a set of measurements.
}
\examples{
dir <- runif(min = 30, max = 80, n = 20)
dip <- runif(min = 10, max = 60, n = 20)
stereo_mean(dir, dip)

}
