% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_stat.R
\name{T_stat}
\alias{T_stat}
\title{T-statistic for layer boundary determination}
\usage{
T_stat(x, k = 6)
}
\arguments{
\item{x}{A data frame containing the depth of perforation in the first column, and the value of interest in the second column}

\item{k}{The window length for the number of data points to include in the calculation of T-statistic. Always and even (par) number}
}
\value{
ggplot and plotly objects showing the Cohen's d statistic and lines marking the Cohen's U3 values for 0.95, 0.99, and 0.999, and suggested boundaries
}
\description{
Calculates the T-statistic and Cohen's d effect size for a perforation log. The coefficient is used to find layer boundaries in a perforation log.
}
\details{
The example data given is intended to show the structure needed for input data. The user should follow this structure, which in general corresponds with a data frame with a sequence in the first column and the observed/measured values in the second column
}
\examples{
T_stat(DPM_data, k = 6)

}
\references{
Mora, R. (2013). Uso de metodos estadisticos para la identifacion de capas de suelos volcanicos con el ensayo del cono de pentracion en los terrenos de la Universidad de Costa Rica, Montes de Oca, San Jose, Costa Rica. - Rev. Geol. Amer. Central, 49: 109-120.
}
